Generate an EES API URL

This function returns a single URL to connect to a chosen EES API endpoint. The resulting URL can be used with GET or POST (as appropriate) in order to connect. Whether the publication_id or dataset_id (or neither) parameter are required depends on the endpoint chose.

Endpoints	id required

get-publications	Neither

get-data-catalogue	publication_id

get-summary, get-meta, get-data, post-data	dataset_id

Usage

api_url(
  endpoint = "get-publications",
  publication_id = NULL,
  dataset_id = NULL,
  indicators = NULL,
  time_periods = NULL,
  geographic_levels = NULL,
  locations = NULL,
  filter_items = NULL,
  dataset_version = NULL,
  page_size = NULL,
  page = NULL,
  api_version = "1.0",
  environment = "dev",
  verbose = FALSE
)

Arguments

endpoint

Name of endpoint, can be "get-publications", "get-data-catalogue", "get-summary", "get-meta", "get-data" or "post-data"

publication_id

ID of the publication to be connected to. This is required if the endpoint is "get-data-catalogue"

dataset_id

ID of data set to be connected to. This is required if the endpoint is one of "get-summary", "get-meta", "get-data" or "post-data"

indicators

Indicators required as a string or vector of strings (required)

time_periods

Time periods required as a string ("period|code") or vector of strings

geographic_levels

Geographic levels required as a string or vector of strings

locations

Location code required as a string or vector of strings

filter_items

Filter items required as a string or vector of strings

dataset_version

Version of data set to be connected to

page_size

Number of results to return in a single query

page

Page number of query results to return

api_version

EES API version

environment

EES environment to connect to: "dev", "test", "preprod" or "prod"

verbose

Run with additional contextual messaging. Logical, default = FALSE

Value

A string containing the URL for connecting to the EES API

Examples

api_url()
#> [1] "https://dev.statistics.api.education.gov.uk/api/v1.0/publications?"
api_url("get-publications")
#> [1] "https://dev.statistics.api.education.gov.uk/api/v1.0/publications?"
api_url("get-data-catalogue", publication_id = eesyapi::example_id("publication"))
#> [1] "https://dev.statistics.api.education.gov.uk/api/v1.0/publications/d823e4df-626f-4450-9b21-08dc8b95fc02/data-sets?"
api_url("get-summary", dataset_id = eesyapi::example_id("dataset"))
#> [1] "https://dev.statistics.api.education.gov.uk/api/v1.0/data-sets/830f9201-9e11-ad75-8dcd-d2efe2834457"
api_url("get-meta", dataset_id = eesyapi::example_id("dataset"))
#> [1] "https://dev.statistics.api.education.gov.uk/api/v1.0/data-sets/830f9201-9e11-ad75-8dcd-d2efe2834457/meta"
api_url(
  "get-data",
  dataset_id = eesyapi::example_id("dataset"),
  indicators = example_id("indicator"),
  time_periods = c("2024|W12", "2024|W13"),
  geographic_levels = c("NAT"),
  filter_items = example_id("filter_item")
)
#> [1] "https://dev.statistics.api.education.gov.uk/api/v1.0/data-sets/830f9201-9e11-ad75-8dcd-d2efe2834457/query?timePeriods.in=2024|W12%2C2024|W13&geographicLevels.in=NAT&filters.in=wEZcb&indicators=PbNeb"
api_url(
  "post-data",
  dataset_id = eesyapi::example_id("dataset"),
  indicators = example_id("indicator")
)
#> [1] "https://dev.statistics.api.education.gov.uk/api/v1.0/data-sets/830f9201-9e11-ad75-8dcd-d2efe2834457/query"