Get publication specific data set catalogue

Get publication specific data set catalogue

Usage

get_data_catalogue(
  publication_id,
  page_size = NULL,
  page = NULL,
  verbose = FALSE
)

Arguments

publication_id

The publication ID as used by the API

page_size

Number of results to collect in a single query to the API (max 40)

page

Page number to return (Default is NULL which will loop through until all pages of the query are collated).

verbose

Add extra contextual information whilst running

Value

Data frame listing the data sets contained within a single publication

Examples

get_data_catalogue(example_id("publication"))
#>                                     id                          title summary
#> 1 830f9201-9e11-ad75-8dcd-d2efe2834457 10MB data file for API dataset        
#> 2 830f9201-5343-7c73-98d7-8e5ac02933b2                    Loc names 3    test
#>      status latestVersion.version          latestVersion.published
#> 1 Published                   1.0 2024-09-20T13:20:32.887001+00:00
#> 2 Published                   1.0 2024-09-20T13:20:32.887001+00:00
#>   latestVersion.totalResults                     latestVersion.id
#> 1                      64801 41710a3a-1fce-4bbd-b6ee-697acf62a62f
#> 2                          2 88a4cd30-3f84-4634-89f8-bfa0d8e85e5a
#>   latestVersion.timePeriods.start latestVersion.timePeriods.end
#> 1                         2017/18                       2019/20
#> 2                            2018                          2018
#>        latestVersion.geographicLevels
#> 1 Local authority, National, Regional
#> 2                     Local authority
#>                                       latestVersion.filters
#> 1 Ethnicity, First language, Free School Meals, School type
#> 2                                                          
#>   latestVersion.indicators
#> 1       Headcount, Percent
#> 2        Admission Numbers