Query a data set using GET and a query URL

This function provides a method for generating and sending a URL based data query to the EES API. As a minimum, it requires the dataset_id and indicators flags to be provided.

Note that the GET command is very limited on the level of logic it can process. For example there is no way of using GET to make a query that combines different filters with AND logic. So if you give GET a set of filter items to search on, it can only return all rows containing any of those items.

Usage

get_dataset(
  dataset_id,
  indicators,
  time_periods = NULL,
  geographic_levels = NULL,
  locations = NULL,
  filter_items = NULL,
  dataset_version = NULL,
  api_version = NULL,
  page = NULL,
  page_size = 1000,
  parse = TRUE,
  verbose = FALSE
)

Arguments

dataset_id

ID of data set to be connected to. This is required if the endpoint is one of "get-summary", "get-meta", "get-data" or "post-data"

indicators

Indicators required as a string or vector of strings (required)

time_periods

Time periods required as a string ("period|code") or vector of strings

geographic_levels

Geographic levels required as a string or vector of strings

locations

Location code required as a string or vector of strings

filter_items

Filter items required as a string or vector of strings

dataset_version

Version of data set to be connected to

api_version

EES API version

page

Page number of query results to return

page_size

Number of results to return in a single query

parse

Logical flag to activate parsing of the results. Default: TRUE

verbose

Add extra contextual information whilst running

Value

Data frame containing query results of an API data set

Examples

get_dataset(
  example_id(),
  geographic_levels = "NAT",
  filter_items = example_id("filter_item"),
  indicators = example_id("indicator")
)
#>    code    period geographic_level   NAT 1Nei3 deq0Z VTQ5X Pmjeb   mRj9K
#> 1    AY 2019/2020              NAT dP0Zw qFujG kKOhs P2yeb LrEWj   74338
#> 2    AY 2019/2020              NAT dP0Zw qFujG kKOhs bfQtT LrEWj   97786
#> 3    AY 2019/2020              NAT dP0Zw qFujG kKOhs DBNQe LrEWj   88631
#> 4    AY 2019/2020              NAT dP0Zw qFujG kKOhs e2euS LrEWj   80711
#> 5    AY 2019/2020              NAT dP0Zw qFujG kKOhs 5y0di LrEWj  154410
#> 6    AY 2019/2020              NAT dP0Zw qFujG kKOhs 4skUZ LrEWj  208654
#> 7    AY 2019/2020              NAT dP0Zw qFujG kKOhs m8R9K LrEWj   36438
#> 8    AY 2019/2020              NAT dP0Zw qFujG kKOhs 0LoT5 LrEWj  176065
#> 9    AY 2019/2020              NAT dP0Zw qFujG kKOhs WNZ2b LrEWj   42937
#> 10   AY 2019/2020              NAT dP0Zw qFujG kKOhs EzCqF LrEWj   23061
#> 11   AY 2019/2020              NAT dP0Zw 0LxT5 kKOhs NBODC LrEWj 1660128
#> 12   AY 2019/2020              NAT dP0Zw qFujG kKOhs 70eXo LrEWj 3064344
#> 13   AY 2019/2020              NAT dP0Zw qFujG kKOhs OS5CL LrEWj    5059
#> 14   AY 2019/2020              NAT dP0Zw qFujG kKOhs daL0Z LrEWj   11518
#> 15   AY 2019/2020              NAT dP0Zw qFujG kKOhs 9ZZ4v LrEWj   18811
#> 16   AY 2019/2020              NAT dP0Zw qFujG kKOhs o3cmX LrEWj  346876
#> 17   AY 2019/2020              NAT dP0Zw qFujG kKOhs wf7bx LrEWj   53527
#> 18   AY 2019/2020              NAT dP0Zw qFujG kKOhs NBODC LrEWj 4714772
#> 19   AY 2019/2020              NAT dP0Zw qFujG kKOhs NBODC iOf6X   10226
#> 20   AY 2019/2020              NAT dP0Zw qFujG kKOhs NBODC I5eBz 1002387
#> 21   AY 2019/2020              NAT dP0Zw qFujG kKOhs NBODC pRaSo 3702159
#> 22   AY 2019/2020              NAT dP0Zw YeLHK kKOhs NBODC LrEWj 4522618
#> 23   AY 2019/2020              NAT dP0Zw CV4Id kKOhs NBODC LrEWj  824237
#> 24   AY 2019/2020              NAT dP0Zw HuazL kKOhs NBODC LrEWj  835776
#> 25   AY 2019/2020              NAT dP0Zw qFujG kKOhs 9ZW4v LrEWj   73567
#> 26   AY 2019/2020              NAT dP0Zw qFujG kKOhs QXGwb LrEWj  115600
#> 27   AY 2019/2020              NAT dP0Zw qFujG kKOhs 2D1lz LrEWj   42439
#> 28   AY 2019/2020              NAT dP0Zw 6Czrf kKOhs NBODC LrEWj  668680
#> 29   AY 2018/2019              NAT dP0Zw HuazL kKOhs NBODC LrEWj  745453
#> 30   AY 2018/2019              NAT dP0Zw 0LxT5 kKOhs NBODC LrEWj 1666230
#> 31   AY 2018/2019              NAT dP0Zw qFujG kKOhs 70eXo LrEWj 3094756
#> 32   AY 2018/2019              NAT dP0Zw qFujG kKOhs OS5CL LrEWj    4959
#> 33   AY 2018/2019              NAT dP0Zw qFujG kKOhs daL0Z LrEWj   11839
#> 34   AY 2018/2019              NAT dP0Zw qFujG kKOhs 9ZZ4v LrEWj   18556
#> 35   AY 2018/2019              NAT dP0Zw qFujG kKOhs o3cmX LrEWj  345269
#> 36   AY 2018/2019              NAT dP0Zw qFujG kKOhs wf7bx LrEWj   48771
#> 37   AY 2018/2019              NAT dP0Zw qFujG kKOhs NBODC LrEWj 4727089
#> 38   AY 2018/2019              NAT dP0Zw qFujG kKOhs NBODC iOf6X    8515
#> 39   AY 2018/2019              NAT dP0Zw qFujG kKOhs NBODC I5eBz 1002292
#> 40   AY 2018/2019              NAT dP0Zw qFujG kKOhs NBODC pRaSo 3716282
#> 41   AY 2018/2019              NAT dP0Zw YeLHK kKOhs NBODC LrEWj 4521272
#> 42   AY 2018/2019              NAT dP0Zw 6Czrf kKOhs NBODC LrEWj  596162
#> 43   AY 2018/2019              NAT dP0Zw qFujG kKOhs P2yeb LrEWj   74345
#> 44   AY 2018/2019              NAT dP0Zw qFujG kKOhs 2D1lz LrEWj   41651
#> 45   AY 2018/2019              NAT dP0Zw qFujG kKOhs 9ZW4v LrEWj   71432
#> 46   AY 2018/2019              NAT dP0Zw qFujG kKOhs QXGwb LrEWj  111438
#> 47   AY 2018/2019              NAT dP0Zw qFujG kKOhs EzCqF LrEWj   22844
#> 48   AY 2018/2019              NAT dP0Zw qFujG kKOhs WNZ2b LrEWj   44582
#> 49   AY 2018/2019              NAT dP0Zw qFujG kKOhs 0LoT5 LrEWj  177656
#> 50   AY 2018/2019              NAT dP0Zw qFujG kKOhs m8R9K LrEWj   36536
#> 51   AY 2018/2019              NAT dP0Zw qFujG kKOhs 4skUZ LrEWj  208869
#> 52   AY 2018/2019              NAT dP0Zw qFujG kKOhs 5y0di LrEWj  149936
#> 53   AY 2018/2019              NAT dP0Zw qFujG kKOhs e2euS LrEWj   80450
#> 54   AY 2018/2019              NAT dP0Zw qFujG kKOhs DBNQe LrEWj   88136
#> 55   AY 2018/2019              NAT dP0Zw qFujG kKOhs bfQtT LrEWj   95064
#> 56   AY 2018/2019              NAT dP0Zw CV4Id kKOhs NBODC LrEWj  734088