Get data set versions

Get data set versions

Usage

get_dataset_versions(
  dataset_id,
  detail = "light",
  ees_environment = NULL,
  api_version = NULL,
  page_size = 40,
  page = NULL,
  verbose = FALSE
)

Arguments

dataset_id

ID of data set to be connected to. This is required if the endpoint is one of "get-dataset-versions", "get-summary", "get-meta", "get-csv", "get-data" or "post-data"

detail

Level of detail to return. Given as a character string, it should be one of: "light" (default) or "full". Using "light" gives the following:

version - version number

type - version type (major or minor)

total_rows - total number of rows in the listed version

release_date - release date of the data set version

release_name - name of the release in which the data set version was published

time_period_start - first available time period in the data

time_period_start - latest available time period in the data

Using "full" adds the following in addition to the above:

version_id - unique ID code for the data set version

status - publication status of the data set version

geographic_levels - geographic levels provided

filters - names of filters in the data set version

indicators - names of indicators in the data set version

notes - any additional information

ees_environment

EES ees_environment to connect to: "dev", "test", "preprod" or "prod"

api_version

EES API version

page_size

Number of results to return in a single query

page

Page number of query results to return

verbose

Run with additional contextual messaging. Logical, default = FALSE

Value

A data frame with X columns:

Data frame listing all available versions of the given data set

Examples

get_dataset_versions(dataset_id = example_id(group = "attendance"))
#>   version  type total_rows release_date release_name time_period_start
#> 1     1.0 Major    2750478   2025-03-06  Week 7 2025      2024 Week 37
#>   time_period_end
#> 1     2025 Week 8