Get a parsed version of the API response for a data set's meta data

Get a list of metadata information for a data set available from the EES API. Provides either look-up tables from human readable labels to ids used in the API, or the raw response from the meta endpoint.

Usage

get_meta(
  dataset_id,
  dataset_version = NULL,
  api_version = NULL,
  verbose = FALSE
)

Arguments

dataset_id

ID of data set to be connected to. This is required if the endpoint is one of "get-summary", "get-meta", "get-csv", "get-data" or "post-data"

dataset_version

Version of data set to be connected to

api_version

EES API version

verbose

Run with additional contextual messaging. Logical, default = FALSE

Value

List of data frames containing a data set's meta data

Examples

get_meta(example_id())
#> $time_periods
#>   code    period   label
#> 1   AY 2022/2023 2022/23
#> 
#> $locations
#>    item_id                          label      code oldCode
#> 1    ea0uS                        Croydon E09000008     306
#> 2    4c9UZ                        Enfield E09000010     308
#> 3    QIqwb                         Harrow E09000015     310
#> 4    rGcNj                          Leeds E08000035     383
#> 5    eLGuS                   Lincolnshire E10000019     925
#> 6    H0RzL                     Manchester E08000003     352
#> 7    VUG5X                       Somerset E10000027     933
#> 8    lrocB                    West Sussex E10000032     938
#> 9    jfhAM                           York E06000014     816
#> 10   dP0Zw                        England E92000001    <NA>
#> 11   42nEU                 Bootham School      <NA>    <NA>
#> 12   y8Jra           Brigshaw High School      <NA>    <NA>
#> 13   k2qHh      Felpham Community College      <NA>    <NA>
#> 14   mLJt9                 Loreto College      <NA>    <NA>
#> 15   59BKd North London Collegiate School      <NA>    <NA>
#> 16   dNu0Z             The Latymer School      <NA>    <NA>
#> 17   0UXNT             The Laurels School      <NA>    <NA>
#> 18   BzBp7      The Priory Ruskin Academy      <NA>    <NA>
#> 19   hciHy    University Academy Holbeach      <NA>    <NA>
#> 20   HS5ez          West Somerset College      <NA>    <NA>
#>    geographic_level_code geographic_level    urn laEstab
#> 1                     LA  Local authority   <NA>    <NA>
#> 2                     LA  Local authority   <NA>    <NA>
#> 3                     LA  Local authority   <NA>    <NA>
#> 4                     LA  Local authority   <NA>    <NA>
#> 5                     LA  Local authority   <NA>    <NA>
#> 6                     LA  Local authority   <NA>    <NA>
#> 7                     LA  Local authority   <NA>    <NA>
#> 8                     LA  Local authority   <NA>    <NA>
#> 9                     LA  Local authority   <NA>    <NA>
#> 10                   NAT         National   <NA>    <NA>
#> 11                   SCH           School 121722 8166000
#> 12                   SCH           School 143238 3834113
#> 13                   SCH           School 126080 9384059
#> 14                   SCH           School 130503 3528601
#> 15                   SCH           School 102257 3106075
#> 16                   SCH           School 102055 3085400
#> 17                   SCH           School 140356 3066028
#> 18                   SCH           School 136194 9256910
#> 19                   SCH           School 137282 9254001
#> 20                   SCH           School 136791 9334291
#> 
#> $filter_columns
#>   col_id           col_name              label
#> 1  tu1Em establishment_type establishment type
#> 2  f2GYF               type               type
#> 
#> $filter_items
#>    col_id             item_label item_id default_item
#> 1   tu1Em      Academy Converter   9RpY4           NA
#> 2   tu1Em    Academy Sponsor Led   23odl           NA
#> 3   tu1Em       Community School   Poz7e           NA
#> 4   tu1Em     Independent School   NwcLD           NA
#> 5   tu1Em     Sixth Form College   wEZcb           NA
#> 6   tu1Em Voluntary Aided School   o4vFm           NA
#> 7   f2GYF                   1618   9Rpn4           NA
#> 8   f2GYF               1618_DIS   dPff0           NA
#> 9   f2GYF            1618_NOTDIS   OBNDC           NA
#> 10  f2GYF                   ACAD   7YbEX           NA
#> 11  f2GYF               ACAD_DIS   p5IXS           NA
#> 12  f2GYF            ACAD_NOTDIS   IXQEB           NA
#> 13  f2GYF                   AGEN   iNW86           NA
#> 14  f2GYF               AGEN_DIS   6D2xr           NA
#> 15  f2GYF            AGEN_NOTDIS   0eYST           NA
#> 16  f2GYF              AGENDENOM   LleUW           NA
#> 17  f2GYF                   ALEV   HJjXz           NA
#> 18  f2GYF               ALEV_DIS   C7DgI           NA
#> 19  f2GYF            ALEV_NOTDIS   YFqfH           NA
#> 20  f2GYF               ANYL2VOC   qGKqj           NA
#> 21  f2GYF              SCOPE_L3M   a2eWL           NA
#> 22  f2GYF               TechCert   Bfbf7           NA
#> 23  f2GYF           TechCert_DIS   zvGnF           NA
#> 24  f2GYF        TechCert_NOTDIS   ka0Wh           NA
#> 25  f2GYF                   TLEV   u9vRo           NA
#> 26  f2GYF               TLEV_DIS   TuzUP           NA
#> 27  f2GYF            TLEV_NOTDIS   cZ4i3           NA
#> 28  f2GYF              TLEVDENOM   XHEvf           NA
#> 
#> $indicators
#>   col_id             col_name                label
#> 1  E40qF            ppe_value            ppe value
#> 2  QuVwb         ppegrd_value         ppegrd value
#> 3  9Aw4v retained_2ndyr_value retained 2ndyr value
#> 4  2YIlz       retained_value       retained value
#> 5  PbNeb    total_pupil_value    total pupil value
#>