Get publications

Get publications

Usage

get_publication_catalogue(page_size = 40, page = NULL, verbose = FALSE)

Arguments

page_size

Number of results to collect in a single query to the API (max 40)

page

Page number to return (Default is NULL which will loop through until all pages of the query are collated).

verbose

Add extra contextual information whilst running

Value

Data frame listing all available publications

Examples

get_publication_catalogue()
#>                                      id
#> 1  b7c85f7e-3f1b-4839-2404-08dcb61b5e69
#> 2  0384fab9-a8f9-4e2c-b388-08dc8ccb7475
#> 3  b6d9ed96-be68-4791-abc3-08dcaba68c04
#> 4  82c3b19e-1060-4f5e-3410-08dc0b96cfea
#> 5  a42c23bb-0cb1-4dda-3d7a-08dccbde7f0d
#> 6  09bf5798-50f3-4698-3d7b-08dccbde7f0d
#> 7  1bc176ef-4165-487a-9e1c-08dcb6e95ade
#> 8  797cd607-b35f-4c97-6093-08dca593ae8f
#> 9  d823e4df-626f-4450-9b21-08dc8b95fc02
#> 10 f3d0caa1-faf2-4dd2-fb33-08dcbcf030f0
#> 11 3f49cc72-d30d-4820-41fc-08dcd5e3d21c
#> 12 0d377b9b-7164-427c-f791-08dcac872413
#> 13 bc5e9cc6-a716-433c-4145-08dcb23ba7a0
#> 14 e54954bd-c40f-4c4e-9ded-08dcb6e95ade
#> 15 ad60ce0a-3bb2-42ab-1551-08dcc3397c5f
#> 16 eae84b36-d42a-4b47-6a3f-08dcad647d8c
#> 17 6a6953f5-75f8-43a6-a0ca-08dcb200bd26
#> 18 80c39b96-93bc-48c9-a6a1-08dcc3729eee
#> 19 8e53c084-f84b-4858-6fbd-08dcc270658d
#> 20 dcfc984b-7d79-4065-23c8-08dcb61b5e69
#> 21 763c067a-c2e3-47fd-1bb2-08dcc8124275
#> 22 dcac68d3-1bb0-495a-0470-08dcd227ff37
#> 23 f8ac7265-1085-4617-0472-08dcd227ff37
#> 24 e0283aa9-9494-4f55-3340-08dccda775e7
#> 25 9d0ed9b2-fc58-4883-40ea-08dcd182d841
#> 26 b5f549d9-37da-4fb4-40ec-08dcd182d841
#> 27 30329ec7-1f06-49ac-107a-08dcd30c0aeb
#> 28 df406dc8-36e3-459b-6f1b-08dcd3140d7b
#> 29 12bf6979-d756-4d5a-b477-08dcd3278032
#> 30 8e40122e-e9bf-4996-b479-08dcd3278032
#> 31 2e952cae-52ea-4e79-d007-08dcd3e026e0
#> 32 c6d5f84c-93bf-4462-d009-08dcd3e026e0
#> 33 9c3b06f6-54e6-45de-2db7-08dcd40056ed
#> 34 88d600a8-c2e0-4f62-5aef-08dcd406aa3f
#> 35 622fbfa0-d3f3-40d0-fb38-08dcbcf030f0
#> 36 90b988b6-0616-4bcd-4f78-08dcbd21b014
#> 37 e185a089-d61e-4c62-4f7e-08dcbd21b014
#> 38 0e15143b-deff-416b-d1ae-08dcbd437af9
#> 39 00684d70-6145-48c8-d1b3-08dcbd437af9
#> 40 81356f76-3b1d-42e5-d1b5-08dcbd437af9
#> 41 3c27c90c-93a9-4633-2a5a-08dcbdebd2d7
#> 42 6a4ab7d9-a78c-44e0-2a5b-08dcbdebd2d7
#> 43 dcd528a1-d825-4a1c-2a5e-08dcbdebd2d7
#> 44 2ec8b977-e7c2-461d-fdd6-08dcc01501e4
#> 45 94fbea55-f8b3-43aa-fdd9-08dcc01501e4
#> 46 8e8b752f-d851-4a1b-fddc-08dcc01501e4
#> 47 8bbc40d4-6064-4687-36e8-08dcc11ec214
#> 48 c67a1fdf-d151-4c4d-36e9-08dcc11ec214
#> 49 3b7a072a-e816-4707-dd8b-08dcc1c817b8
#> 50 626a5077-20fe-4eb8-2de6-08dcc2c36ace
#> 51 51939cb3-d565-466a-8d4e-08dcc5951828
#> 52 4b55e9e3-0976-4342-8d4f-08dcc5951828
#> 53 3f3b1def-33dc-4845-1bb4-08dcc8124275
#> 54 a8bac0ad-98d9-4fdc-0cd6-08dcc90be889
#> 55 ddfe5285-a81f-40e0-531c-08dccb4c54a6
#> 56 c4eafefa-bc65-46e4-531d-08dccb4c54a6
#> 57 45472e70-cd4c-4756-3d7d-08dccbde7f0d
#> 58 cd2b91aa-aa6e-4d81-3d7e-08dccbde7f0d
#> 59 2ae92a9c-b105-412b-3d7f-08dccbde7f0d
#> 60 51731d51-9cbc-4b18-40ef-08dcd182d841
#> 61 367599a5-3511-4fad-40f2-08dcd182d841
#> 62 5ff14750-70b2-4651-40f4-08dcd182d841
#> 63 0a5c63b8-f55b-466f-0471-08dcd227ff37
#>                                                                              title
#> 1                                                                  07-08 new  test
#> 2                                                               API data sets test
#> 3                                                              Attendance API data
#> 4                                                     CSV filename validation test
#> 5                                                                    EES-4859 Test
#> 6                                                                  EES-4859 Test 2
#> 7                                                                       public api
#> 8                                                             Public API data sets
#> 9                                         Public API dataset testing-Updated title
#> 10                                                                 public api test
#> 11                                                         Pupil absence (example)
#> 12                                                              SV  filtering test
#> 13                                                             SV api error test 2
#> 14                                                    SV data set version new test
#> 15                                                 SV data set versions test fresh
#> 16                               SV different release to previous data set v. test
#> 17                                            SV location and filter mappings test
#> 18                                            SV location columns in mappings test
#> 19                                                         SV new changelog test 2
#> 20                                                           SV preview token test
#> 21                                                 SV subscription test - versions
#> 22                                                           Test publication- two
#> 23                                                                    test_version
#> 24 UI tests - Public API - Generate and Preview API token 20240905-154004-dnbp0k-0
#> 25 UI tests - Public API - Generate and Preview API token 20240910-102412-7xzk6r-0
#> 26 UI tests - Public API - Generate and Preview API token 20240910-104342-lpfafr-0
#> 27         UI tests - public api resolve mapping statuses 20240912-092909-2z8zjw-0
#> 28         UI tests - public api resolve mapping statuses 20240912-104059-5bc4iq-0
#> 29         UI tests - public api resolve mapping statuses 20240912-123545-qps47t-0
#> 30         UI tests - public api resolve mapping statuses 20240912-140204-p68rgv-0
#> 31         UI tests - public api resolve mapping statuses 20240913-103729-b0rmxl-0
#> 32         UI tests - public api resolve mapping statuses 20240913-134647-tmvdt8-0
#> 33         UI tests - public api resolve mapping statuses 20240913-144827-sxnbu7-0
#> 34         UI tests - public api resolve mapping statuses 20240913-151337-4eaacd-0
#> 35                       UI tests - public api restricted 20240815-104142-5oawar-0
#> 36                       UI tests - public api restricted 20240815-121821-q82h4l-0
#> 37                       UI tests - public api restricted 20240815-132100-dhgfhq-0
#> 38                       UI tests - public api restricted 20240815-160023-hijo8m-0
#> 39                       UI tests - public api restricted 20240815-173741-27mbiz-0
#> 40                       UI tests - public api restricted 20240815-182019-d4ulsy-0
#> 41                       UI tests - public api restricted 20240816-120540-lfoygh-0
#> 42                       UI tests - public api restricted 20240816-122705-tmbu3n-0
#> 43                       UI tests - public api restricted 20240816-130357-f1csy9-0
#> 44                       UI tests - public api restricted 20240819-132529-f4t4ql-0
#> 45                       UI tests - public api restricted 20240819-134741-dg912l-0
#> 46                       UI tests - public api restricted 20240819-141655-3hzywz-0
#> 47                       UI tests - public api restricted 20240820-141415-sad28v-0
#> 48                       UI tests - public api restricted 20240820-142835-rfwh1a-0
#> 49                       UI tests - public api restricted 20240821-101725-39g358-0
#> 50                       UI tests - public api restricted 20240822-160725-pqyxoq-0
#> 51                       UI tests - public api restricted 20240827-133156-84xwl2-0
#> 52                       UI tests - public api restricted 20240827-134827-e2awin-0
#> 53                       UI tests - public api restricted 20240829-101651-4dfg3c-0
#> 54                       UI tests - public api restricted 20240830-155327-ru0ery-0
#> 55                       UI tests - public api restricted 20240902-124015-55geng-0
#> 56                       UI tests - public api restricted 20240902-124429-hodesh-0
#> 57                       UI tests - public api restricted 20240903-090649-t1p11g-0
#> 58                       UI tests - public api restricted 20240903-094705-ato012-0
#> 59                       UI tests - public api restricted 20240903-095523-uies1z-0
#> 60                       UI tests - public api restricted 20240910-124108-oo4ebr-0
#> 61                       UI tests - public api restricted 20240910-131643-0kvip6-0
#> 62                       UI tests - public api restricted 20240910-151652-m6czfd-0
#> 63                                                                       version 2
#>                                                                           slug
#> 1                                                               07-08-new-test
#> 2                                                           api-data-sets-test
#> 3                                                          attendance-api-data
#> 4                                                 csv-filename-validation-test
#> 5                                                                ees-4859-test
#> 6                                                              ees-4859-test-2
#> 7                                                                   public-api
#> 8                                                         public-api-data-sets
#> 9                                     public-api-dataset-testing-updated-title
#> 10                                                             public-api-test
#> 11                                                       pupil-absence-example
#> 12                                                           sv-filtering-test
#> 13                                                         sv-api-error-test-2
#> 14                                                sv-data-set-version-new-test
#> 15                                             sv-data-set-versions-test-fresh
#> 16                            sv-different-release-to-previous-data-set-v-test
#> 17                                        sv-location-and-filter-mappings-test
#> 18                                        sv-location-columns-in-mappings-test
#> 19                                                     sv-new-changelog-test-2
#> 20                                                       sv-preview-token-test
#> 21                                               sv-subscription-test-versions
#> 22                                                        test-publication-two
#> 23                                                                test-version
#> 24 ui-tests-public-api-generate-and-preview-api-token-20240905-154004-dnbp0k-0
#> 25 ui-tests-public-api-generate-and-preview-api-token-20240910-102412-7xzk6r-0
#> 26 ui-tests-public-api-generate-and-preview-api-token-20240910-104342-lpfafr-0
#> 27       ui-tests-public-api-resolve-mapping-statuses-20240912-092909-2z8zjw-0
#> 28       ui-tests-public-api-resolve-mapping-statuses-20240912-104059-5bc4iq-0
#> 29       ui-tests-public-api-resolve-mapping-statuses-20240912-123545-qps47t-0
#> 30       ui-tests-public-api-resolve-mapping-statuses-20240912-140204-p68rgv-0
#> 31       ui-tests-public-api-resolve-mapping-statuses-20240913-103729-b0rmxl-0
#> 32       ui-tests-public-api-resolve-mapping-statuses-20240913-134647-tmvdt8-0
#> 33       ui-tests-public-api-resolve-mapping-statuses-20240913-144827-sxnbu7-0
#> 34       ui-tests-public-api-resolve-mapping-statuses-20240913-151337-4eaacd-0
#> 35                     ui-tests-public-api-restricted-20240815-104142-5oawar-0
#> 36                     ui-tests-public-api-restricted-20240815-121821-q82h4l-0
#> 37                     ui-tests-public-api-restricted-20240815-132100-dhgfhq-0
#> 38                     ui-tests-public-api-restricted-20240815-160023-hijo8m-0
#> 39                     ui-tests-public-api-restricted-20240815-173741-27mbiz-0
#> 40                     ui-tests-public-api-restricted-20240815-182019-d4ulsy-0
#> 41                     ui-tests-public-api-restricted-20240816-120540-lfoygh-0
#> 42                     ui-tests-public-api-restricted-20240816-122705-tmbu3n-0
#> 43                     ui-tests-public-api-restricted-20240816-130357-f1csy9-0
#> 44                     ui-tests-public-api-restricted-20240819-132529-f4t4ql-0
#> 45                     ui-tests-public-api-restricted-20240819-134741-dg912l-0
#> 46                     ui-tests-public-api-restricted-20240819-141655-3hzywz-0
#> 47                     ui-tests-public-api-restricted-20240820-141415-sad28v-0
#> 48                     ui-tests-public-api-restricted-20240820-142835-rfwh1a-0
#> 49                     ui-tests-public-api-restricted-20240821-101725-39g358-0
#> 50                     ui-tests-public-api-restricted-20240822-160725-pqyxoq-0
#> 51                     ui-tests-public-api-restricted-20240827-133156-84xwl2-0
#> 52                     ui-tests-public-api-restricted-20240827-134827-e2awin-0
#> 53                     ui-tests-public-api-restricted-20240829-101651-4dfg3c-0
#> 54                     ui-tests-public-api-restricted-20240830-155327-ru0ery-0
#> 55                     ui-tests-public-api-restricted-20240902-124015-55geng-0
#> 56                     ui-tests-public-api-restricted-20240902-124429-hodesh-0
#> 57                     ui-tests-public-api-restricted-20240903-090649-t1p11g-0
#> 58                     ui-tests-public-api-restricted-20240903-094705-ato012-0
#> 59                     ui-tests-public-api-restricted-20240903-095523-uies1z-0
#> 60                     ui-tests-public-api-restricted-20240910-124108-oo4ebr-0
#> 61                     ui-tests-public-api-restricted-20240910-131643-0kvip6-0
#> 62                     ui-tests-public-api-restricted-20240910-151652-m6czfd-0
#> 63                                                                   version-2
#>                                                                                    summary
#> 1                                                                                     asdf
#> 2                                                                                Test test
#> 3                                          A place to demo tidy attendance data on the API
#> 4                                                                                     TEST
#> 5                                                                                        P
#> 6                                                                                        T
#> 7                                                                                    dnadk
#> 8                                                                                     asdf
#> 9                                                                                     test
#> 10                                                                                   skank
#> 11                                             Demo publication showing absence statistics
#> 12                                                                                      tt
#> 13                                                                                    test
#> 14                                                                                    asdf
#> 15                                                                                     ads
#> 16                                                                                     ttt
#> 17                                                                                    test
#> 18                                                                                   asdfg
#> 19                                                                                     asd
#> 20                                                                                    test
#> 21                                                                                    test
#> 22                                                                                 Summary
#> 23                                                                                 version
#> 24 UI tests - Public API - Generate and Preview API token 20240905-154004-dnbp0k-0 summary
#> 25 UI tests - Public API - Generate and Preview API token 20240910-102412-7xzk6r-0 summary
#> 26 UI tests - Public API - Generate and Preview API token 20240910-104342-lpfafr-0 summary
#> 27         UI tests - public api resolve mapping statuses 20240912-092909-2z8zjw-0 summary
#> 28         UI tests - public api resolve mapping statuses 20240912-104059-5bc4iq-0 summary
#> 29         UI tests - public api resolve mapping statuses 20240912-123545-qps47t-0 summary
#> 30         UI tests - public api resolve mapping statuses 20240912-140204-p68rgv-0 summary
#> 31         UI tests - public api resolve mapping statuses 20240913-103729-b0rmxl-0 summary
#> 32         UI tests - public api resolve mapping statuses 20240913-134647-tmvdt8-0 summary
#> 33         UI tests - public api resolve mapping statuses 20240913-144827-sxnbu7-0 summary
#> 34         UI tests - public api resolve mapping statuses 20240913-151337-4eaacd-0 summary
#> 35                       UI tests - public api restricted 20240815-104142-5oawar-0 summary
#> 36                       UI tests - public api restricted 20240815-121821-q82h4l-0 summary
#> 37                       UI tests - public api restricted 20240815-132100-dhgfhq-0 summary
#> 38                       UI tests - public api restricted 20240815-160023-hijo8m-0 summary
#> 39                       UI tests - public api restricted 20240815-173741-27mbiz-0 summary
#> 40                       UI tests - public api restricted 20240815-182019-d4ulsy-0 summary
#> 41                       UI tests - public api restricted 20240816-120540-lfoygh-0 summary
#> 42                       UI tests - public api restricted 20240816-122705-tmbu3n-0 summary
#> 43                       UI tests - public api restricted 20240816-130357-f1csy9-0 summary
#> 44                       UI tests - public api restricted 20240819-132529-f4t4ql-0 summary
#> 45                       UI tests - public api restricted 20240819-134741-dg912l-0 summary
#> 46                       UI tests - public api restricted 20240819-141655-3hzywz-0 summary
#> 47                       UI tests - public api restricted 20240820-141415-sad28v-0 summary
#> 48                       UI tests - public api restricted 20240820-142835-rfwh1a-0 summary
#> 49                       UI tests - public api restricted 20240821-101725-39g358-0 summary
#> 50                       UI tests - public api restricted 20240822-160725-pqyxoq-0 summary
#> 51                       UI tests - public api restricted 20240827-133156-84xwl2-0 summary
#> 52                       UI tests - public api restricted 20240827-134827-e2awin-0 summary
#> 53                       UI tests - public api restricted 20240829-101651-4dfg3c-0 summary
#> 54                       UI tests - public api restricted 20240830-155327-ru0ery-0 summary
#> 55                       UI tests - public api restricted 20240902-124015-55geng-0 summary
#> 56                       UI tests - public api restricted 20240902-124429-hodesh-0 summary
#> 57                       UI tests - public api restricted 20240903-090649-t1p11g-0 summary
#> 58                       UI tests - public api restricted 20240903-094705-ato012-0 summary
#> 59                       UI tests - public api restricted 20240903-095523-uies1z-0 summary
#> 60                       UI tests - public api restricted 20240910-124108-oo4ebr-0 summary
#> 61                       UI tests - public api restricted 20240910-131643-0kvip6-0 summary
#> 62                       UI tests - public api restricted 20240910-151652-m6czfd-0 summary
#> 63                                                                                    test
#>                        lastPublished
#> 1  2024-08-08T10:00:02.2934882+00:00
#> 2  2024-08-27T15:33:26.2534352+00:00
#> 3  2024-09-17T11:13:41.8711175+00:00
#> 4  2024-09-16T08:26:35.9533351+00:00
#> 5   2024-09-03T08:14:08.960874+00:00
#> 6  2024-09-03T08:35:47.3987126+00:00
#> 7  2024-08-14T13:36:18.6631746+00:00
#> 8  2024-08-30T19:44:04.7735973+00:00
#> 9  2024-09-06T09:43:27.1491755+00:00
#> 10 2024-08-15T07:41:41.7761049+00:00
#> 11 2024-09-16T00:13:59.1295677+00:00
#> 12 2024-08-29T10:09:58.2034103+00:00
#> 13 2024-08-01T15:09:03.6297181+00:00
#> 14 2024-08-08T11:07:06.7644435+00:00
#> 15 2024-08-23T09:44:05.7204217+00:00
#> 16 2024-08-08T10:22:19.9236641+00:00
#> 17 2024-08-08T12:02:03.1816198+00:00
#> 18 2024-08-29T10:02:13.5516433+00:00
#> 19 2024-08-22T12:16:32.7751639+00:00
#> 20 2024-08-06T21:13:56.6808974+00:00
#> 21 2024-08-29T10:23:47.9366621+00:00
#> 22 2024-09-11T10:02:04.3541199+00:00
#> 23 2024-09-11T12:19:47.2252341+00:00
#> 24 2024-09-05T15:43:27.5207737+00:00
#> 25  2024-09-10T10:28:05.304492+00:00
#> 26 2024-09-10T10:46:15.6825521+00:00
#> 27 2024-09-12T09:30:26.6421893+00:00
#> 28 2024-09-12T10:42:18.9957315+00:00
#> 29 2024-09-12T12:37:32.2971201+00:00
#> 30 2024-09-12T14:03:22.8623098+00:00
#> 31 2024-09-13T10:39:38.4566627+00:00
#> 32 2024-09-13T13:48:49.4468138+00:00
#> 33 2024-09-13T14:49:45.9984798+00:00
#> 34 2024-09-13T15:15:04.1199017+00:00
#> 35 2024-08-15T10:46:51.9802821+00:00
#> 36 2024-08-15T12:19:54.3404049+00:00
#> 37 2024-08-15T13:25:02.6714358+00:00
#> 38 2024-08-15T16:05:55.2826284+00:00
#> 39  2024-08-15T17:39:15.783522+00:00
#> 40 2024-08-15T18:23:38.3395014+00:00
#> 41 2024-08-16T12:17:53.9213595+00:00
#> 42 2024-08-16T12:29:16.2725999+00:00
#> 43 2024-08-16T13:06:17.1243932+00:00
#> 44 2024-08-19T13:28:59.8082831+00:00
#> 45   2024-08-19T13:50:23.18643+00:00
#> 46 2024-08-19T14:19:49.7998889+00:00
#> 47 2024-08-20T14:15:57.5821238+00:00
#> 48  2024-08-20T14:30:53.099727+00:00
#> 49 2024-08-21T10:20:10.7410575+00:00
#> 50 2024-08-22T16:09:43.1779826+00:00
#> 51 2024-08-27T13:33:35.9204674+00:00
#> 52  2024-08-27T13:50:37.299422+00:00
#> 53 2024-08-29T10:19:22.5366641+00:00
#> 54 2024-08-30T15:56:59.1494367+00:00
#> 55   2024-09-02T12:43:15.06004+00:00
#> 56 2024-09-02T12:46:30.9172256+00:00
#> 57 2024-09-03T09:09:22.7557929+00:00
#> 58 2024-09-03T09:49:16.9291002+00:00
#> 59 2024-09-03T09:57:41.5733993+00:00
#> 60 2024-09-10T12:42:57.1837219+00:00
#> 61 2024-09-10T13:18:24.5634739+00:00
#> 62 2024-09-10T15:19:40.1834534+00:00
#> 63 2024-09-11T11:58:42.1983859+00:00