Get publication specific data set catalogue

Get publication specific data set catalogue

Usage

get_publication_datasets(
  publication_id,
  page_size = NULL,
  page = NULL,
  verbose = FALSE
)

Arguments

publication_id

The publication ID as used by the API

page_size

Number of results to collect in a single query to the API (max 40)

page

Page number to return (Default is NULL which will loop through until all pages of the query are collated).

verbose

Add extra contextual information whilst running

Value

Data frame listing the data sets contained within a single publication

Examples

get_publication_datasets(example_id("publication"))
#>                                     id                           title summary
#> 1 a3ff9101-bce1-9774-b1f5-b698f3311168 Attendance in schools test data        
#> 2 d7329101-f275-d277-bbfe-d8cfaa709833           Attendance in schools        
#>      status latestVersion.version          latestVersion.published
#> 1 Published                   1.0 2024-09-17T11:13:59.135518+00:00
#> 2 Published                   1.0 2024-08-15T08:45:50.015731+00:00
#>   latestVersion.totalResults                     latestVersion.id
#> 1                    3723444 77787b8c-a833-4825-aa3f-b030cd429e81
#> 2                    3723444 00000000-0000-0000-0000-000000000000
#>   latestVersion.timePeriods.start latestVersion.timePeriods.end
#> 1                    2023 Week 37                  2024 Week 26
#> 2                    2023 Week 37                  2024 Week 26
#>        latestVersion.geographicLevels
#> 1 Local authority, National, Regional
#> 2 Local authority, National, Regional
#>                                                         latestVersion.filters
#> 1 Attendance status, Attendance type, Day number, Establishment phase, Reason
#> 2 Attendance status, Attendance type, Day number, Establishment phase, Reason
#>                         latestVersion.indicators
#> 1 Session count, Session percent, Session scaled
#> 2 Session count, Session percent, Session scaled