Parse API meta to give the filter item codes

Parse API meta to give the filter item codes

Usage

parse_meta_filter_item_ids(api_meta_filters)

Arguments

api_meta_filters

Filter information provided by the API output

Value

Data frame containing filter item codes matched to filter item labels and col_name

Examples

get_meta_response(example_id())$filters |>
  parse_meta_filter_item_ids()
#>    col_name
#> 1     VTQ5X
#> 2     VTQ5X
#> 3     VTQ5X
#> 4     VTQ5X
#> 5     VTQ5X
#> 6     VTQ5X
#> 7     VTQ5X
#> 8     VTQ5X
#> 9     VTQ5X
#> 10    VTQ5X
#> 11    VTQ5X
#> 12    VTQ5X
#> 13    VTQ5X
#> 14    VTQ5X
#> 15    VTQ5X
#> 16    VTQ5X
#> 17    VTQ5X
#> 18    VTQ5X
#> 19    VTQ5X
#> 20    VTQ5X
#> 21    Pmjeb
#> 22    Pmjeb
#> 23    Pmjeb
#> 24    Pmjeb
#> 25    1Nei3
#> 26    1Nei3
#> 27    1Nei3
#> 28    1Nei3
#> 29    1Nei3
#> 30    1Nei3
#> 31    deq0Z
#> 32    deq0Z
#> 33    deq0Z
#> 34    deq0Z
#> 35    deq0Z
#> 36    deq0Z
#> 37    deq0Z
#>                                                                         item_label
#> 1                                                           Any other ethnic group
#> 2                                               Asian - Any other Asian background
#> 3                                                              Asian - Bangladeshi
#> 4                                                                   Asian - Indian
#> 5                                                                Asian - Pakistani
#> 6                                               Black - Any other Black background
#> 7                                                            Black - Black African
#> 8                                                          Black - Black Caribbean
#> 9                                                                          Chinese
#> 10                                              Mixed - Any other Mixed background
#> 11                                                         Mixed - White and Asian
#> 12                                                 Mixed - White and Black African
#> 13                                               Mixed - White and Black Caribbean
#> 14                                                                           Total
#> 15                                                                    Unclassified
#> 16                                              White - Any other White background
#> 17                                                              White - Gypsy/Roma
#> 18                                                                   White - Irish
#> 19                                             White - Traveller of Irish heritage
#> 20                                                           White - White British
#> 21                                                 Known or believed to be English
#> 22                                      Known or believed to be other than English
#> 23                                                           Language unclassified
#> 24                                                                           Total
#> 25                                    FSM eligible pupils taking free school meals
#> 26                                  infants taken a free school meal on census day
#> 27                                      known to be eligible for free school meals
#> 28 known to be eligible for free school meals (used for FSM in Performance Tables)
#> 29                           number of pupils (used for FSM in Performance Tables)
#> 30                                                                           Total
#> 31                                                   Non-maintained special school
#> 32                                                             Pupil referral unit
#> 33                                                            State-funded nursery
#> 34                                                            State-funded primary
#> 35                                                          State-funded secondary
#> 36                                                     State-funded special school
#> 37                                                                           Total
#>    item_id default_item
#> 1    bfQtT           NA
#> 2    DBNQe           NA
#> 3    e2euS           NA
#> 4    5y0di           NA
#> 5    4skUZ           NA
#> 6    m8R9K           NA
#> 7    0LoT5           NA
#> 8    WNZ2b           NA
#> 9    EzCqF           NA
#> 10   QXGwb           NA
#> 11   9ZW4v           NA
#> 12   2D1lz           NA
#> 13   P2yeb           NA
#> 14   NBODC         TRUE
#> 15   wf7bx           NA
#> 16   o3cmX           NA
#> 17   9ZZ4v           NA
#> 18   daL0Z           NA
#> 19   OS5CL           NA
#> 20   70eXo           NA
#> 21   pRaSo           NA
#> 22   I5eBz           NA
#> 23   iOf6X           NA
#> 24   LrEWj         TRUE
#> 25   6Czrf           NA
#> 26   0LxT5           NA
#> 27   HuazL           NA
#> 28   CV4Id           NA
#> 29   YeLHK           NA
#> 30   qFujG         TRUE
#> 31   a6OLP           NA
#> 32   Bda7J           NA
#> 33   zo8FQ           NA
#> 34   kKOhs           NA
#> 35   ucEo4           NA
#> 36   XT3fK           NA
#> 37   TWtPJ         TRUE