Parse IDs in a set of indicators

The API uses unique IDs (sqids) to identify each indicator column. This function parses those into the data creators' column names based on the meta data stored on the API for the data set.

Usage

parse_sqids_indicators(indicators, meta, verbose = FALSE)

Arguments

indicators

A set of indicator columns as taken from a data set downloaded from the API

meta

Meta data for the data set as provided by get_meta()

verbose

Run in verbose mode with debugging messages

Value

Data frame

Examples

example_data_raw(group = "attendance") |>
  magrittr::use_series("values") |>
  eesyapi:::parse_sqids_indicators(get_meta(example_id(group = "attendance")))
#>    session_count
#> 1       61554120
#> 2         516772
#> 3       62070892
#> 4        4648820
#> 5        2886610
#> 6        1762210
#> 7        1354113
#> 8       66719712
#> 9       29757304
#> 10      30567654
#> 11      60324958
#> 12       1229162
#> 13       2155394
#> 14        205372
#> 15          1596
#> 16            58
#> 17            63
#> 18        161417
#> 19        226475
#> 20          3023
#> 21          2312
#> 22        123676
#> 23        145513
#> 24             5
#> 25         57588
#> 26        268767
#> 27          8021
#> 28         36878
#> 29        326838
#> 30        124951
#> 31       1275659
#> 32         34762