Parse location sqids

The API uses unique IDs (sqids) to identify each location in a data set. This function parses those into the corresponding location codes and names based on the meta data stored on the API for the data set.

Usage

parse_sqids_locations(locations, meta, verbose = FALSE)

Arguments

locations

A set of location columns as taken from a data set downloaded from the API

meta

Meta data for the data set as provided by get_meta()

verbose

Run in verbose mode with debugging messages

Value

Data frame of parsed geography information

Examples

example_data_raw() |>
  magrittr::use_series("locations") |>
  parse_sqids_locations(get_meta(example_id(group = "attendance")))
#>     la_name   la_code la_oldCode nat_name  nat_code reg_name  reg_code
#> 1  Havering E09000016        311  England E92000001   London E12000007
#> 2  Havering E09000016        311  England E92000001   London E12000007
#> 3  Havering E09000016        311  England E92000001   London E12000007
#> 4  Havering E09000016        311  England E92000001   London E12000007
#> 5  Havering E09000016        311  England E92000001   London E12000007
#> 6  Havering E09000016        311  England E92000001   London E12000007
#> 7  Havering E09000016        311  England E92000001   London E12000007
#> 8  Havering E09000016        311  England E92000001   London E12000007
#> 9  Havering E09000016        311  England E92000001   London E12000007
#> 10 Havering E09000016        311  England E92000001   London E12000007
#> 11 Havering E09000016        311  England E92000001   London E12000007
#> 12 Havering E09000016        311  England E92000001   London E12000007
#> 13 Havering E09000016        311  England E92000001   London E12000007
#> 14 Havering E09000016        311  England E92000001   London E12000007
#> 15 Havering E09000016        311  England E92000001   London E12000007
#> 16 Havering E09000016        311  England E92000001   London E12000007
#> 17 Havering E09000016        311  England E92000001   London E12000007
#> 18 Havering E09000016        311  England E92000001   London E12000007
#> 19 Havering E09000016        311  England E92000001   London E12000007
#> 20 Havering E09000016        311  England E92000001   London E12000007
#> 21 Havering E09000016        311  England E92000001   London E12000007
#> 22 Havering E09000016        311  England E92000001   London E12000007
#> 23 Havering E09000016        311  England E92000001   London E12000007
#> 24 Havering E09000016        311  England E92000001   London E12000007
#> 25 Havering E09000016        311  England E92000001   London E12000007
#> 26 Havering E09000016        311  England E92000001   London E12000007
#> 27 Havering E09000016        311  England E92000001   London E12000007
#> 28 Havering E09000016        311  England E92000001   London E12000007
#> 29 Havering E09000016        311  England E92000001   London E12000007
#> 30 Havering E09000016        311  England E92000001   London E12000007
#> 31 Havering E09000016        311  England E92000001   London E12000007
#> 32 Havering E09000016        311  England E92000001   London E12000007