Example raw data

Download some example raw data. Mainly intended for use in developing / testing the sqid parsing or as an example of getting raw data if any end users would prefer to do the sqid parsing themselves.

Usage

example_data_raw(group = "attendance", size = 32)

Arguments

group

Choose the publication group of examples to use. Options are:

"attendance": Large example data set, careful what you ask for

"public-api-testing": Smaller example data set

size

Number of rows to return (max = 1000)

Value

Nested list form of example data from the API

Examples

example_data_raw()
#>    timePeriod.code timePeriod.period geographicLevel locations.LA locations.NAT
#> 1              W25              2024              LA        5Zdi9         mRj9K
#> 2              W25              2024              LA        5Zdi9         mRj9K
#> 3              W25              2024              LA        5Zdi9         mRj9K
#> 4              W25              2024              LA        5Zdi9         mRj9K
#> 5              W25              2024              LA        5Zdi9         mRj9K
#> 6              W25              2024              LA        5Zdi9         mRj9K
#> 7              W25              2024              LA        5Zdi9         mRj9K
#> 8              W25              2024              LA        5Zdi9         mRj9K
#> 9              W25              2024              LA        5Zdi9         mRj9K
#> 10             W25              2024              LA        5Zdi9         mRj9K
#> 11             W25              2024              LA        5Zdi9         mRj9K
#> 12             W25              2024              LA        5Zdi9         mRj9K
#> 13             W25              2024              LA        5Zdi9         mRj9K
#> 14             W25              2024              LA        5Zdi9         mRj9K
#> 15             W25              2024              LA        5Zdi9         mRj9K
#> 16             W25              2024              LA        5Zdi9         mRj9K
#> 17             W25              2024              LA        5Zdi9         mRj9K
#> 18             W25              2024              LA        5Zdi9         mRj9K
#> 19             W25              2024              LA        5Zdi9         mRj9K
#> 20             W25              2024              LA        5Zdi9         mRj9K
#> 21             W25              2024              LA        5Zdi9         mRj9K
#> 22             W25              2024              LA        5Zdi9         mRj9K
#> 23             W25              2024              LA        5Zdi9         mRj9K
#> 24             W25              2024              LA        5Zdi9         mRj9K
#> 25             W25              2024              LA        5Zdi9         mRj9K
#> 26             W25              2024              LA        5Zdi9         mRj9K
#> 27             W25              2024              LA        5Zdi9         mRj9K
#> 28             W25              2024              LA        5Zdi9         mRj9K
#> 29             W25              2024              LA        5Zdi9         mRj9K
#> 30             W25              2024              LA        5Zdi9         mRj9K
#> 31             W25              2024              LA        5Zdi9         mRj9K
#> 32             W25              2024              LA        5Zdi9         mRj9K
#>    locations.REG filters.juAMt filters.eNuSW filters.tdEm5 filters.5Zdi9
#> 1          lZVcB         1rPi3         9Ru4v         tj0Em         BfP7J
#> 2          lZVcB         1rPi3         9Ru4v         TuxPJ         BfP7J
#> 3          lZVcB         1rPi3         9Ru4v         fzaYF         BfP7J
#> 4          lZVcB         1rPi3         9Ru4v         fzaYF         qGJjG
#> 5          lZVcB         1rPi3         9Ru4v         cZO31         qGJjG
#> 6          lZVcB         1rPi3         9Ru4v         jgoAM         qGJjG
#> 7          lZVcB         1rPi3         9Ru4v         fzaYF         zvUFQ
#> 8          lZVcB         1rPi3         9Ru4v         fzaYF         u9Mo4
#> 9          lZVcB         1rPi3         mU59K         tj0Em         BfP7J
#> 10         lZVcB         1rPi3         0evT5         tj0Em         BfP7J
#> 11         lZVcB         1rPi3         4kdUZ         tj0Em         BfP7J
#> 12         lZVcB         1rPi3         yW5aB         5Tsdi         BfP7J
#> 13         lZVcB         1rPi3         3se8u         cZO31         qGJjG
#> 14         lZVcB         1rPi3         RySka         cZO31         qGJjG
#> 15         lZVcB         1rPi3         WcR2b         cZO31         qGJjG
#> 16         lZVcB         1rPi3         EADqF         cZO31         qGJjG
#> 17         lZVcB         1rPi3         Qt6wb         cZO31         qGJjG
#> 18         lZVcB         1rPi3         GCsgr         cZO31         qGJjG
#> 19         lZVcB         1rPi3         o47mX         cZO31         qGJjG
#> 20         lZVcB         1rPi3         l8ScB         cZO31         qGJjG
#> 21         lZVcB         1rPi3         u9Oo4         TuxPJ         BfP7J
#> 22         lZVcB         1rPi3         UhYRF         fzaYF         kaNhs
#> 23         lZVcB         1rPi3         5TYdi         TuxPJ         BfP7J
#> 24         lZVcB         1rPi3         PoTeb         TuxPJ         BfP7J
#> 25         lZVcB         1rPi3         NwNDC         TuxPJ         BfP7J
#> 26         lZVcB         1rPi3         rQkNj         jgoAM         qGJjG
#> 27         lZVcB         1rPi3         23Llz         jgoAM         qGJjG
#> 28         lZVcB         1rPi3         emJuS         jgoAM         qGJjG
#> 29         lZVcB         1rPi3         bqZtT         jgoAM         qGJjG
#> 30         lZVcB         1rPi3         Db3Qe         XH4fK         a2NLP
#> 31         lZVcB         1rPi3         wqtbx         fzaYF         kaNhs
#> 32         lZVcB         1rPi3         oU4mX         XH4fK         a2NLP
#>    filters.fyYFZ tj0Em
#> 1          emDuS   438
#> 2          emDuS    22
#> 3          emDuS   460
#> 4          emDuS    78
#> 5          emDuS    50
#> 6          emDuS    28
#> 7          emDuS     9
#> 8          emDuS   538
#> 9          emDuS   214
#> 10         emDuS   219
#> 11         emDuS   433
#> 12         emDuS     5
#> 13         emDuS    24
#> 14         emDuS     6
#> 15         emDuS     0
#> 16         emDuS     0
#> 17         emDuS     0
#> 18         emDuS    14
#> 19         emDuS     0
#> 20         emDuS     6
#> 21         emDuS    15
#> 22         emDuS     0
#> 23         emDuS     0
#> 24         emDuS     7
#> 25         emDuS     0
#> 26         emDuS     8
#> 27         emDuS     4
#> 28         emDuS    16
#> 29         emDuS     0
#> 30         emDuS     0
#> 31         emDuS     0
#> 32         emDuS     0