Example raw data

Download some example raw data. Mainly intended for use in developing / testing the sqid parsing or as an example of getting raw data if any end users would prefer to do the sqid parsing themselves.

Usage

example_data_raw(group = "attendance", size = 32)

Arguments

group

Choose the publication group of examples to use. Options are:

"attendance": Large example data set, careful what you ask for

"public-api-testing": Smaller example data set

size

Number of rows to return (max = 1000)

Value

Nested list form of example data from the API

Examples

example_data_raw()
#>    timePeriod.code timePeriod.period geographicLevel locations.LA locations.NAT
#> 1              W25              2024              LA        5Zdi9         mRj9K
#> 2              W25              2024              LA        5Zdi9         mRj9K
#> 3              W25              2024              LA        5Zdi9         mRj9K
#> 4              W25              2024              LA        5Zdi9         mRj9K
#> 5              W25              2024              LA        5Zdi9         mRj9K
#> 6              W25              2024              LA        5Zdi9         mRj9K
#> 7              W25              2024              LA        5Zdi9         mRj9K
#> 8              W25              2024              LA        5Zdi9         mRj9K
#> 9              W25              2024              LA        5Zdi9         mRj9K
#> 10             W25              2024              LA        5Zdi9         mRj9K
#> 11             W25              2024              LA        5Zdi9         mRj9K
#> 12             W25              2024              LA        5Zdi9         mRj9K
#> 13             W25              2024              LA        5Zdi9         mRj9K
#> 14             W25              2024              LA        5Zdi9         mRj9K
#> 15             W25              2024              LA        5Zdi9         mRj9K
#> 16             W25              2024              LA        5Zdi9         mRj9K
#> 17             W25              2024              LA        5Zdi9         mRj9K
#> 18             W25              2024              LA        5Zdi9         mRj9K
#> 19             W25              2024              LA        5Zdi9         mRj9K
#> 20             W25              2024              LA        5Zdi9         mRj9K
#> 21             W25              2024              LA        5Zdi9         mRj9K
#> 22             W25              2024              LA        5Zdi9         mRj9K
#> 23             W25              2024              LA        5Zdi9         mRj9K
#> 24             W25              2024              LA        5Zdi9         mRj9K
#> 25             W25              2024              LA        5Zdi9         mRj9K
#> 26             W25              2024              LA        5Zdi9         mRj9K
#> 27             W25              2024              LA        5Zdi9         mRj9K
#> 28             W25              2024              LA        5Zdi9         mRj9K
#> 29             W25              2024              LA        5Zdi9         mRj9K
#> 30             W25              2024              LA        5Zdi9         mRj9K
#> 31             W25              2024              LA        5Zdi9         mRj9K
#> 32             W25              2024              LA        5Zdi9         mRj9K
#>    locations.REG filters.juAMt filters.eNuSW filters.5Zdi9 filters.fyYFZ
#> 1          lZVcB         1rPi3         9Ru4v         BfP7J         emDuS
#> 2          lZVcB         1rPi3         9Ru4v         BfP7J         emDuS
#> 3          lZVcB         1rPi3         9Ru4v         BfP7J         emDuS
#> 4          lZVcB         1rPi3         9Ru4v         qGJjG         emDuS
#> 5          lZVcB         1rPi3         9Ru4v         qGJjG         emDuS
#> 6          lZVcB         1rPi3         9Ru4v         qGJjG         emDuS
#> 7          lZVcB         1rPi3         9Ru4v         zvUFQ         emDuS
#> 8          lZVcB         1rPi3         9Ru4v         u9Mo4         emDuS
#> 9          lZVcB         1rPi3         mU59K         BfP7J         emDuS
#> 10         lZVcB         1rPi3         0evT5         BfP7J         emDuS
#> 11         lZVcB         1rPi3         4kdUZ         BfP7J         emDuS
#> 12         lZVcB         1rPi3         yW5aB         BfP7J         emDuS
#> 13         lZVcB         1rPi3         3se8u         qGJjG         emDuS
#> 14         lZVcB         1rPi3         RySka         qGJjG         emDuS
#> 15         lZVcB         1rPi3         WcR2b         qGJjG         emDuS
#> 16         lZVcB         1rPi3         EADqF         qGJjG         emDuS
#> 17         lZVcB         1rPi3         Qt6wb         qGJjG         emDuS
#> 18         lZVcB         1rPi3         GCsgr         qGJjG         emDuS
#> 19         lZVcB         1rPi3         o47mX         qGJjG         emDuS
#> 20         lZVcB         1rPi3         l8ScB         qGJjG         emDuS
#> 21         lZVcB         1rPi3         u9Oo4         BfP7J         emDuS
#> 22         lZVcB         1rPi3         UhYRF         kaNhs         emDuS
#> 23         lZVcB         1rPi3         5TYdi         BfP7J         emDuS
#> 24         lZVcB         1rPi3         PoTeb         BfP7J         emDuS
#> 25         lZVcB         1rPi3         NwNDC         BfP7J         emDuS
#> 26         lZVcB         1rPi3         rQkNj         qGJjG         emDuS
#> 27         lZVcB         1rPi3         23Llz         qGJjG         emDuS
#> 28         lZVcB         1rPi3         emJuS         qGJjG         emDuS
#> 29         lZVcB         1rPi3         bqZtT         qGJjG         emDuS
#> 30         lZVcB         1rPi3         Db3Qe         a2NLP         emDuS
#> 31         lZVcB         1rPi3         wqtbx         kaNhs         emDuS
#> 32         lZVcB         1rPi3         oU4mX         a2NLP         emDuS
#>    filters.tdEm5 tj0Em
#> 1          tj0Em   438
#> 2          TuxPJ    22
#> 3          fzaYF   460
#> 4          fzaYF    78
#> 5          cZO31    50
#> 6          jgoAM    28
#> 7          fzaYF     9
#> 8          fzaYF   538
#> 9          tj0Em   214
#> 10         tj0Em   219
#> 11         tj0Em   433
#> 12         5Tsdi     5
#> 13         cZO31    24
#> 14         cZO31     6
#> 15         cZO31     0
#> 16         cZO31     0
#> 17         cZO31     0
#> 18         cZO31    14
#> 19         cZO31     0
#> 20         cZO31     6
#> 21         TuxPJ    15
#> 22         fzaYF     0
#> 23         TuxPJ     0
#> 24         TuxPJ     7
#> 25         TuxPJ     0
#> 26         jgoAM     8
#> 27         jgoAM     4
#> 28         jgoAM    16
#> 29         jgoAM     0
#> 30         XH4fK     0
#> 31         fzaYF     0
#> 32         XH4fK     0