Example raw data

Download some example raw data. Mainly intended for use in developing / testing the sqid parsing or as an example of getting raw data if any end users would prefer to do the sqid parsing themselves.

Usage

example_data_raw(group = "attendance", size = 32)

Arguments

group

Choose the publication group of examples to use. Can be "attendance".

size

Number of rows to return (max = 1000)

Value

Nested list form of example data from the API

Examples

example_data_raw()
#>    timePeriod.code timePeriod.period geographicLevel locations.LA locations.NAT
#> 1              W26              2024              LA        9cy4v         dP0Zw
#> 2              W26              2024              LA        9cy4v         dP0Zw
#> 3              W26              2024              LA        9cy4v         dP0Zw
#> 4              W26              2024              LA        9cy4v         dP0Zw
#> 5              W26              2024              LA        9cy4v         dP0Zw
#> 6              W26              2024              LA        9cy4v         dP0Zw
#> 7              W26              2024              LA        9cy4v         dP0Zw
#> 8              W26              2024              LA        9cy4v         dP0Zw
#> 9              W26              2024              LA        9cy4v         dP0Zw
#> 10             W26              2024              LA        9cy4v         dP0Zw
#> 11             W26              2024              LA        9cy4v         dP0Zw
#> 12             W26              2024              LA        9cy4v         dP0Zw
#> 13             W26              2024              LA        9cy4v         dP0Zw
#> 14             W26              2024              LA        9cy4v         dP0Zw
#> 15             W26              2024              LA        9cy4v         dP0Zw
#> 16             W26              2024              LA        9cy4v         dP0Zw
#> 17             W26              2024              LA        9cy4v         dP0Zw
#> 18             W26              2024              LA        9cy4v         dP0Zw
#> 19             W26              2024              LA        9cy4v         dP0Zw
#> 20             W26              2024              LA        9cy4v         dP0Zw
#> 21             W26              2024              LA        9cy4v         dP0Zw
#> 22             W26              2024              LA        9cy4v         dP0Zw
#> 23             W26              2024              LA        9cy4v         dP0Zw
#> 24             W26              2024              LA        9cy4v         dP0Zw
#> 25             W26              2024              LA        9cy4v         dP0Zw
#> 26             W26              2024              LA        9cy4v         dP0Zw
#> 27             W26              2024              LA        9cy4v         dP0Zw
#> 28             W26              2024              LA        9cy4v         dP0Zw
#> 29             W26              2024              LA        9cy4v         dP0Zw
#> 30             W26              2024              LA        9cy4v         dP0Zw
#> 31             W26              2024              LA        9cy4v         dP0Zw
#> 32             W26              2024              LA        9cy4v         dP0Zw
#>    locations.REG filters.Db3Qe filters.emJuS filters.4kdUZ filters.5TYdi
#> 1          S4HVx         pmRSo         CvuId         XVWfK         qf0jG
#> 2          S4HVx         pmRSo         6AXrf         XVWfK         qf0jG
#> 3          S4HVx         pmRSo         H0RzL         XVWfK         qf0jG
#> 4          S4HVx         7SdXo         H0RzL         XVWfK         qf0jG
#> 5          S4HVx         7SdXo         0k3T5         XVWfK         qf0jG
#> 6          S4HVx         7SdXo         YdkHK         XVWfK         qf0jG
#> 7          S4HVx         IndBz         H0RzL         XVWfK         qf0jG
#> 8          S4HVx         LKQWj         H0RzL         XVWfK         qf0jG
#> 9          S4HVx         pmRSo         CvuId         XVWfK         qf0jG
#> 10         S4HVx         pmRSo         CvuId         XVWfK         qf0jG
#> 11         S4HVx         pmRSo         CvuId         XVWfK         qf0jG
#> 12         S4HVx         pmRSo         CvuId         XVWfK         qf0jG
#> 13         S4HVx         7SdXo         0k3T5         XVWfK         qf0jG
#> 14         S4HVx         7SdXo         0k3T5         XVWfK         qf0jG
#> 15         S4HVx         7SdXo         0k3T5         XVWfK         qf0jG
#> 16         S4HVx         7SdXo         0k3T5         XVWfK         qf0jG
#> 17         S4HVx         7SdXo         0k3T5         XVWfK         qf0jG
#> 18         S4HVx         7SdXo         0k3T5         XVWfK         qf0jG
#> 19         S4HVx         7SdXo         0k3T5         XVWfK         qf0jG
#> 20         S4HVx         7SdXo         0k3T5         XVWfK         qf0jG
#> 21         S4HVx         pmRSo         6AXrf         XVWfK         qf0jG
#> 22         S4HVx         pmRSo         6AXrf         XVWfK         qf0jG
#> 23         S4HVx         pmRSo         6AXrf         XVWfK         qf0jG
#> 24         S4HVx         pmRSo         6AXrf         XVWfK         qf0jG
#> 25         S4HVx         pmRSo         6AXrf         XVWfK         qf0jG
#> 26         S4HVx         7SdXo         YdkHK         XVWfK         qf0jG
#> 27         S4HVx         7SdXo         YdkHK         XVWfK         qf0jG
#> 28         S4HVx         7SdXo         YdkHK         XVWfK         qf0jG
#> 29         S4HVx         ijI6X         H0RzL         XVWfK         qf0jG
#> 30         S4HVx         ijI6X         H0RzL         XVWfK         qf0jG
#> 31         S4HVx         ijI6X         H0RzL         XVWfK         qf0jG
#> 32         S4HVx         ijI6X         H0RzL         XVWfK         qf0jG
#>    filters.mU59K      bqZtT
#> 1          bBrtT      32342
#> 2          bBrtT        605
#> 3          bBrtT      32947
#> 4          bBrtT       1955
#> 5          bBrtT       1455
#> 6          bBrtT        500
#> 7          bBrtT        404
#> 8          bBrtT      34902
#> 9          GzNgr      15783
#> 10         3Nk8u      16180
#> 11         rGcNj      31963
#> 12         PyBeb        379
#> 13         AH3GK       1153
#> 14         1qRi3         84
#> 15         S9YVx          4
#> 16         yxmaB          0
#> 17         R8uka         14
#> 18         gF8O9         30
#> 19         eaAuS          3
#> 20         f5ZYF        167
#> 21         t74Em        120
#> 22         jYyAM          0
#> 23         oWvmX         16
#> 24         ea0uS        383
#> 25         5USdi          0
#> 26         m2B9K        241
#> 27         DOiQe         25
#> 28         l6ucB        192
#> 29         dRI0Z         42
#> 30         uLHo4          0
#> 31         4c9UZ          2
#> 32         m2m9K        124