Example raw data

Download some example raw data. Mainly intended for use in developing / testing the sqid parsing or as an example of getting raw data if any end users would prefer to do the sqid parsing themselves.

Usage

example_data_raw(group = "attendance", size = 32)

Arguments

group

Choose the publication group of examples to use. Can be "attendance".

size

Number of rows to return (max = 1000)

Value

Nested list form of example data from the API

Examples

example_data_raw()
#>    timePeriod.code timePeriod.period geographicLevel locations.LA locations.NAT
#> 1              W26              2024              LA        9cy4v         dP0Zw
#> 2              W26              2024              LA        9cy4v         dP0Zw
#> 3              W26              2024              LA        9cy4v         dP0Zw
#> 4              W26              2024              LA        9cy4v         dP0Zw
#> 5              W26              2024              LA        9cy4v         dP0Zw
#> 6              W26              2024              LA        9cy4v         dP0Zw
#> 7              W26              2024              LA        9cy4v         dP0Zw
#> 8              W26              2024              LA        9cy4v         dP0Zw
#> 9              W26              2024              LA        9cy4v         dP0Zw
#> 10             W26              2024              LA        9cy4v         dP0Zw
#> 11             W26              2024              LA        9cy4v         dP0Zw
#> 12             W26              2024              LA        9cy4v         dP0Zw
#> 13             W26              2024              LA        9cy4v         dP0Zw
#> 14             W26              2024              LA        9cy4v         dP0Zw
#> 15             W26              2024              LA        9cy4v         dP0Zw
#> 16             W26              2024              LA        9cy4v         dP0Zw
#> 17             W26              2024              LA        9cy4v         dP0Zw
#> 18             W26              2024              LA        9cy4v         dP0Zw
#> 19             W26              2024              LA        9cy4v         dP0Zw
#> 20             W26              2024              LA        9cy4v         dP0Zw
#> 21             W26              2024              LA        9cy4v         dP0Zw
#> 22             W26              2024              LA        9cy4v         dP0Zw
#> 23             W26              2024              LA        9cy4v         dP0Zw
#> 24             W26              2024              LA        9cy4v         dP0Zw
#> 25             W26              2024              LA        9cy4v         dP0Zw
#> 26             W26              2024              LA        9cy4v         dP0Zw
#> 27             W26              2024              LA        9cy4v         dP0Zw
#> 28             W26              2024              LA        9cy4v         dP0Zw
#> 29             W26              2024              LA        9cy4v         dP0Zw
#> 30             W26              2024              LA        9cy4v         dP0Zw
#> 31             W26              2024              LA        9cy4v         dP0Zw
#> 32             W26              2024              LA        9cy4v         dP0Zw
#>    locations.REG filters.5TYdi filters.mU59K filters.Db3Qe filters.emJuS
#> 1          S4HVx         qf0jG         bBrtT         pmRSo         CvuId
#> 2          S4HVx         qf0jG         bBrtT         pmRSo         6AXrf
#> 3          S4HVx         qf0jG         bBrtT         pmRSo         H0RzL
#> 4          S4HVx         qf0jG         bBrtT         7SdXo         H0RzL
#> 5          S4HVx         qf0jG         bBrtT         7SdXo         0k3T5
#> 6          S4HVx         qf0jG         bBrtT         7SdXo         YdkHK
#> 7          S4HVx         qf0jG         bBrtT         IndBz         H0RzL
#> 8          S4HVx         qf0jG         bBrtT         LKQWj         H0RzL
#> 9          S4HVx         qf0jG         GzNgr         pmRSo         CvuId
#> 10         S4HVx         qf0jG         3Nk8u         pmRSo         CvuId
#> 11         S4HVx         qf0jG         rGcNj         pmRSo         CvuId
#> 12         S4HVx         qf0jG         PyBeb         pmRSo         CvuId
#> 13         S4HVx         qf0jG         AH3GK         7SdXo         0k3T5
#> 14         S4HVx         qf0jG         1qRi3         7SdXo         0k3T5
#> 15         S4HVx         qf0jG         S9YVx         7SdXo         0k3T5
#> 16         S4HVx         qf0jG         yxmaB         7SdXo         0k3T5
#> 17         S4HVx         qf0jG         R8uka         7SdXo         0k3T5
#> 18         S4HVx         qf0jG         gF8O9         7SdXo         0k3T5
#> 19         S4HVx         qf0jG         eaAuS         7SdXo         0k3T5
#> 20         S4HVx         qf0jG         f5ZYF         7SdXo         0k3T5
#> 21         S4HVx         qf0jG         t74Em         pmRSo         6AXrf
#> 22         S4HVx         qf0jG         jYyAM         pmRSo         6AXrf
#> 23         S4HVx         qf0jG         oWvmX         pmRSo         6AXrf
#> 24         S4HVx         qf0jG         ea0uS         pmRSo         6AXrf
#> 25         S4HVx         qf0jG         5USdi         pmRSo         6AXrf
#> 26         S4HVx         qf0jG         m2B9K         7SdXo         YdkHK
#> 27         S4HVx         qf0jG         DOiQe         7SdXo         YdkHK
#> 28         S4HVx         qf0jG         l6ucB         7SdXo         YdkHK
#> 29         S4HVx         qf0jG         dRI0Z         ijI6X         H0RzL
#> 30         S4HVx         qf0jG         uLHo4         ijI6X         H0RzL
#> 31         S4HVx         qf0jG         4c9UZ         ijI6X         H0RzL
#> 32         S4HVx         qf0jG         m2m9K         ijI6X         H0RzL
#>    filters.4kdUZ      bqZtT
#> 1          XVWfK      32342
#> 2          XVWfK        605
#> 3          XVWfK      32947
#> 4          XVWfK       1955
#> 5          XVWfK       1455
#> 6          XVWfK        500
#> 7          XVWfK        404
#> 8          XVWfK      34902
#> 9          XVWfK      15783
#> 10         XVWfK      16180
#> 11         XVWfK      31963
#> 12         XVWfK        379
#> 13         XVWfK       1153
#> 14         XVWfK         84
#> 15         XVWfK          4
#> 16         XVWfK          0
#> 17         XVWfK         14
#> 18         XVWfK         30
#> 19         XVWfK          3
#> 20         XVWfK        167
#> 21         XVWfK        120
#> 22         XVWfK          0
#> 23         XVWfK         16
#> 24         XVWfK        383
#> 25         XVWfK          0
#> 26         XVWfK        241
#> 27         XVWfK         25
#> 28         XVWfK        192
#> 29         XVWfK         42
#> 30         XVWfK          0
#> 31         XVWfK          2
#> 32         XVWfK        124