Example raw data

Download some example raw data. Mainly intended for use in developing / testing the sqid parsing or as an example of getting raw data if any end users would prefer to do the sqid parsing themselves.

Usage

example_data_raw(group = "attendance", size = 32)

Arguments

group

Choose the publication group of examples to use. Options are:

"attendance": Large example data set, careful what you ask for

"public-api-testing": Smaller example data set

size

Number of rows to return (max = 1000)

Value

Nested list form of example data from the API

Examples

example_data_raw()
#>    timePeriod.code timePeriod.period geographicLevel   NAT filters.X9fKb
#> 1              W14              2025             NAT dP0Zw         rbyNj
#> 2              W14              2025             NAT dP0Zw         rbyNj
#> 3              W14              2025             NAT dP0Zw         rbyNj
#> 4              W14              2025             NAT dP0Zw         rbyNj
#> 5              W14              2025             NAT dP0Zw         rbyNj
#> 6              W14              2025             NAT dP0Zw         rbyNj
#> 7              W14              2025             NAT dP0Zw         rbyNj
#> 8              W14              2025             NAT dP0Zw         rbyNj
#> 9              W14              2025             NAT dP0Zw         rbyNj
#> 10             W14              2025             NAT dP0Zw         rbyNj
#> 11             W14              2025             NAT dP0Zw         rbyNj
#> 12             W14              2025             NAT dP0Zw         rbyNj
#> 13             W14              2025             NAT dP0Zw         rbyNj
#> 14             W14              2025             NAT dP0Zw         rbyNj
#> 15             W14              2025             NAT dP0Zw         rbyNj
#> 16             W14              2025             NAT dP0Zw         rbyNj
#> 17             W14              2025             NAT dP0Zw         rbyNj
#> 18             W14              2025             NAT dP0Zw         rbyNj
#> 19             W14              2025             NAT dP0Zw         rbyNj
#> 20             W14              2025             NAT dP0Zw         rbyNj
#> 21             W14              2025             NAT dP0Zw         rbyNj
#> 22             W14              2025             NAT dP0Zw         rbyNj
#> 23             W14              2025             NAT dP0Zw         rbyNj
#> 24             W14              2025             NAT dP0Zw         rbyNj
#> 25             W14              2025             NAT dP0Zw         rbyNj
#> 26             W14              2025             NAT dP0Zw         rbyNj
#> 27             W14              2025             NAT dP0Zw         rbyNj
#> 28             W14              2025             NAT dP0Zw         rbyNj
#> 29             W14              2025             NAT dP0Zw         rbyNj
#> 30             W14              2025             NAT dP0Zw         rbyNj
#> 31             W14              2025             NAT dP0Zw         rbyNj
#> 32             W14              2025             NAT dP0Zw         rbyNj
#>    filters.krhsL filters.uxo41 filters.TsPJP filters.z4FQE      X9fKb
#> 1          TmQPJ         UazRF         4yTUZ         bjstT 1619003783
#> 2          TmQPJ         P9Aeb         4yTUZ         3xu8u   14815860
#> 3          TmQPJ         ls5cB         4yTUZ         HnuzL 1633819643
#> 4          e4wuS         uUIo4         4yTUZ         0oaT5  118548447
#> 5          e4wuS         1Dui3         4yTUZ         S0OVx   80814242
#> 6          e4wuS         oRfmX         4yTUZ         Dt8Qe   37734205
#> 7          h2IyW         AIaGK         4yTUZ         y2daB   37028063
#> 8          g7LO9         VPw5X         4yTUZ         RaVka 1752368090
#> 9          TmQPJ         UazRF         4yTUZ         qOnjG  782878151
#> 10         TmQPJ         UazRF         4yTUZ         BHu7J  802942904
#> 11         TmQPJ         UazRF         4yTUZ         YvoHK 1585821055
#> 12         TmQPJ         UazRF         4yTUZ         acQLP   33182728
#> 13         e4wuS         1Dui3         4yTUZ         28Wlz   60111966
#> 14         e4wuS         1Dui3         4yTUZ         duo0Z    5569344
#> 15         e4wuS         1Dui3         4yTUZ         uU1o4    1211933
#> 16         e4wuS         1Dui3         4yTUZ         XFRfK      36993
#> 17         e4wuS         1Dui3         4yTUZ         5PHdi       1715
#> 18         e4wuS         1Dui3         4yTUZ         9ko4v    2731162
#> 19         e4wuS         1Dui3         4yTUZ         L31Wj    6872530
#> 20         e4wuS         1Dui3         4yTUZ         kWQhs      67637
#> 21         e4wuS         1Dui3         4yTUZ         NtQDC     211774
#> 22         e4wuS         1Dui3         4yTUZ         TmKPJ    3871881
#> 23         TmQPJ         P9Aeb         4yTUZ         QErwb    3417207
#> 24         TmQPJ         P9Aeb         4yTUZ         PmPeb        553
#> 25         TmQPJ         P9Aeb         4yTUZ         e9vuS    1479929
#> 26         TmQPJ         P9Aeb         4yTUZ         W3k2b    8624874
#> 27         TmQPJ         P9Aeb         4yTUZ         jC9AM     668876
#> 28         TmQPJ         P9Aeb         4yTUZ         EGQqF     624421
#> 29         e4wuS         oRfmX         4yTUZ         tGPEm    6361189
#> 30         e4wuS         oRfmX         4yTUZ         oRUmX    3845335
#> 31         e4wuS         oRfmX         4yTUZ         6wgrf   26318365
#> 32         e4wuS         oRfmX         4yTUZ         7zvXo    1209316