Example raw data

Download some example raw data. Mainly intended for use in developing / testing the sqid parsing or as an example of getting raw data if any end users would prefer to do the sqid parsing themselves.

Usage

example_data_raw(group = "attendance", size = 32)

Arguments

group

Choose the publication group of examples to use. Options are:

"attendance": Large example data set, careful what you ask for

"public-api-testing": Smaller example data set

size

Number of rows to return (max = 1000)

Value

Nested list form of example data from the API

Examples

example_data_raw()
#>    timePeriod.code timePeriod.period geographicLevel locations.LA locations.NAT
#> 1              W21              2025              LA        6DArf         dP0Zw
#> 2              W21              2025              LA        6DArf         dP0Zw
#> 3              W21              2025              LA        6DArf         dP0Zw
#> 4              W21              2025              LA        6DArf         dP0Zw
#> 5              W21              2025              LA        6DArf         dP0Zw
#> 6              W21              2025              LA        6DArf         dP0Zw
#> 7              W21              2025              LA        6DArf         dP0Zw
#> 8              W21              2025              LA        6DArf         dP0Zw
#> 9              W21              2025              LA        6DArf         dP0Zw
#> 10             W21              2025              LA        6DArf         dP0Zw
#> 11             W21              2025              LA        6DArf         dP0Zw
#> 12             W21              2025              LA        6DArf         dP0Zw
#> 13             W21              2025              LA        6DArf         dP0Zw
#> 14             W21              2025              LA        6DArf         dP0Zw
#> 15             W21              2025              LA        6DArf         dP0Zw
#> 16             W21              2025              LA        6DArf         dP0Zw
#> 17             W21              2025              LA        6DArf         dP0Zw
#> 18             W21              2025              LA        6DArf         dP0Zw
#> 19             W21              2025              LA        6DArf         dP0Zw
#> 20             W21              2025              LA        6DArf         dP0Zw
#> 21             W21              2025              LA        6DArf         dP0Zw
#> 22             W21              2025              LA        6DArf         dP0Zw
#> 23             W21              2025              LA        6DArf         dP0Zw
#> 24             W21              2025              LA        6DArf         dP0Zw
#> 25             W21              2025              LA        6DArf         dP0Zw
#> 26             W21              2025              LA        6DArf         dP0Zw
#> 27             W21              2025              LA        6DArf         dP0Zw
#> 28             W21              2025              LA        6DArf         dP0Zw
#> 29             W21              2025              LA        6DArf         dP0Zw
#> 30             W21              2025              LA        6DArf         dP0Zw
#> 31             W21              2025              LA        6DArf         dP0Zw
#> 32             W21              2025              LA        6DArf         dP0Zw
#>    locations.REG filters.X9fKb filters.TsPJP filters.z4FQE filters.krhsL
#> 1          VTQ5X         GBMgr         5ezdi         iQj6X         mkA9K
#> 2          VTQ5X         GBMgr         5ezdi         IYvBz         mkA9K
#> 3          VTQ5X         GBMgr         5ezdi         9hh4v         mkA9K
#> 4          VTQ5X         3Mo8u         4yTUZ         bjstT         TmQPJ
#> 5          VTQ5X         3Mo8u         4yTUZ         3xu8u         TmQPJ
#> 6          VTQ5X         3Mo8u         4yTUZ         HnuzL         TmQPJ
#> 7          VTQ5X         3Mo8u         4yTUZ         0oaT5         e4wuS
#> 8          VTQ5X         3Mo8u         4yTUZ         S0OVx         e4wuS
#> 9          VTQ5X         3Mo8u         4yTUZ         Dt8Qe         e4wuS
#> 10         VTQ5X         3Mo8u         4yTUZ         y2daB         h2IyW
#> 11         VTQ5X         3Mo8u         4yTUZ         RaVka         g7LO9
#> 12         VTQ5X         3Mo8u         4yTUZ         qOnjG         TmQPJ
#> 13         VTQ5X         3Mo8u         4yTUZ         BHu7J         TmQPJ
#> 14         VTQ5X         3Mo8u         4yTUZ         YvoHK         TmQPJ
#> 15         VTQ5X         3Mo8u         4yTUZ         acQLP         TmQPJ
#> 16         VTQ5X         3Mo8u         4yTUZ         28Wlz         e4wuS
#> 17         VTQ5X         3Mo8u         4yTUZ         duo0Z         e4wuS
#> 18         VTQ5X         3Mo8u         4yTUZ         uU1o4         e4wuS
#> 19         VTQ5X         3Mo8u         4yTUZ         XFRfK         e4wuS
#> 20         VTQ5X         3Mo8u         4yTUZ         5PHdi         e4wuS
#> 21         VTQ5X         3Mo8u         4yTUZ         9ko4v         e4wuS
#> 22         VTQ5X         3Mo8u         4yTUZ         L31Wj         e4wuS
#> 23         VTQ5X         3Mo8u         4yTUZ         kWQhs         e4wuS
#> 24         VTQ5X         3Mo8u         4yTUZ         NtQDC         e4wuS
#> 25         VTQ5X         3Mo8u         4yTUZ         TmKPJ         e4wuS
#> 26         VTQ5X         3Mo8u         4yTUZ         QErwb         TmQPJ
#> 27         VTQ5X         3Mo8u         4yTUZ         PmPeb         TmQPJ
#> 28         VTQ5X         3Mo8u         4yTUZ         e9vuS         TmQPJ
#> 29         VTQ5X         3Mo8u         4yTUZ         W3k2b         TmQPJ
#> 30         VTQ5X         3Mo8u         4yTUZ         jC9AM         TmQPJ
#> 31         VTQ5X         3Mo8u         4yTUZ         EGQqF         TmQPJ
#> 32         VTQ5X         3Mo8u         4yTUZ         tGPEm         e4wuS
#>    filters.uxo41   X9fKb
#> 1          duj0Z      54
#> 2          duj0Z    6018
#> 3          duj0Z       0
#> 4          UazRF 3580704
#> 5          P9Aeb   34563
#> 6          ls5cB 3615267
#> 7          uUIo4  368853
#> 8          1Dui3  171953
#> 9          oRfmX  196900
#> 10         AIaGK  179448
#> 11         VPw5X 3984120
#> 12         UazRF 1648852
#> 13         UazRF 1771727
#> 14         UazRF 3420579
#> 15         UazRF  160125
#> 16         1Dui3  111941
#> 17         1Dui3   11724
#> 18         1Dui3    2144
#> 19         1Dui3     487
#> 20         1Dui3       0
#> 21         1Dui3   20595
#> 22         1Dui3   15195
#> 23         1Dui3     290
#> 24         1Dui3     565
#> 25         1Dui3    8733
#> 26         P9Aeb   17737
#> 27         P9Aeb       2
#> 28         P9Aeb    1571
#> 29         P9Aeb   11942
#> 30         P9Aeb     217
#> 31         P9Aeb    3094
#> 32         oRfmX   17816