Example raw data

Download some example raw data. Mainly intended for use in developing / testing the sqid parsing or as an example of getting raw data if any end users would prefer to do the sqid parsing themselves.

Usage

example_data_raw(group = "attendance", size = 32)

Arguments

group

Choose the publication group of examples to use. Options are:

"attendance": Large example data set, careful what you ask for

"absence": Smaller example data set

size

Number of rows to return (max = 1000)

Value

Nested list form of example data from the API

Examples

example_data_raw()
#>    timePeriod.code timePeriod.period geographicLevel locations.LA locations.NAT
#> 1              W30              2025              LA        Qiwb4         dP0Zw
#> 2              W30              2025              LA        Qiwb4         dP0Zw
#> 3              W30              2025              LA        Qiwb4         dP0Zw
#> 4              W30              2025              LA        Qiwb4         dP0Zw
#> 5              W30              2025              LA        Qiwb4         dP0Zw
#> 6              W30              2025              LA        Qiwb4         dP0Zw
#> 7              W30              2025              LA        Qiwb4         dP0Zw
#> 8              W30              2025              LA        Qiwb4         dP0Zw
#> 9              W30              2025              LA        Qiwb4         dP0Zw
#> 10             W30              2025              LA        Qiwb4         dP0Zw
#> 11             W30              2025              LA        Qiwb4         dP0Zw
#> 12             W30              2025              LA        Qiwb4         dP0Zw
#> 13             W30              2025              LA        Qiwb4         dP0Zw
#> 14             W30              2025              LA        Qiwb4         dP0Zw
#> 15             W30              2025              LA        Qiwb4         dP0Zw
#> 16             W30              2025              LA        Qiwb4         dP0Zw
#> 17             W30              2025              LA        Qiwb4         dP0Zw
#> 18             W30              2025              LA        Qiwb4         dP0Zw
#> 19             W30              2025              LA        Qiwb4         dP0Zw
#> 20             W30              2025              LA        Qiwb4         dP0Zw
#> 21             W30              2025              LA        Qiwb4         dP0Zw
#> 22             W30              2025              LA        Qiwb4         dP0Zw
#> 23             W30              2025              LA        Qiwb4         dP0Zw
#> 24             W30              2025              LA        Qiwb4         dP0Zw
#> 25             W30              2025              LA        Qiwb4         dP0Zw
#> 26             W30              2025              LA        Qiwb4         dP0Zw
#> 27             W30              2025              LA        Qiwb4         dP0Zw
#> 28             W30              2025              LA        Qiwb4         dP0Zw
#> 29             W30              2025              LA        Qiwb4         dP0Zw
#> 30             W30              2025              LA        Qiwb4         dP0Zw
#> 31             W30              2025              LA        Qiwb4         dP0Zw
#> 32             W30              2025              LA        Qiwb4         dP0Zw
#>    locations.REG filters.z4FQE filters.X9fKb filters.TsPJP filters.krhsL
#> 1          ACyGK         bjstT         GBMgr         4yTUZ         TmQPJ
#> 2          ACyGK         3xu8u         GBMgr         4yTUZ         TmQPJ
#> 3          ACyGK         HnuzL         GBMgr         4yTUZ         TmQPJ
#> 4          ACyGK         0oaT5         GBMgr         4yTUZ         e4wuS
#> 5          ACyGK         S0OVx         GBMgr         4yTUZ         e4wuS
#> 6          ACyGK         Dt8Qe         GBMgr         4yTUZ         e4wuS
#> 7          ACyGK         y2daB         GBMgr         4yTUZ         h2IyW
#> 8          ACyGK         RaVka         GBMgr         4yTUZ         g7LO9
#> 9          ACyGK         qOnjG         GBMgr         4yTUZ         TmQPJ
#> 10         ACyGK         BHu7J         GBMgr         4yTUZ         TmQPJ
#> 11         ACyGK         YvoHK         GBMgr         4yTUZ         TmQPJ
#> 12         ACyGK         acQLP         GBMgr         4yTUZ         TmQPJ
#> 13         ACyGK         28Wlz         GBMgr         4yTUZ         e4wuS
#> 14         ACyGK         duo0Z         GBMgr         4yTUZ         e4wuS
#> 15         ACyGK         uU1o4         GBMgr         4yTUZ         e4wuS
#> 16         ACyGK         XFRfK         GBMgr         4yTUZ         e4wuS
#> 17         ACyGK         5PHdi         GBMgr         4yTUZ         e4wuS
#> 18         ACyGK         9ko4v         GBMgr         4yTUZ         e4wuS
#> 19         ACyGK         L31Wj         GBMgr         4yTUZ         e4wuS
#> 20         ACyGK         kWQhs         GBMgr         4yTUZ         e4wuS
#> 21         ACyGK         NtQDC         GBMgr         4yTUZ         e4wuS
#> 22         ACyGK         TmKPJ         GBMgr         4yTUZ         e4wuS
#> 23         ACyGK         QErwb         GBMgr         4yTUZ         TmQPJ
#> 24         ACyGK         PmPeb         GBMgr         4yTUZ         TmQPJ
#> 25         ACyGK         e9vuS         GBMgr         4yTUZ         TmQPJ
#> 26         ACyGK         W3k2b         GBMgr         4yTUZ         TmQPJ
#> 27         ACyGK         jC9AM         GBMgr         4yTUZ         TmQPJ
#> 28         ACyGK         EGQqF         GBMgr         4yTUZ         TmQPJ
#> 29         ACyGK         tGPEm         GBMgr         4yTUZ         e4wuS
#> 30         ACyGK         oRUmX         GBMgr         4yTUZ         e4wuS
#> 31         ACyGK         6wgrf         GBMgr         4yTUZ         e4wuS
#> 32         ACyGK         7zvXo         GBMgr         4yTUZ         e4wuS
#>    filters.uxo41   X9fKb
#> 1          UazRF 5639966
#> 2          P9Aeb   58092
#> 3          ls5cB 5698058
#> 4          uUIo4  295703
#> 5          1Dui3  186682
#> 6          oRfmX  109021
#> 7          AIaGK   58805
#> 8          VPw5X 5993761
#> 9          UazRF 2777002
#> 10         UazRF 2811203
#> 11         UazRF 5588205
#> 12         UazRF   51761
#> 13         1Dui3  147978
#> 14         1Dui3   15484
#> 15         1Dui3     771
#> 16         1Dui3       2
#> 17         1Dui3       0
#> 18         1Dui3    1972
#> 19         1Dui3   15195
#> 20         1Dui3     189
#> 21         1Dui3     226
#> 22         1Dui3    4378
#> 23         P9Aeb    7600
#> 24         P9Aeb       0
#> 25         P9Aeb    1932
#> 26         P9Aeb   48095
#> 27         P9Aeb       1
#> 28         P9Aeb     464
#> 29         oRfmX   53217
#> 30         oRfmX    7044
#> 31         oRfmX   46966
#> 32         oRfmX    1794