Example raw data

Download some example raw data. Mainly intended for use in developing / testing the sqid parsing or as an example of getting raw data if any end users would prefer to do the sqid parsing themselves.

Usage

example_data_raw(group = "attendance", size = 32)

Arguments

group

Choose the publication group of examples to use. Options are:

"attendance": Large example data set, careful what you ask for

"public-api-testing": Smaller example data set

size

Number of rows to return (max = 1000)

Value

Nested list form of example data from the API

Examples

example_data_raw()
#>    timePeriod.code timePeriod.period geographicLevel locations.LA locations.NAT
#> 1               W8              2025              LA        O7CLF         dP0Zw
#> 2               W8              2025              LA        O7CLF         dP0Zw
#> 3               W8              2025              LA        O7CLF         dP0Zw
#> 4               W8              2025              LA        O7CLF         dP0Zw
#> 5               W8              2025              LA        O7CLF         dP0Zw
#> 6               W8              2025              LA        O7CLF         dP0Zw
#> 7               W8              2025              LA        O7CLF         dP0Zw
#> 8               W8              2025              LA        O7CLF         dP0Zw
#> 9               W8              2025              LA        O7CLF         dP0Zw
#> 10              W8              2025              LA        O7CLF         dP0Zw
#> 11              W8              2025              LA        O7CLF         dP0Zw
#> 12              W8              2025              LA        O7CLF         dP0Zw
#> 13              W8              2025              LA        O7CLF         dP0Zw
#> 14              W8              2025              LA        O7CLF         dP0Zw
#> 15              W8              2025              LA        O7CLF         dP0Zw
#> 16              W8              2025              LA        O7CLF         dP0Zw
#> 17              W8              2025              LA        O7CLF         dP0Zw
#> 18              W8              2025              LA        O7CLF         dP0Zw
#> 19              W8              2025              LA        O7CLF         dP0Zw
#> 20              W8              2025              LA        O7CLF         dP0Zw
#> 21              W8              2025              LA        O7CLF         dP0Zw
#> 22              W8              2025              LA        O7CLF         dP0Zw
#> 23              W8              2025              LA        O7CLF         dP0Zw
#> 24              W8              2025              LA        O7CLF         dP0Zw
#> 25              W8              2025              LA        O7CLF         dP0Zw
#> 26              W8              2025              LA        O7CLF         dP0Zw
#> 27              W8              2025              LA        O7CLF         dP0Zw
#> 28              W8              2025              LA        O7CLF         dP0Zw
#> 29              W8              2025              LA        O7CLF         dP0Zw
#> 30              W8              2025              LA        O7CLF         dP0Zw
#> 31              W8              2025              LA        O7CLF         dP0Zw
#> 32              W8              2025              LA        O7CLF         dP0Zw
#>    locations.REG filters.TsPJP filters.X9fKb filters.z4FQE filters.krhsL
#> 1          ACyGK         4yTUZ         GBMgr         bjstT         TmQPJ
#> 2          ACyGK         4yTUZ         GBMgr         3xu8u         TmQPJ
#> 3          ACyGK         4yTUZ         GBMgr         HnuzL         TmQPJ
#> 4          ACyGK         4yTUZ         GBMgr         0oaT5         e4wuS
#> 5          ACyGK         4yTUZ         GBMgr         S0OVx         e4wuS
#> 6          ACyGK         4yTUZ         GBMgr         Dt8Qe         e4wuS
#> 7          ACyGK         4yTUZ         GBMgr         y2daB         h2IyW
#> 8          ACyGK         4yTUZ         GBMgr         RaVka         g7LO9
#> 9          ACyGK         4yTUZ         GBMgr         qOnjG         TmQPJ
#> 10         ACyGK         4yTUZ         GBMgr         BHu7J         TmQPJ
#> 11         ACyGK         4yTUZ         GBMgr         YvoHK         TmQPJ
#> 12         ACyGK         4yTUZ         GBMgr         acQLP         TmQPJ
#> 13         ACyGK         4yTUZ         GBMgr         28Wlz         e4wuS
#> 14         ACyGK         4yTUZ         GBMgr         duo0Z         e4wuS
#> 15         ACyGK         4yTUZ         GBMgr         uU1o4         e4wuS
#> 16         ACyGK         4yTUZ         GBMgr         XFRfK         e4wuS
#> 17         ACyGK         4yTUZ         GBMgr         5PHdi         e4wuS
#> 18         ACyGK         4yTUZ         GBMgr         9ko4v         e4wuS
#> 19         ACyGK         4yTUZ         GBMgr         L31Wj         e4wuS
#> 20         ACyGK         4yTUZ         GBMgr         kWQhs         e4wuS
#> 21         ACyGK         4yTUZ         GBMgr         NtQDC         e4wuS
#> 22         ACyGK         4yTUZ         GBMgr         TmKPJ         e4wuS
#> 23         ACyGK         4yTUZ         GBMgr         QErwb         TmQPJ
#> 24         ACyGK         4yTUZ         GBMgr         PmPeb         TmQPJ
#> 25         ACyGK         4yTUZ         GBMgr         e9vuS         TmQPJ
#> 26         ACyGK         4yTUZ         GBMgr         W3k2b         TmQPJ
#> 27         ACyGK         4yTUZ         GBMgr         jC9AM         TmQPJ
#> 28         ACyGK         4yTUZ         GBMgr         EGQqF         TmQPJ
#> 29         ACyGK         4yTUZ         GBMgr         tGPEm         e4wuS
#> 30         ACyGK         4yTUZ         GBMgr         oRUmX         e4wuS
#> 31         ACyGK         4yTUZ         GBMgr         6wgrf         e4wuS
#> 32         ACyGK         4yTUZ         GBMgr         7zvXo         e4wuS
#>    filters.uxo41   X9fKb
#> 1          UazRF 1273300
#> 2          P9Aeb    6642
#> 3          ls5cB 1279942
#> 4          uUIo4   70857
#> 5          1Dui3   49494
#> 6          oRfmX   21363
#> 7          AIaGK   17533
#> 8          VPw5X 1350799
#> 9          UazRF  621998
#> 10         UazRF  635842
#> 11         UazRF 1257840
#> 12         UazRF   15460
#> 13         1Dui3   42823
#> 14         1Dui3    2884
#> 15         1Dui3       7
#> 16         1Dui3       0
#> 17         1Dui3       0
#> 18         1Dui3      84
#> 19         1Dui3    3302
#> 20         1Dui3     115
#> 21         1Dui3      33
#> 22         1Dui3     245
#> 23         P9Aeb     843
#> 24         P9Aeb       0
#> 25         P9Aeb     734
#> 26         P9Aeb    5006
#> 27         P9Aeb       0
#> 28         P9Aeb      59
#> 29         oRfmX    8110
#> 30         oRfmX    2073
#> 31         oRfmX   10283
#> 32         oRfmX     897