Example raw data

Download some example raw data. Mainly intended for use in developing / testing the sqid parsing or as an example of getting raw data if any end users would prefer to do the sqid parsing themselves.

Usage

example_data_raw(group = "attendance", size = 32)

Arguments

group

Choose the publication group of examples to use. Options are:

"attendance": Large example data set, careful what you ask for

"public-api-testing": Smaller example data set

size

Number of rows to return (max = 1000)

Value

Nested list form of example data from the API

Examples

example_data_raw()
#>    timePeriod.code timePeriod.period geographicLevel locations.LA locations.NAT
#> 1              W26              2024              LA        bqZtT         mRj9K
#> 2              W26              2024              LA        bqZtT         mRj9K
#> 3              W26              2024              LA        bqZtT         mRj9K
#> 4              W26              2024              LA        bqZtT         mRj9K
#> 5              W26              2024              LA        bqZtT         mRj9K
#> 6              W26              2024              LA        bqZtT         mRj9K
#> 7              W26              2024              LA        bqZtT         mRj9K
#> 8              W26              2024              LA        bqZtT         mRj9K
#> 9              W26              2024              LA        bqZtT         mRj9K
#> 10             W26              2024              LA        bqZtT         mRj9K
#> 11             W26              2024              LA        bqZtT         mRj9K
#> 12             W26              2024              LA        bqZtT         mRj9K
#> 13             W26              2024              LA        bqZtT         mRj9K
#> 14             W26              2024              LA        bqZtT         mRj9K
#> 15             W26              2024              LA        bqZtT         mRj9K
#> 16             W26              2024              LA        bqZtT         mRj9K
#> 17             W26              2024              LA        bqZtT         mRj9K
#> 18             W26              2024              LA        bqZtT         mRj9K
#> 19             W26              2024              LA        bqZtT         mRj9K
#> 20             W26              2024              LA        bqZtT         mRj9K
#> 21             W26              2024              LA        bqZtT         mRj9K
#> 22             W26              2024              LA        bqZtT         mRj9K
#> 23             W26              2024              LA        bqZtT         mRj9K
#> 24             W26              2024              LA        bqZtT         mRj9K
#> 25             W26              2024              LA        bqZtT         mRj9K
#> 26             W26              2024              LA        bqZtT         mRj9K
#> 27             W26              2024              LA        bqZtT         mRj9K
#> 28             W26              2024              LA        bqZtT         mRj9K
#> 29             W26              2024              LA        bqZtT         mRj9K
#> 30             W26              2024              LA        bqZtT         mRj9K
#> 31             W26              2024              LA        bqZtT         mRj9K
#> 32             W26              2024              LA        bqZtT         mRj9K
#>    locations.REG filters.juAMt filters.eNuSW filters.5Zdi9 filters.fyYFZ
#> 1          kxGhs         V0D5X         9Ru4v         BfP7J         emDuS
#> 2          kxGhs         V0D5X         9Ru4v         BfP7J         emDuS
#> 3          kxGhs         V0D5X         9Ru4v         BfP7J         emDuS
#> 4          kxGhs         V0D5X         9Ru4v         qGJjG         emDuS
#> 5          kxGhs         V0D5X         9Ru4v         qGJjG         emDuS
#> 6          kxGhs         V0D5X         9Ru4v         qGJjG         emDuS
#> 7          kxGhs         V0D5X         9Ru4v         zvUFQ         emDuS
#> 8          kxGhs         V0D5X         9Ru4v         u9Mo4         emDuS
#> 9          kxGhs         V0D5X         mU59K         BfP7J         emDuS
#> 10         kxGhs         V0D5X         0evT5         BfP7J         emDuS
#> 11         kxGhs         V0D5X         4kdUZ         BfP7J         emDuS
#> 12         kxGhs         V0D5X         yW5aB         BfP7J         emDuS
#> 13         kxGhs         V0D5X         3se8u         qGJjG         emDuS
#> 14         kxGhs         V0D5X         RySka         qGJjG         emDuS
#> 15         kxGhs         V0D5X         WcR2b         qGJjG         emDuS
#> 16         kxGhs         V0D5X         EADqF         qGJjG         emDuS
#> 17         kxGhs         V0D5X         Qt6wb         qGJjG         emDuS
#> 18         kxGhs         V0D5X         GCsgr         qGJjG         emDuS
#> 19         kxGhs         V0D5X         o47mX         qGJjG         emDuS
#> 20         kxGhs         V0D5X         l8ScB         qGJjG         emDuS
#> 21         kxGhs         V0D5X         u9Oo4         BfP7J         emDuS
#> 22         kxGhs         V0D5X         UhYRF         BfP7J         emDuS
#> 23         kxGhs         V0D5X         5TYdi         BfP7J         emDuS
#> 24         kxGhs         V0D5X         PoTeb         BfP7J         emDuS
#> 25         kxGhs         V0D5X         NwNDC         BfP7J         emDuS
#> 26         kxGhs         V0D5X         rQkNj         qGJjG         emDuS
#> 27         kxGhs         V0D5X         23Llz         qGJjG         emDuS
#> 28         kxGhs         V0D5X         emJuS         qGJjG         emDuS
#> 29         kxGhs         V0D5X         bqZtT         mk09K         emDuS
#> 30         kxGhs         V0D5X         Db3Qe         mk09K         emDuS
#> 31         kxGhs         V0D5X         wqtbx         mk09K         emDuS
#> 32         kxGhs         V0D5X         oU4mX         mk09K         emDuS
#>    filters.tdEm5      tj0Em
#> 1          tj0Em      32342
#> 2          TuxPJ        605
#> 3          0opT5      32947
#> 4          0opT5       1955
#> 5          cZO31       1455
#> 6          jgoAM        500
#> 7          0opT5        404
#> 8          0opT5      34902
#> 9          tj0Em      15783
#> 10         tj0Em      16180
#> 11         tj0Em      31963
#> 12         tj0Em        379
#> 13         cZO31       1153
#> 14         cZO31         84
#> 15         cZO31          4
#> 16         cZO31          0
#> 17         cZO31         14
#> 18         cZO31         30
#> 19         cZO31          3
#> 20         cZO31        167
#> 21         TuxPJ        120
#> 22         TuxPJ          0
#> 23         TuxPJ         16
#> 24         TuxPJ        383
#> 25         TuxPJ          0
#> 26         jgoAM        241
#> 27         jgoAM         25
#> 28         jgoAM        192
#> 29         0opT5         42
#> 30         0opT5          0
#> 31         0opT5          2
#> 32         0opT5        124