Example raw data

Download some example raw data. Mainly intended for use in developing / testing the sqid parsing or as an example of getting raw data if any end users would prefer to do the sqid parsing themselves.

Usage

example_data_raw(group = "attendance", size = 32)

Arguments

group

Choose the publication group of examples to use. Options are:

"attendance": Large example data set, careful what you ask for

"public-api-testing": Smaller example data set

size

Number of rows to return (max = 1000)

Value

Nested list form of example data from the API

Examples

example_data_raw()
#>    timePeriod.code timePeriod.period geographicLevel locations.LA locations.NAT
#> 1              W18              2025              LA        kaNhs         dP0Zw
#> 2              W18              2025              LA        kaNhs         dP0Zw
#> 3              W18              2025              LA        kaNhs         dP0Zw
#> 4              W18              2025              LA        kaNhs         dP0Zw
#> 5              W18              2025              LA        kaNhs         dP0Zw
#> 6              W18              2025              LA        kaNhs         dP0Zw
#> 7              W18              2025              LA        kaNhs         dP0Zw
#> 8              W18              2025              LA        kaNhs         dP0Zw
#> 9              W18              2025              LA        kaNhs         dP0Zw
#> 10             W18              2025              LA        kaNhs         dP0Zw
#> 11             W18              2025              LA        kaNhs         dP0Zw
#> 12             W18              2025              LA        kaNhs         dP0Zw
#> 13             W18              2025              LA        kaNhs         dP0Zw
#> 14             W18              2025              LA        kaNhs         dP0Zw
#> 15             W18              2025              LA        kaNhs         dP0Zw
#> 16             W18              2025              LA        kaNhs         dP0Zw
#> 17             W18              2025              LA        kaNhs         dP0Zw
#> 18             W18              2025              LA        kaNhs         dP0Zw
#> 19             W18              2025              LA        kaNhs         dP0Zw
#> 20             W18              2025              LA        kaNhs         dP0Zw
#> 21             W18              2025              LA        kaNhs         dP0Zw
#> 22             W18              2025              LA        kaNhs         dP0Zw
#> 23             W18              2025              LA        kaNhs         dP0Zw
#> 24             W18              2025              LA        kaNhs         dP0Zw
#> 25             W18              2025              LA        kaNhs         dP0Zw
#> 26             W18              2025              LA        kaNhs         dP0Zw
#> 27             W18              2025              LA        kaNhs         dP0Zw
#> 28             W18              2025              LA        kaNhs         dP0Zw
#> 29             W18              2025              LA        kaNhs         dP0Zw
#> 30             W18              2025              LA        kaNhs         dP0Zw
#> 31             W18              2025              LA        kaNhs         dP0Zw
#> 32             W18              2025              LA        kaNhs         dP0Zw
#>    locations.REG filters.X9fKb filters.z4FQE filters.TsPJP filters.krhsL
#> 1          VTQ5X         3Mo8u         3xu8u         bAxtT         TmQPJ
#> 2          VTQ5X         3Mo8u         HnuzL         bAxtT         TmQPJ
#> 3          VTQ5X         3Mo8u         0oaT5         bAxtT         e4wuS
#> 4          VTQ5X         3Mo8u         S0OVx         bAxtT         e4wuS
#> 5          VTQ5X         3Mo8u         Dt8Qe         bAxtT         e4wuS
#> 6          VTQ5X         3Mo8u         y2daB         bAxtT         h2IyW
#> 7          VTQ5X         3Mo8u         RaVka         bAxtT         g7LO9
#> 8          VTQ5X         3Mo8u         qOnjG         bAxtT         TmQPJ
#> 9          VTQ5X         3Mo8u         BHu7J         bAxtT         TmQPJ
#> 10         VTQ5X         3Mo8u         YvoHK         bAxtT         TmQPJ
#> 11         VTQ5X         3Mo8u         acQLP         bAxtT         TmQPJ
#> 12         VTQ5X         3Mo8u         28Wlz         bAxtT         e4wuS
#> 13         VTQ5X         3Mo8u         duo0Z         bAxtT         e4wuS
#> 14         VTQ5X         3Mo8u         uU1o4         bAxtT         e4wuS
#> 15         VTQ5X         3Mo8u         XFRfK         bAxtT         e4wuS
#> 16         VTQ5X         3Mo8u         5PHdi         bAxtT         e4wuS
#> 17         VTQ5X         3Mo8u         9ko4v         bAxtT         e4wuS
#> 18         VTQ5X         3Mo8u         L31Wj         bAxtT         e4wuS
#> 19         VTQ5X         3Mo8u         kWQhs         bAxtT         e4wuS
#> 20         VTQ5X         3Mo8u         NtQDC         bAxtT         e4wuS
#> 21         VTQ5X         3Mo8u         TmKPJ         bAxtT         e4wuS
#> 22         VTQ5X         3Mo8u         QErwb         bAxtT         TmQPJ
#> 23         VTQ5X         3Mo8u         PmPeb         bAxtT         TmQPJ
#> 24         VTQ5X         3Mo8u         e9vuS         bAxtT         TmQPJ
#> 25         VTQ5X         3Mo8u         W3k2b         bAxtT         TmQPJ
#> 26         VTQ5X         3Mo8u         jC9AM         bAxtT         TmQPJ
#> 27         VTQ5X         3Mo8u         EGQqF         bAxtT         TmQPJ
#> 28         VTQ5X         3Mo8u         tGPEm         bAxtT         e4wuS
#> 29         VTQ5X         3Mo8u         oRUmX         bAxtT         e4wuS
#> 30         VTQ5X         3Mo8u         6wgrf         bAxtT         e4wuS
#> 31         VTQ5X         3Mo8u         7zvXo         bAxtT         e4wuS
#> 32         VTQ5X         3Mo8u         0unT5         bAxtT         mkA9K
#>    filters.uxo41 X9fKb
#> 1          P9Aeb   529
#> 2          ls5cB 34756
#> 3          uUIo4  3419
#> 4          1Dui3  1836
#> 5          oRfmX  1583
#> 6          AIaGK  1505
#> 7          VPw5X 38175
#> 8          UazRF 16057
#> 9          UazRF 16814
#> 10         UazRF 32871
#> 11         UazRF  1356
#> 12         1Dui3  1032
#> 13         1Dui3   184
#> 14         1Dui3     0
#> 15         1Dui3    14
#> 16         1Dui3     0
#> 17         1Dui3    79
#> 18         1Dui3   313
#> 19         1Dui3     0
#> 20         1Dui3     2
#> 21         1Dui3   212
#> 22         P9Aeb   324
#> 23         P9Aeb     0
#> 24         P9Aeb   108
#> 25         P9Aeb    75
#> 26         P9Aeb     8
#> 27         P9Aeb    14
#> 28         oRfmX   120
#> 29         oRfmX   149
#> 30         oRfmX  1244
#> 31         oRfmX    70
#> 32         duj0Z   253