Example raw data

Download some example raw data. Mainly intended for use in developing / testing the sqid parsing or as an example of getting raw data if any end users would prefer to do the sqid parsing themselves.

Usage

example_data_raw(group = "attendance", size = 32)

Arguments

group

Choose the publication group of examples to use. Options are:

"attendance": Large example data set, careful what you ask for

"absence": Smaller example data set

size

Number of rows to return (max = 1000)

Value

Nested list form of example data from the API

Examples

example_data_raw()
#>    timePeriod.code timePeriod.period geographicLevel locations.LA locations.NAT
#> 1              W30              2025              LA        Qiwb4         dP0Zw
#> 2              W30              2025              LA        Qiwb4         dP0Zw
#> 3              W30              2025              LA        Qiwb4         dP0Zw
#> 4              W30              2025              LA        Qiwb4         dP0Zw
#> 5              W30              2025              LA        Qiwb4         dP0Zw
#> 6              W30              2025              LA        Qiwb4         dP0Zw
#> 7              W30              2025              LA        Qiwb4         dP0Zw
#> 8              W30              2025              LA        Qiwb4         dP0Zw
#> 9              W30              2025              LA        Qiwb4         dP0Zw
#> 10             W30              2025              LA        Qiwb4         dP0Zw
#> 11             W30              2025              LA        Qiwb4         dP0Zw
#> 12             W30              2025              LA        Qiwb4         dP0Zw
#> 13             W30              2025              LA        Qiwb4         dP0Zw
#> 14             W30              2025              LA        Qiwb4         dP0Zw
#> 15             W30              2025              LA        Qiwb4         dP0Zw
#> 16             W30              2025              LA        Qiwb4         dP0Zw
#> 17             W30              2025              LA        Qiwb4         dP0Zw
#> 18             W30              2025              LA        Qiwb4         dP0Zw
#> 19             W30              2025              LA        Qiwb4         dP0Zw
#> 20             W30              2025              LA        Qiwb4         dP0Zw
#> 21             W30              2025              LA        Qiwb4         dP0Zw
#> 22             W30              2025              LA        Qiwb4         dP0Zw
#> 23             W30              2025              LA        Qiwb4         dP0Zw
#> 24             W30              2025              LA        Qiwb4         dP0Zw
#> 25             W30              2025              LA        Qiwb4         dP0Zw
#> 26             W30              2025              LA        Qiwb4         dP0Zw
#> 27             W30              2025              LA        Qiwb4         dP0Zw
#> 28             W30              2025              LA        Qiwb4         dP0Zw
#> 29             W30              2025              LA        Qiwb4         dP0Zw
#> 30             W30              2025              LA        Qiwb4         dP0Zw
#> 31             W30              2025              LA        Qiwb4         dP0Zw
#> 32             W30              2025              LA        Qiwb4         dP0Zw
#>    locations.REG filters.X9fKb filters.TsPJP filters.krhsL filters.uxo41
#> 1          ACyGK         GBMgr         4yTUZ         TmQPJ         UazRF
#> 2          ACyGK         GBMgr         4yTUZ         TmQPJ         P9Aeb
#> 3          ACyGK         GBMgr         4yTUZ         TmQPJ         ls5cB
#> 4          ACyGK         GBMgr         4yTUZ         e4wuS         uUIo4
#> 5          ACyGK         GBMgr         4yTUZ         e4wuS         1Dui3
#> 6          ACyGK         GBMgr         4yTUZ         e4wuS         oRfmX
#> 7          ACyGK         GBMgr         4yTUZ         h2IyW         AIaGK
#> 8          ACyGK         GBMgr         4yTUZ         g7LO9         VPw5X
#> 9          ACyGK         GBMgr         4yTUZ         TmQPJ         UazRF
#> 10         ACyGK         GBMgr         4yTUZ         TmQPJ         UazRF
#> 11         ACyGK         GBMgr         4yTUZ         TmQPJ         UazRF
#> 12         ACyGK         GBMgr         4yTUZ         TmQPJ         UazRF
#> 13         ACyGK         GBMgr         4yTUZ         e4wuS         1Dui3
#> 14         ACyGK         GBMgr         4yTUZ         e4wuS         1Dui3
#> 15         ACyGK         GBMgr         4yTUZ         e4wuS         1Dui3
#> 16         ACyGK         GBMgr         4yTUZ         e4wuS         1Dui3
#> 17         ACyGK         GBMgr         4yTUZ         e4wuS         1Dui3
#> 18         ACyGK         GBMgr         4yTUZ         e4wuS         1Dui3
#> 19         ACyGK         GBMgr         4yTUZ         e4wuS         1Dui3
#> 20         ACyGK         GBMgr         4yTUZ         e4wuS         1Dui3
#> 21         ACyGK         GBMgr         4yTUZ         e4wuS         1Dui3
#> 22         ACyGK         GBMgr         4yTUZ         e4wuS         1Dui3
#> 23         ACyGK         GBMgr         4yTUZ         TmQPJ         P9Aeb
#> 24         ACyGK         GBMgr         4yTUZ         TmQPJ         P9Aeb
#> 25         ACyGK         GBMgr         4yTUZ         TmQPJ         P9Aeb
#> 26         ACyGK         GBMgr         4yTUZ         TmQPJ         P9Aeb
#> 27         ACyGK         GBMgr         4yTUZ         TmQPJ         P9Aeb
#> 28         ACyGK         GBMgr         4yTUZ         TmQPJ         P9Aeb
#> 29         ACyGK         GBMgr         4yTUZ         e4wuS         oRfmX
#> 30         ACyGK         GBMgr         4yTUZ         e4wuS         oRfmX
#> 31         ACyGK         GBMgr         4yTUZ         e4wuS         oRfmX
#> 32         ACyGK         GBMgr         4yTUZ         e4wuS         oRfmX
#>    filters.z4FQE   X9fKb
#> 1          bjstT 5639966
#> 2          3xu8u   58092
#> 3          HnuzL 5698058
#> 4          0oaT5  295703
#> 5          S0OVx  186682
#> 6          Dt8Qe  109021
#> 7          y2daB   58805
#> 8          RaVka 5993761
#> 9          qOnjG 2777002
#> 10         BHu7J 2811203
#> 11         YvoHK 5588205
#> 12         acQLP   51761
#> 13         28Wlz  147978
#> 14         duo0Z   15484
#> 15         uU1o4     771
#> 16         XFRfK       2
#> 17         5PHdi       0
#> 18         9ko4v    1972
#> 19         L31Wj   15195
#> 20         kWQhs     189
#> 21         NtQDC     226
#> 22         TmKPJ    4378
#> 23         QErwb    7600
#> 24         PmPeb       0
#> 25         e9vuS    1932
#> 26         W3k2b   48095
#> 27         jC9AM       1
#> 28         EGQqF     464
#> 29         tGPEm   53217
#> 30         oRUmX    7044
#> 31         6wgrf   46966
#> 32         7zvXo    1794