Example raw data

Download some example raw data. Mainly intended for use in developing / testing the sqid parsing or as an example of getting raw data if any end users would prefer to do the sqid parsing themselves.

Usage

example_data_raw(group = "attendance", size = 32)

Arguments

group

Choose the publication group of examples to use. Options are:

"attendance": Large example data set, careful what you ask for

"public-api-testing": Smaller example data set

size

Number of rows to return (max = 1000)

Value

Nested list form of example data from the API

Examples

example_data_raw()
#>    timePeriod.code timePeriod.period geographicLevel locations.NAT
#> 1              W10              2025             REG         dP0Zw
#> 2              W10              2025             REG         dP0Zw
#> 3              W10              2025             REG         dP0Zw
#> 4              W10              2025             REG         dP0Zw
#> 5              W10              2025             REG         dP0Zw
#> 6              W10              2025             REG         dP0Zw
#> 7              W10              2025             REG         dP0Zw
#> 8              W10              2025             REG         dP0Zw
#> 9              W10              2025             REG         dP0Zw
#> 10             W10              2025             REG         dP0Zw
#> 11             W10              2025             REG         dP0Zw
#> 12             W10              2025             REG         dP0Zw
#> 13             W10              2025             REG         dP0Zw
#> 14             W10              2025             REG         dP0Zw
#> 15             W10              2025             REG         dP0Zw
#> 16             W10              2025             REG         dP0Zw
#> 17             W10              2025             REG         dP0Zw
#> 18             W10              2025             REG         dP0Zw
#> 19             W10              2025             REG         dP0Zw
#> 20             W10              2025             REG         dP0Zw
#> 21             W10              2025             REG         dP0Zw
#> 22             W10              2025             REG         dP0Zw
#> 23             W10              2025             REG         dP0Zw
#> 24             W10              2025             REG         dP0Zw
#> 25             W10              2025             REG         dP0Zw
#> 26             W10              2025             REG         dP0Zw
#> 27             W10              2025             REG         dP0Zw
#> 28             W10              2025             REG         dP0Zw
#> 29             W10              2025             REG         dP0Zw
#> 30             W10              2025             REG         dP0Zw
#> 31             W10              2025             REG         dP0Zw
#> 32             W10              2025             REG         dP0Zw
#>    locations.REG filters.X9fKb filters.TsPJP filters.krhsL filters.uxo41
#> 1          ACyGK         rbyNj         4yTUZ         TmQPJ         UazRF
#> 2          ACyGK         rbyNj         4yTUZ         TmQPJ         P9Aeb
#> 3          ACyGK         rbyNj         4yTUZ         TmQPJ         ls5cB
#> 4          ACyGK         rbyNj         4yTUZ         e4wuS         uUIo4
#> 5          ACyGK         rbyNj         4yTUZ         e4wuS         1Dui3
#> 6          ACyGK         rbyNj         4yTUZ         e4wuS         oRfmX
#> 7          ACyGK         rbyNj         4yTUZ         h2IyW         AIaGK
#> 8          ACyGK         rbyNj         4yTUZ         g7LO9         VPw5X
#> 9          ACyGK         rbyNj         4yTUZ         TmQPJ         UazRF
#> 10         ACyGK         rbyNj         4yTUZ         TmQPJ         UazRF
#> 11         ACyGK         rbyNj         4yTUZ         TmQPJ         UazRF
#> 12         ACyGK         rbyNj         4yTUZ         TmQPJ         UazRF
#> 13         ACyGK         rbyNj         4yTUZ         e4wuS         1Dui3
#> 14         ACyGK         rbyNj         4yTUZ         e4wuS         1Dui3
#> 15         ACyGK         rbyNj         4yTUZ         e4wuS         1Dui3
#> 16         ACyGK         rbyNj         4yTUZ         e4wuS         1Dui3
#> 17         ACyGK         rbyNj         4yTUZ         e4wuS         1Dui3
#> 18         ACyGK         rbyNj         4yTUZ         e4wuS         1Dui3
#> 19         ACyGK         rbyNj         4yTUZ         e4wuS         1Dui3
#> 20         ACyGK         rbyNj         4yTUZ         e4wuS         1Dui3
#> 21         ACyGK         rbyNj         4yTUZ         e4wuS         1Dui3
#> 22         ACyGK         rbyNj         4yTUZ         e4wuS         1Dui3
#> 23         ACyGK         rbyNj         4yTUZ         TmQPJ         P9Aeb
#> 24         ACyGK         rbyNj         4yTUZ         TmQPJ         P9Aeb
#> 25         ACyGK         rbyNj         4yTUZ         TmQPJ         P9Aeb
#> 26         ACyGK         rbyNj         4yTUZ         TmQPJ         P9Aeb
#> 27         ACyGK         rbyNj         4yTUZ         TmQPJ         P9Aeb
#> 28         ACyGK         rbyNj         4yTUZ         TmQPJ         P9Aeb
#> 29         ACyGK         rbyNj         4yTUZ         e4wuS         oRfmX
#> 30         ACyGK         rbyNj         4yTUZ         e4wuS         oRfmX
#> 31         ACyGK         rbyNj         4yTUZ         e4wuS         oRfmX
#> 32         ACyGK         rbyNj         4yTUZ         e4wuS         oRfmX
#>    filters.z4FQE    X9fKb
#> 1          bjstT 61554120
#> 2          3xu8u   516772
#> 3          HnuzL 62070892
#> 4          0oaT5  4648820
#> 5          S0OVx  2886610
#> 6          Dt8Qe  1762210
#> 7          y2daB  1354113
#> 8          RaVka 66719712
#> 9          qOnjG 29757304
#> 10         BHu7J 30567654
#> 11         YvoHK 60324958
#> 12         acQLP  1229162
#> 13         28Wlz  2155394
#> 14         duo0Z   205372
#> 15         uU1o4     1596
#> 16         XFRfK       58
#> 17         5PHdi       63
#> 18         9ko4v   161417
#> 19         L31Wj   226475
#> 20         kWQhs     3023
#> 21         NtQDC     2312
#> 22         TmKPJ   123676
#> 23         QErwb   145513
#> 24         PmPeb        5
#> 25         e9vuS    57588
#> 26         W3k2b   268767
#> 27         jC9AM     8021
#> 28         EGQqF    36878
#> 29         tGPEm   326838
#> 30         oRUmX   124951
#> 31         6wgrf  1275659
#> 32         7zvXo    34762