Example raw data

Download some example raw data. Mainly intended for use in developing / testing the sqid parsing or as an example of getting raw data if any end users would prefer to do the sqid parsing themselves.

Usage

example_data_raw(group = "attendance", size = 32)

Arguments

group

Choose the publication group of examples to use. Options are:

"attendance": Large example data set, careful what you ask for

"public-api-testing": Smaller example data set

size

Number of rows to return (max = 1000)

Value

Nested list form of example data from the API

Examples

example_data_raw()
#>    timePeriod.code timePeriod.period geographicLevel locations.LA locations.NAT
#> 1              W26              2024              LA        bqZtT         mRj9K
#> 2              W26              2024              LA        bqZtT         mRj9K
#> 3              W26              2024              LA        bqZtT         mRj9K
#> 4              W26              2024              LA        bqZtT         mRj9K
#> 5              W26              2024              LA        bqZtT         mRj9K
#> 6              W26              2024              LA        bqZtT         mRj9K
#> 7              W26              2024              LA        bqZtT         mRj9K
#> 8              W26              2024              LA        bqZtT         mRj9K
#> 9              W26              2024              LA        bqZtT         mRj9K
#> 10             W26              2024              LA        bqZtT         mRj9K
#> 11             W26              2024              LA        bqZtT         mRj9K
#> 12             W26              2024              LA        bqZtT         mRj9K
#> 13             W26              2024              LA        bqZtT         mRj9K
#> 14             W26              2024              LA        bqZtT         mRj9K
#> 15             W26              2024              LA        bqZtT         mRj9K
#> 16             W26              2024              LA        bqZtT         mRj9K
#> 17             W26              2024              LA        bqZtT         mRj9K
#> 18             W26              2024              LA        bqZtT         mRj9K
#> 19             W26              2024              LA        bqZtT         mRj9K
#> 20             W26              2024              LA        bqZtT         mRj9K
#> 21             W26              2024              LA        bqZtT         mRj9K
#> 22             W26              2024              LA        bqZtT         mRj9K
#> 23             W26              2024              LA        bqZtT         mRj9K
#> 24             W26              2024              LA        bqZtT         mRj9K
#> 25             W26              2024              LA        bqZtT         mRj9K
#> 26             W26              2024              LA        bqZtT         mRj9K
#> 27             W26              2024              LA        bqZtT         mRj9K
#> 28             W26              2024              LA        bqZtT         mRj9K
#> 29             W26              2024              LA        bqZtT         mRj9K
#> 30             W26              2024              LA        bqZtT         mRj9K
#> 31             W26              2024              LA        bqZtT         mRj9K
#> 32             W26              2024              LA        bqZtT         mRj9K
#>    locations.REG filters.tdEm5 filters.5Zdi9 filters.fyYFZ filters.juAMt
#> 1          kxGhs         tj0Em         BfP7J         emDuS         V0D5X
#> 2          kxGhs         TuxPJ         BfP7J         emDuS         V0D5X
#> 3          kxGhs         0opT5         BfP7J         emDuS         V0D5X
#> 4          kxGhs         0opT5         qGJjG         emDuS         V0D5X
#> 5          kxGhs         cZO31         qGJjG         emDuS         V0D5X
#> 6          kxGhs         jgoAM         qGJjG         emDuS         V0D5X
#> 7          kxGhs         0opT5         zvUFQ         emDuS         V0D5X
#> 8          kxGhs         0opT5         u9Mo4         emDuS         V0D5X
#> 9          kxGhs         tj0Em         BfP7J         emDuS         V0D5X
#> 10         kxGhs         tj0Em         BfP7J         emDuS         V0D5X
#> 11         kxGhs         tj0Em         BfP7J         emDuS         V0D5X
#> 12         kxGhs         tj0Em         BfP7J         emDuS         V0D5X
#> 13         kxGhs         cZO31         qGJjG         emDuS         V0D5X
#> 14         kxGhs         cZO31         qGJjG         emDuS         V0D5X
#> 15         kxGhs         cZO31         qGJjG         emDuS         V0D5X
#> 16         kxGhs         cZO31         qGJjG         emDuS         V0D5X
#> 17         kxGhs         cZO31         qGJjG         emDuS         V0D5X
#> 18         kxGhs         cZO31         qGJjG         emDuS         V0D5X
#> 19         kxGhs         cZO31         qGJjG         emDuS         V0D5X
#> 20         kxGhs         cZO31         qGJjG         emDuS         V0D5X
#> 21         kxGhs         TuxPJ         BfP7J         emDuS         V0D5X
#> 22         kxGhs         TuxPJ         BfP7J         emDuS         V0D5X
#> 23         kxGhs         TuxPJ         BfP7J         emDuS         V0D5X
#> 24         kxGhs         TuxPJ         BfP7J         emDuS         V0D5X
#> 25         kxGhs         TuxPJ         BfP7J         emDuS         V0D5X
#> 26         kxGhs         jgoAM         qGJjG         emDuS         V0D5X
#> 27         kxGhs         jgoAM         qGJjG         emDuS         V0D5X
#> 28         kxGhs         jgoAM         qGJjG         emDuS         V0D5X
#> 29         kxGhs         0opT5         mk09K         emDuS         V0D5X
#> 30         kxGhs         0opT5         mk09K         emDuS         V0D5X
#> 31         kxGhs         0opT5         mk09K         emDuS         V0D5X
#> 32         kxGhs         0opT5         mk09K         emDuS         V0D5X
#>    filters.eNuSW      tj0Em
#> 1          9Ru4v      32342
#> 2          9Ru4v        605
#> 3          9Ru4v      32947
#> 4          9Ru4v       1955
#> 5          9Ru4v       1455
#> 6          9Ru4v        500
#> 7          9Ru4v        404
#> 8          9Ru4v      34902
#> 9          mU59K      15783
#> 10         0evT5      16180
#> 11         4kdUZ      31963
#> 12         yW5aB        379
#> 13         3se8u       1153
#> 14         RySka         84
#> 15         WcR2b          4
#> 16         EADqF          0
#> 17         Qt6wb         14
#> 18         GCsgr         30
#> 19         o47mX          3
#> 20         l8ScB        167
#> 21         u9Oo4        120
#> 22         UhYRF          0
#> 23         5TYdi         16
#> 24         PoTeb        383
#> 25         NwNDC          0
#> 26         rQkNj        241
#> 27         23Llz         25
#> 28         emJuS        192
#> 29         bqZtT         42
#> 30         Db3Qe          0
#> 31         wqtbx          2
#> 32         oU4mX        124