Example raw data

Download some example raw data. Mainly intended for use in developing / testing the sqid parsing or as an example of getting raw data if any end users would prefer to do the sqid parsing themselves.

Usage

example_data_raw(group = "attendance", size = 32)

Arguments

group

Choose the publication group of examples to use. Options are:

"attendance": Large example data set, careful what you ask for

"public-api-testing": Smaller example data set

size

Number of rows to return (max = 1000)

Value

Nested list form of example data from the API

Examples

example_data_raw()
#>    timePeriod.code timePeriod.period geographicLevel locations.NAT
#> 1              W10              2025             REG         dP0Zw
#> 2              W10              2025             REG         dP0Zw
#> 3              W10              2025             REG         dP0Zw
#> 4              W10              2025             REG         dP0Zw
#> 5              W10              2025             REG         dP0Zw
#> 6              W10              2025             REG         dP0Zw
#> 7              W10              2025             REG         dP0Zw
#> 8              W10              2025             REG         dP0Zw
#> 9              W10              2025             REG         dP0Zw
#> 10             W10              2025             REG         dP0Zw
#> 11             W10              2025             REG         dP0Zw
#> 12             W10              2025             REG         dP0Zw
#> 13             W10              2025             REG         dP0Zw
#> 14             W10              2025             REG         dP0Zw
#> 15             W10              2025             REG         dP0Zw
#> 16             W10              2025             REG         dP0Zw
#> 17             W10              2025             REG         dP0Zw
#> 18             W10              2025             REG         dP0Zw
#> 19             W10              2025             REG         dP0Zw
#> 20             W10              2025             REG         dP0Zw
#> 21             W10              2025             REG         dP0Zw
#> 22             W10              2025             REG         dP0Zw
#> 23             W10              2025             REG         dP0Zw
#> 24             W10              2025             REG         dP0Zw
#> 25             W10              2025             REG         dP0Zw
#> 26             W10              2025             REG         dP0Zw
#> 27             W10              2025             REG         dP0Zw
#> 28             W10              2025             REG         dP0Zw
#> 29             W10              2025             REG         dP0Zw
#> 30             W10              2025             REG         dP0Zw
#> 31             W10              2025             REG         dP0Zw
#> 32             W10              2025             REG         dP0Zw
#>    locations.REG filters.X9fKb filters.TsPJP filters.z4FQE filters.krhsL
#> 1          ACyGK         rbyNj         4yTUZ         bjstT         TmQPJ
#> 2          ACyGK         rbyNj         4yTUZ         3xu8u         TmQPJ
#> 3          ACyGK         rbyNj         4yTUZ         HnuzL         TmQPJ
#> 4          ACyGK         rbyNj         4yTUZ         0oaT5         e4wuS
#> 5          ACyGK         rbyNj         4yTUZ         S0OVx         e4wuS
#> 6          ACyGK         rbyNj         4yTUZ         Dt8Qe         e4wuS
#> 7          ACyGK         rbyNj         4yTUZ         y2daB         h2IyW
#> 8          ACyGK         rbyNj         4yTUZ         RaVka         g7LO9
#> 9          ACyGK         rbyNj         4yTUZ         qOnjG         TmQPJ
#> 10         ACyGK         rbyNj         4yTUZ         BHu7J         TmQPJ
#> 11         ACyGK         rbyNj         4yTUZ         YvoHK         TmQPJ
#> 12         ACyGK         rbyNj         4yTUZ         acQLP         TmQPJ
#> 13         ACyGK         rbyNj         4yTUZ         28Wlz         e4wuS
#> 14         ACyGK         rbyNj         4yTUZ         duo0Z         e4wuS
#> 15         ACyGK         rbyNj         4yTUZ         uU1o4         e4wuS
#> 16         ACyGK         rbyNj         4yTUZ         XFRfK         e4wuS
#> 17         ACyGK         rbyNj         4yTUZ         5PHdi         e4wuS
#> 18         ACyGK         rbyNj         4yTUZ         9ko4v         e4wuS
#> 19         ACyGK         rbyNj         4yTUZ         L31Wj         e4wuS
#> 20         ACyGK         rbyNj         4yTUZ         kWQhs         e4wuS
#> 21         ACyGK         rbyNj         4yTUZ         NtQDC         e4wuS
#> 22         ACyGK         rbyNj         4yTUZ         TmKPJ         e4wuS
#> 23         ACyGK         rbyNj         4yTUZ         QErwb         TmQPJ
#> 24         ACyGK         rbyNj         4yTUZ         PmPeb         TmQPJ
#> 25         ACyGK         rbyNj         4yTUZ         e9vuS         TmQPJ
#> 26         ACyGK         rbyNj         4yTUZ         W3k2b         TmQPJ
#> 27         ACyGK         rbyNj         4yTUZ         jC9AM         TmQPJ
#> 28         ACyGK         rbyNj         4yTUZ         EGQqF         TmQPJ
#> 29         ACyGK         rbyNj         4yTUZ         tGPEm         e4wuS
#> 30         ACyGK         rbyNj         4yTUZ         oRUmX         e4wuS
#> 31         ACyGK         rbyNj         4yTUZ         6wgrf         e4wuS
#> 32         ACyGK         rbyNj         4yTUZ         7zvXo         e4wuS
#>    filters.uxo41    X9fKb
#> 1          UazRF 61554120
#> 2          P9Aeb   516772
#> 3          ls5cB 62070892
#> 4          uUIo4  4648820
#> 5          1Dui3  2886610
#> 6          oRfmX  1762210
#> 7          AIaGK  1354113
#> 8          VPw5X 66719712
#> 9          UazRF 29757304
#> 10         UazRF 30567654
#> 11         UazRF 60324958
#> 12         UazRF  1229162
#> 13         1Dui3  2155394
#> 14         1Dui3   205372
#> 15         1Dui3     1596
#> 16         1Dui3       58
#> 17         1Dui3       63
#> 18         1Dui3   161417
#> 19         1Dui3   226475
#> 20         1Dui3     3023
#> 21         1Dui3     2312
#> 22         1Dui3   123676
#> 23         P9Aeb   145513
#> 24         P9Aeb        5
#> 25         P9Aeb    57588
#> 26         P9Aeb   268767
#> 27         P9Aeb     8021
#> 28         P9Aeb    36878
#> 29         oRfmX   326838
#> 30         oRfmX   124951
#> 31         oRfmX  1275659
#> 32         oRfmX    34762