Get publication specific data set catalogue

Get publication specific data set catalogue

Usage

get_data_catalogue(
  publication_id,
  ees_environment = NULL,
  api_version = NULL,
  page_size = NULL,
  page = NULL,
  verbose = FALSE
)

Arguments

publication_id

ID of the publication to be connected to. This is required if the endpoint is "get-data-catalogue"

ees_environment

EES ees_environment to connect to: "dev", "test", "preprod" or "prod"

api_version

EES API version

page_size

Number of results to return in a single query

page

Page number of query results to return

verbose

Run with additional contextual messaging. Logical, default = FALSE

Value

Data frame listing the data sets contained within a single publication

Examples

get_data_catalogue(example_id("publication"))
#>                                     id                           title
#> 1 3bb59201-0d1e-1e74-a745-8f65fe5895cf A level and equivalents results
#> 2 57b69201-033a-2c77-a19f-abcce2b11341       Pupil attendance sessions
#> 3 e1ae9201-2fff-d376-8fa3-bd3c3660d4c8                   Pupil absence
#>     summary    status latestVersion.version          latestVersion.published
#> 1           Published                   1.0 2024-10-22T22:34:44.356583+00:00
#> 2           Published                   1.0 2024-10-22T22:34:44.356583+00:00
#> 3 Test file Published                   1.0 2024-10-21T12:15:26.171566+00:00
#>   latestVersion.totalResults                     latestVersion.id
#> 1                      22685 49ac60e3-915f-4080-9b07-fb7625e02011
#> 2                    3315543 cc6745d0-d4bb-4871-bf26-a9ed653a66f1
#> 3                         12 5296244d-5487-4416-bec0-43ed4b50b72a
#>   latestVersion.timePeriods.start latestVersion.timePeriods.end
#> 1                         2021/22                       2021/22
#> 2                    2023 Week 37                  2024 Week 25
#> 3                         2021/22                       2023/24
#>        latestVersion.geographicLevels
#> 1                              School
#> 2 Local authority, National, Regional
#> 3                            National
#>                                                         latestVersion.filters
#> 1        Disadvantaged status, Establishment type group, Qualification cohort
#> 2 Attendance status, Attendance type, Day number, Establishment phase, Reason
#> 3                                                              Type of school
#>                                                                                                                                              latestVersion.indicators
#> 1 Number of students retained, Number of students retained (2nd year), Points per entry, Points per entry (grade), Retained, Total students, Total triggered students
#> 2                                                                                                                      Session count, Session percent, Session scaled
#> 3                                                                         Number of overall absence sessions, Number of pupil enrolments, Number of sessions possible