Get publication specific data set catalogue

Get publication specific data set catalogue

Usage

get_data_catalogue(
  publication_id,
  ees_environment = NULL,
  api_version = NULL,
  page_size = NULL,
  page = NULL,
  verbose = FALSE
)

Arguments

publication_id

ID of the publication to be connected to. This is required if the endpoint is "get-data-catalogue"

ees_environment

EES ees_environment to connect to: "dev", "test", "preprod" or "prod"

api_version

EES API version

page_size

Number of results to return in a single query

page

Page number of query results to return

verbose

Run with additional contextual messaging. Logical, default = FALSE

Value

Data frame listing the data sets contained within a single publication

Examples

get_data_catalogue(example_id("publication"))
#>                                     id                              title
#> 1 55629501-e98b-0c75-adba-f95a0cfbb5e9      Persistent absence in schools
#> 2 63629501-d3ca-c471-9780-ec4cb6fdf172 Reasons for absence and attendance
#> 3 55629501-1945-0174-956c-594f21c90404 Numbers of schools submitting data
#>                                                           summary    status
#> 1                                     Persistent absence measures Published
#> 2                        Reasons for pupil attendance and absence Published
#> 3 The number of schools returning data as used in the publication Published
#>   latestVersion.version          latestVersion.published
#> 1                   1.1  2025-03-20T09:30:16.09499+00:00
#> 2                   1.1 2025-03-20T09:30:16.029694+00:00
#> 3                   1.1 2025-03-20T09:30:15.467745+00:00
#>   latestVersion.totalResults                     latestVersion.id
#> 1                        988 6f69c2ae-e17a-4853-8440-7278306045f7
#> 2                    3013000 77e0afa7-7b92-47cb-988a-04d9694d9381
#> 3                      65500 ba8cbe51-ec5a-4684-8163-fe3642665061
#>   latestVersion.timePeriods.start latestVersion.timePeriods.end
#> 1                     2025 Week 8                  2025 Week 10
#> 2                    2024 Week 37                  2025 Week 10
#> 3                    2024 Week 37                  2025 Week 10
#>        latestVersion.geographicLevels
#> 1 Local authority, National, Regional
#> 2 Local authority, National, Regional
#> 3 Local authority, National, Regional
#>                                                                latestVersion.filters
#> 1                                                                    Education phase
#> 2 Attendance reason, Attendance status, Attendance type, Education phase, Time frame
#> 3                                                        Education phase, Time frame
#>                                                               latestVersion.indicators
#> 1                                                           Persistent absence percent
#> 2 Number of sessions, Number of sessions (scaled), Percent of sessions, Reference date
#> 3                Number of schools submitting, Reference date, Total number of schools