Query a data set using GET and a query URL

This function provides a method for generating and sending a URL based data query to the EES API. As a minimum, it requires the dataset_id and indicators flags to be provided.

Note that the GET command is very limited on the level of logic it can process. For example there is no way of using GET to make a query that combines different filters with AND logic. So if you give GET a set of filter items to search on, it can only return all rows containing any of those items.

Usage

get_dataset(
  dataset_id,
  indicators,
  time_periods = NULL,
  geographic_levels = NULL,
  locations = NULL,
  filter_items = NULL,
  dataset_version = NULL,
  api_version = NULL,
  page = NULL,
  page_size = 1000,
  parse = TRUE,
  verbose = FALSE
)

Arguments

dataset_id

ID of data set to be connected to. This is required if the endpoint is one of "get-summary", "get-meta", "get-data" or "post-data"

indicators

Indicators required as a string or vector of strings (required)

time_periods

Time periods required as a string ("period|code") or vector of strings

geographic_levels

Geographic levels required as a string or vector of strings

locations

Location code required as a string or vector of strings

filter_items

Filter items required as a string or vector of strings

dataset_version

Version of data set to be connected to

api_version

EES API version

page

Page number of query results to return

page_size

Number of results to return in a single query

parse

Logical flag to activate parsing of the results. Default: TRUE

verbose

Add extra contextual information whilst running

Value

Data frame containing query results of an API data set

Examples

get_dataset(
  example_id(),
  geographic_levels = "NAT",
  filter_items = example_id("filter_item"),
  indicators = example_id("indicator")
)
#>    code    period geographic_level   NAT 5mvdi h8fyW
#> 1    AY 2016/2017              NAT dP0Zw HsQzL 21319
#> 2    AY 2015/2016              NAT dP0Zw HsQzL 21245
#> 3    AY 2014/2015              NAT dP0Zw HsQzL 21234
#> 4    AY 2013/2014              NAT dP0Zw HsQzL 21193
#> 5    AY 2012/2013              NAT dP0Zw HsQzL 21157
#> 6    AY 2011/2012              NAT dP0Zw HsQzL 21133
#> 7    AY 2010/2011              NAT dP0Zw HsQzL 21228
#> 8    AY 2009/2010              NAT dP0Zw HsQzL 21352
#> 9    AY 2008/2009              NAT dP0Zw HsQzL 21685
#> 10   AY 2007/2008              NAT dP0Zw HsQzL 21641
#> 11   AY 2006/2007              NAT dP0Zw HsQzL 21864