Get a parsed version of the API response for a data set's meta data

Get a list of metadata information for a data set available from the EES API. Provides either look-up tables from human readable labels to ids used in the API, or the raw response from the meta endpoint.

Usage

get_meta(dataset_id, dataset_version = NULL, api_version = NULL)

Arguments

dataset_id

ID of data set to be connected to

dataset_version

Version of data set to be connected to

api_version

EES API version

Value

List of data frames containing a data set's meta data

Examples

get_meta(example_id())
#> $time_periods
#>    code period        label
#> 1    W1   2024  2024 Week 1
#> 2    W2   2024  2024 Week 2
#> 3    W3   2024  2024 Week 3
#> 4    W4   2024  2024 Week 4
#> 5    W5   2024  2024 Week 5
#> 6    W6   2024  2024 Week 6
#> 7    W7   2024  2024 Week 7
#> 8    W8   2024  2024 Week 8
#> 9    W9   2024  2024 Week 9
#> 10  W10   2024 2024 Week 10
#> 11  W11   2024 2024 Week 11
#> 12  W12   2024 2024 Week 12
#> 13  W13   2024 2024 Week 13
#> 14  W15   2024 2024 Week 15
#> 15  W16   2024 2024 Week 16
#> 16  W17   2024 2024 Week 17
#> 17  W18   2024 2024 Week 18
#> 18  W19   2024 2024 Week 19
#> 19  W20   2024 2024 Week 20
#> 20  W21   2024 2024 Week 21
#> 21  W23   2024 2024 Week 23
#> 22  W24   2024 2024 Week 24
#> 23  W25   2024 2024 Week 25
#> 24  W26   2024 2024 Week 26
#> 25  W37   2023 2023 Week 37
#> 26  W38   2023 2023 Week 38
#> 27  W39   2023 2023 Week 39
#> 28  W40   2023 2023 Week 40
#> 29  W41   2023 2023 Week 41
#> 30  W42   2023 2023 Week 42
#> 31  W43   2023 2023 Week 43
#> 32  W44   2023 2023 Week 44
#> 33  W45   2023 2023 Week 45
#> 34  W46   2023 2023 Week 46
#> 35  W47   2023 2023 Week 47
#> 36  W48   2023 2023 Week 48
#> 37  W49   2023 2023 Week 49
#> 38  W50   2023 2023 Week 50
#> 39  W51   2023 2023 Week 51
#> 
#> $locations
#>     geographic_level      code                               label item_id
#> 1    Local authority E09000002                Barking and Dagenham   09WT5
#> 2    Local authority E09000003                              Barnet   HSmzL
#> 3    Local authority E08000016                            Barnsley   bgKtT
#> 4    Local authority E06000022        Bath and North East Somerset   wgAbx
#> 5    Local authority E06000055                             Bedford   YS4HK
#> 6    Local authority E09000004                              Bexley   7zXob
#> 7    Local authority E08000025                          Birmingham   eRtuS
#> 8    Local authority E06000008               Blackburn with Darwen   WxL2b
#> 9    Local authority E06000009                           Blackpool   EIEqF
#> 10   Local authority E08000001                              Bolton   uk7o4
#> 11   Local authority E06000058 Bournemouth, Christchurch and Poole   k3Ehs
#> 12   Local authority E06000036                    Bracknell Forest   1b9i3
#> 13   Local authority E08000032                            Bradford   4LeUZ
#> 14   Local authority E09000005                               Brent   CH0Id
#> 15   Local authority E06000043                   Brighton and Hove   tCoEm
#> 16   Local authority E06000023                    Bristol, City of   asELP
#> 17   Local authority E09000006                             Bromley   Yp9HK
#> 18   Local authority E06000060                     Buckinghamshire   daf0Z
#> 19   Local authority E08000002                                Bury   lM0cB
#> 20   Local authority E08000033                          Calderdale   myP9K
#> 21   Local authority E10000003                      Cambridgeshire   oyzmX
#> 22   Local authority E09000007                              Camden   ohkmX
#> 23   Local authority E06000056                Central Bedfordshire   qBVjG
#> 24   Local authority E06000049                       Cheshire East   qFjG7
#> 25   Local authority E06000050           Cheshire West and Chester   NAEDC
#> 26   Local authority E09000001                      City of London   wGjbx
#> 27   Local authority E06000052                            Cornwall   wOGbx
#> 28   Local authority E06000047                       County Durham   uyNo4
#> 29   Local authority E08000026                            Coventry   T4bPJ
#> 30   Local authority E09000008                             Croydon   qISjG
#> 31   Local authority E06000063                          Cumberland   lQ4cB
#> 32   Local authority E06000005                          Darlington   c1X31
#> 33   Local authority E06000015                               Derby   z5aFQ
#> 34   Local authority E10000007                          Derbyshire   HTzLj
#> 35   Local authority E10000008                               Devon   Su7Vx
#> 36   Local authority E08000017                           Doncaster   DjxQe
#> 37   Local authority E06000059                              Dorset   zdLFQ
#> 38   Local authority E08000027                              Dudley   hc7yW
#> 39   Local authority E09000009                              Ealing   a8bLP
#> 40   Local authority E06000011            East Riding of Yorkshire   LsrWj
#> 41   Local authority E10000011                         East Sussex   5midi
#> 42   Local authority E09000010                             Enfield   BFm7J
#> 43   Local authority E10000012                               Essex   YPHKM
#> 44   Local authority E08000037                           Gateshead   YVUHK
#> 45   Local authority E10000013                     Gloucestershire   9WW4v
#> 46   Local authority E09000011                           Greenwich   9aa4v
#> 47   Local authority E09000012                             Hackney   dP0Zw
#> 48   Local authority E06000006                              Halton   Gfigr
#> 49   Local authority E09000013              Hammersmith and Fulham   dm90Z
#> 50   Local authority E10000014                           Hampshire   fpcYF
#> 51   Local authority E09000014                            Haringey   zVyFQ
#> 52   Local authority E09000015                              Harrow   pTSoj
#> 53   Local authority E06000001                          Hartlepool   OHFCL
#> 54   Local authority E09000016                            Havering   klbhs
#> 55   Local authority E06000019            Herefordshire, County of   XpafK
#> 56   Local authority E10000015                       Hertfordshire   d4U0Z
#> 57   Local authority E09000017                          Hillingdon   uk6o4
#> 58   Local authority E09000018                            Hounslow   XvhfK
#> 59   Local authority E06000046                       Isle of Wight   OzJCL
#> 60   Local authority E06000053                     Isles of Scilly   q75jG
#> 61   Local authority E09000019                           Islington   OYzCL
#> 62   Local authority E09000020              Kensington and Chelsea   7JpXo
#> 63   Local authority E10000016                                Kent   4WuUZ
#> 64   Local authority E06000010         Kingston upon Hull, City of   Bv47J
#> 65   Local authority E09000021                Kingston upon Thames   T4iPJ
#> 66   Local authority E08000034                            Kirklees   09VT5
#> 67   Local authority E08000011                            Knowsley   Vut5X
#> 68   Local authority E09000022                             Lambeth   O7CLF
#> 69   Local authority E10000017                          Lancashire   045T5
#> 70   Local authority E08000035                               Leeds   Wsa2b
#> 71   Local authority E06000016                           Leicester   h3XyW
#> 72   Local authority E10000018                      Leicestershire   TUtPJ
#> 73   Local authority E09000023                            Lewisham   pPmSo
#> 74   Local authority E10000019                        Lincolnshire   7nVXo
#> 75   Local authority E08000012                           Liverpool   PUgeb
#> 76   Local authority E06000032                               Luton   CzTId
#> 77   Local authority E08000003                          Manchester   UWJRF
#> 78   Local authority E06000035                              Medway   mae9K
#> 79   Local authority E09000024                              Merton   cK731
#> 80   Local authority E06000002                       Middlesbrough   7vSXo
#> 81   Local authority E06000042                       Milton Keynes   BJ97J
#> 82   Local authority E08000021                 Newcastle upon Tyne   QAQwb
#> 83   Local authority E09000025                              Newham   5o1di
#> 84   Local authority E10000020                             Norfolk   pu4So
#> 85   Local authority E06000012             North East Lincolnshire   0kT5D
#> 86   Local authority E06000013                  North Lincolnshire   6EIrf
#> 87   Local authority E06000061              North Northamptonshire   cif31
#> 88   Local authority E06000024                      North Somerset   9LL4v
#> 89   Local authority E08000022                      North Tyneside   9LU4v
#> 90   Local authority E06000065                     North Yorkshire   1gai3
#> 91   Local authority E06000057                      Northumberland   TRrPJ
#> 92   Local authority E06000018                          Nottingham   9WR4v
#> 93   Local authority E10000024                     Nottinghamshire   Qgbwb
#> 94   Local authority E08000004                              Oldham   ohHmX
#> 95   Local authority E10000025                         Oxfordshire   LxNWj
#> 96   Local authority E06000031                        Peterborough   rqWNj
#> 97   Local authority E06000026                            Plymouth   y8eaB
#> 98   Local authority E06000044                          Portsmouth   j7RAM
#> 99   Local authority E06000038                             Reading   CpId1
#> 100  Local authority E09000026                           Redbridge   IzBzg
#> 101  Local authority E06000003                Redcar and Cleveland   pe9So
#> 102  Local authority E09000027                Richmond upon Thames   tOuEm
#> 103  Local authority E08000005                            Rochdale   rELNj
#> 104  Local authority E08000018                           Rotherham   eRSuS
#> 105  Local authority E06000017                             Rutland   maO9K
#> 106  Local authority E08000006                             Salford   GX1gr
#> 107  Local authority E08000028                            Sandwell   myg9K
#> 108  Local authority E08000014                              Sefton   1wmi3
#> 109  Local authority E08000019                           Sheffield   5oJdi
#> 110  Local authority E06000051                          Shropshire   2K8lz
#> 111  Local authority E06000039                              Slough   lKTcB
#> 112  Local authority E08000029                            Solihull   gf2O9
#> 113  Local authority E06000066                            Somerset   ucvo4
#> 114  Local authority E06000025               South Gloucestershire   d940Z
#> 115  Local authority E08000023                      South Tyneside   2M3lz
#> 116  Local authority E06000045                         Southampton   ekEuS
#> 117  Local authority E06000033                     Southend-on-Sea   bO1tT
#> 118  Local authority E09000028                           Southwark   IzpBz
#> 119  Local authority E08000013                          St. Helens   LxWjE
#> 120  Local authority E10000028                       Staffordshire   gYlO9
#> 121  Local authority E08000007                           Stockport   3198u
#> 122  Local authority E06000004                    Stockton-on-Tees   IFSBz
#> 123  Local authority E06000021                      Stoke-on-Trent   AX2GK
#> 124  Local authority E10000029                             Suffolk   6qBrf
#> 125  Local authority E08000024                          Sunderland   PRmeb
#> 126  Local authority E10000030                              Surrey   0UlT5
#> 127  Local authority E09000029                              Sutton   fnlYF
#> 128  Local authority E06000030                             Swindon   PROeb
#> 129  Local authority E08000008                            Tameside   SeBVx
#> 130  Local authority E06000020                  Telford and Wrekin   Pk9eb
#> 131  Local authority E06000034                            Thurrock   DAMQe
#> 132  Local authority E06000027                              Torbay   RlTka
#> 133  Local authority E09000030                       Tower Hamlets   ibG6X
#> 134  Local authority E08000009                            Trafford   y3PaB
#> 135  Local authority E08000036                           Wakefield   ECtqF
#> 136  Local authority E08000030                             Walsall   it6Xr
#> 137  Local authority E09000031                      Waltham Forest   jBUAM
#> 138  Local authority E09000032                          Wandsworth   LZ6Wj
#> 139  Local authority E06000007                          Warrington   3i48u
#> 140  Local authority E10000031                        Warwickshire   H54zL
#> 141  Local authority E06000037                      West Berkshire   uhXo4
#> 142  Local authority E06000062               West Northamptonshire   596di
#> 143  Local authority E10000032                         West Sussex   CFPId
#> 144  Local authority E09000033                         Westminster   6tCrf
#> 145  Local authority E06000064             Westmorland and Furness   Ux0RF
#> 146  Local authority E08000010                               Wigan   R20ka
#> 147  Local authority E06000054                           Wiltshire   VoE5X
#> 148  Local authority E06000040              Windsor and Maidenhead   d9g0Z
#> 149  Local authority E08000015                              Wirral   dmG0Z
#> 150  Local authority E06000041                           Wokingham   U2vRF
#> 151  Local authority E08000031                       Wolverhampton   A7WGK
#> 152  Local authority E10000034                      Worcestershire   5mvdi
#> 153  Local authority E06000014                                York   HV9zL
#> 154         National E92000001                             England   BT7J3
#> 
#> $filter_columns
#>              col_name               label
#> 1   attendance_status   Attendance status
#> 2     attendance_type     Attendance type
#> 3          day_number          Day number
#> 4 establishment_phase Establishment phase
#> 5              reason              Reason
#> 
#> $filter_items
#>               col_name                                 item_label item_id
#> 1    attendance_status                                    Absence   uRBo4
#> 2    attendance_status                                 Attendance   lxdcB
#> 3    attendance_status                              Late sessions   ULFRF
#> 4    attendance_status                             Not determined   okYmX
#> 5    attendance_status                          Possible sessions   rtaNj
#> 6      attendance_type              Approved educational activity   G7Kgr
#> 7      attendance_type                                 Authorised   3Km8u
#> 8      attendance_type                             Not determined   SPIVx
#> 9      attendance_type                                    Present   ycTaB
#> 10     attendance_type                               Unauthorised   RWdka
#> 11          day_number                                          1   bGMtT
#> 12          day_number                                          2   DqsQe
#> 13          day_number                                          3   eUpuS
#> 14          day_number                                          4   5uFdi
#> 15          day_number                                          5   4aPUZ
#> 16          day_number                                      Total   mhT9K
#> 17 establishment_phase                                    Primary   0mST5
#> 18 establishment_phase                                  Secondary   WZy2b
#> 19 establishment_phase                                    Special   EObqF
#> 20 establishment_phase                                      Total   QjDwb
#> 21              reason                       B education off site   9y94v
#> 22              reason                         C authorised other   2xclz
#> 23              reason                        D dual registration   PUoeb
#> 24              reason                      E authorised excluded   NqbDC
#> 25              reason                                   Excluded   wGjbx
#> 26              reason                    F legacy family holiday   ohkmX
#> 27              reason                     G unauthorised holiday   9aa4v
#> 28              reason                       H authorised holiday   dm90Z
#> 29              reason                       I authorised illness   OYzCL
#> 30              reason                                J interview   7JpXo
#> 31              reason     L present late before registers closed   pPmSo
#> 32              reason                M authorised medical dental   IzpBz
#> 33              reason                            N no reason yet   ibG6X
#> 34              reason               Not attending planned closed   LZ6Wj
#> 35              reason                       O other unauthorised   6tCrf
#> 36              reason               P approved sporting activity   HSmzL
#> 37              reason                                         Pa   09WT5
#> 38              reason                                    Present   CH0Id
#> 39              reason                                 Present am   Yp9HK
#> 40              reason                                 Present pm   qISjG
#> 41              reason          R authorised religious observance   a8bLP
#> 42              reason                   S authorised study leave   BFm7J
#> 43              reason                   T authorised grt absence   zVyFQ
#> 44              reason                                      Total   klbhs
#> 45              reason U unauthorised late after registers closed   uk6o4
#> 46              reason                   V educational visit trip   XvhfK
#> 47              reason                          W work experience   T4iPJ
#> 48              reason           Y not attending enforced closure   cK731
#> 49              reason                Z not attending not on roll   5o1di
#>    default_item
#> 1            NA
#> 2            NA
#> 3            NA
#> 4            NA
#> 5            NA
#> 6            NA
#> 7            NA
#> 8            NA
#> 9            NA
#> 10           NA
#> 11           NA
#> 12           NA
#> 13           NA
#> 14           NA
#> 15           NA
#> 16         TRUE
#> 17           NA
#> 18           NA
#> 19           NA
#> 20         TRUE
#> 21           NA
#> 22           NA
#> 23           NA
#> 24           NA
#> 25           NA
#> 26           NA
#> 27           NA
#> 28           NA
#> 29           NA
#> 30           NA
#> 31           NA
#> 32           NA
#> 33           NA
#> 34           NA
#> 35           NA
#> 36           NA
#> 37           NA
#> 38           NA
#> 39           NA
#> 40           NA
#> 41           NA
#> 42           NA
#> 43           NA
#> 44         TRUE
#> 45           NA
#> 46           NA
#> 47           NA
#> 48           NA
#> 49           NA
#> 
#> $indicators
#>          col_name           label
#> 1   session_count   Session count
#> 2 session_percent Session percent
#> 3  session_scaled  Session scaled
#>