Get the base API response for a data set's meta data

Get the metadata information for a data set available from the EES API.

Usage

get_meta_response(
  dataset_id,
  dataset_version = NULL,
  api_version = NULL,
  parse = TRUE
)

Arguments

dataset_id

ID of data set to be connected to

dataset_version

Version of data set to be connected to

api_version

EES API version

parse

Parse result into structured list

Value

Results of query to API meta data endpoint

Examples

get_meta_response(example_id())
#> $filters
#>      id      column       label                  hint
#> 1 5mvdi school_type School type Filter by school type
#>                                                                                                      options
#> 1 LfnWj, 6TUrf, 0wOT5, HsQzL, Special, State-funded primary, State-funded secondary, Total, NA, NA, NA, TRUE
#> 
#> $indicators
#>      id              column
#> 1 tImEm     fixed_excl_rate
#> 2 fS3YF          fixed_excl
#> 3 jXkAM           perm_excl
#> 4 ehquS           headcount
#> 5 TREPJ      one_plus_fixed
#> 6 h8fyW         num_schools
#> 7 mLC9K one_plus_fixed_rate
#> 8 gHcO9      perm_excl_rate
#>                                                           label
#> 1                                   Fixed period exclusion rate
#> 2                             Number of fixed period exclusions
#> 3                                Number of permanent exclusions
#> 4                                              Number of pupils
#> 5     Number of pupils with one or more fixed period exclusions
#> 6                                             Number of schools
#> 7 Percentage of pupils with one or more fixed period exclusions
#> 8                                      Permanent exclusion rate
#> 
#> $geographicLevels
#>   code           label
#> 1   LA Local authority
#> 2  NAT        National
#> 3  REG        Regional
#> 4  SCH          School
#> 
#> $locations
#>   level.code     level.label
#> 1         LA Local authority
#> 2        NAT        National
#> 3        REG        Regional
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  options
#> 1 S9YVx, 21Zlz, QOtwb, PhIeb, pmRSo, yxmaB, gF8O9, XSLfK, TkNPJ, 6AXrf, oLJmX, Dg1Qe, l6ucB, R8uka, EXAqF, p4hSo, bBrtT, Y0hHK, 0k3T5, U3SRF, qYTjG, DOiQe, IndBz, wIObx, oWLmX, 3Nk8u, dRk0Z, AH3GK, ea0uS, aKALP, z0WFQ, eyjuS, BSk7J, zPlFQ, t74Em, ksAhs, V4A5X, ISTBz, 5USdi, 54rdi, uWwo4, 4c9UZ, X52fK, jYyAM, EYgqF, TyQPJ, m2m9K, 0hPT5, k8qhs, Wic2b, cNz31, EfGqF, QIqwb, Y5RHK, 9cy4v, 75TXo, 5RDdi, 26xlz, PaReb, 2islz, iG76X, NOADC, wBCbx, tXUEm, OnSCL, o2WmX, oWvmX, uLQo4, 9cc4v, fdsYF, rGcNj, TkqPJ, jHaAM, dkh0Z, eLGuS, XVWfK, SmfVx, H0RzL, bI6tT, OvpCL, qXdjG, W1l2b, eaAuS, 7V0Xo, TyAPJ, tB9Em, fV8YF, dUC0Z, ThjPJ, ml79K, hxzyW, NgjDC, b7NtT, mWI9K, gJZO9, CvuId, Az8GK, 3rR8u, liecB, rAlNj, QCEwb, 54pdi, p9kSo, axqLP, Ip0Bz, YdkHK, f5ZYF, hZYyW, qf0jG, PyBeb, crH31, 92k4v, 9ll4v, 4lmUZ, 1qRi3, VUG5X, uaYo4, hsHyW, 9lh4v, ysoaB, iAB6X, LibWj, Pa7eb, aMjLP, BnU7J, V9j5X, 1jki3, m2B9K, dkB0Z, L1wWj, GHrgr, BpR7J, AYcGK, R3eka, UcpRF, 6Gfrf, zT2FQ, GzNgr, dRI0Z, 0h8T5, HTkzL, uaDo4, uWzo4, eyduS, lrocB, CnoId, kZjhs, 7SdXo, mWo9K, 5UNdi, 0R0T5, uLHo4, U8VRF, jfhAM, Barking and Dagenham, Barnet, Barnsley, Bath and North East Somerset, Bedford, Bexley, Birmingham, Blackburn with Darwen, Blackpool, Bolton, Bournemouth, Bracknell Forest, Bradford, Brent, Brighton and Hove, Bristol City of, Bromley, Buckinghamshire, Bury, Calderdale, Cambridgeshire, Camden, Central Bedfordshire, Cheshire East, Cheshire West and Chester, City of London, Cornwall, Coventry, Croydon, Cumbria, Darlington, Derby, Derbyshire, Devon, Doncaster, Dorset, Dudley, Durham, Ealing, East Riding of Yorkshire, East Sussex, Enfield, Essex, Gateshead, Gateshead, Gloucestershire, Greenwich, Hackney, Halton, Hammersmith and Fulham, Hampshire, Haringey, Harrow, Hartlepool, Havering, Herefordshire, Hertfordshire, Hillingdon, Hounslow, Isle of Wight, Isles Of Scilly, Islington, Kensington and Chelsea, Kent, Kingston upon Hull City of, Kingston upon Thames, Kirklees, Knowsley, Lambeth, Lancashire, Leeds, Leicester, Leicestershire, Lewisham, Lincolnshire, Liverpool, Luton, Manchester, Medway, Merton, Middlesbrough, Milton Keynes, Newcastle upon Tyne, Newham, Norfolk, North East Lincolnshire, North Lincolnshire, North Somerset, North Tyneside, North Yorkshire, Northamptonshire, Northumberland, Northumberland, Nottingham, Nottinghamshire, Oldham, Oxfordshire, Peterborough, Plymouth, Poole, Portsmouth, Reading, Redbridge, Redcar and Cleveland, Richmond upon Thames, Rochdale, Rotherham, Rutland, Salford, Sandwell, Sefton, Sheffield, Shropshire, Slough, Solihull, Somerset, South Gloucestershire, South Tyneside, Southampton, Southend-on-Sea, Southwark, St Helens, Staffordshire, Stockport, Stockton-on-Tees, Stoke-on-Trent, Suffolk, Sunderland, Surrey, Sutton, Swindon, Tameside, Telford and Wrekin, Thurrock, Torbay, Tower Hamlets, Trafford, Wakefield, Walsall, Waltham Forest, Wandsworth, Warrington, Warwickshire, West Berkshire, West Sussex, Westminster, Wigan, Wiltshire, Windsor and Maidenhead, Wirral, Wokingham, Wolverhampton, Worcestershire, York, E09000002, E09000003, E08000016, E06000022, E06000055, E09000004, E08000025, E06000008, E06000009, E08000001, E06000028, E06000036, E08000032, E09000005, E06000043, E06000023, E09000006, E10000002, E08000002, E08000033, E10000003, E09000007, E06000056, E06000049, E06000050, E09000001, E06000052, E08000026, E09000008, E10000006, E06000005, E06000015, E10000007, E10000008, E08000017, E10000009, E08000027, E06000047, E09000009, E06000011, E10000011, E09000010, E10000012, E08000020, E08000037, E10000013, E09000011, E09000012, E06000006, E09000013, E10000014, E09000014, E09000015, E06000001, E09000016, E06000019, E10000015, E09000017, E09000018, E06000046, E06000053, E09000019, E09000020, E10000016, E06000010, E09000021, E08000034, E08000011, E09000022, E10000017, E08000035, E06000016, E10000018, E09000023, E10000019, E08000012, E06000032, E08000003, E06000035, E09000024, E06000002, E06000042, E08000021, E09000025, E10000020, E06000012, E06000013, E06000024, E08000022, E10000023, E10000021, E06000048, E06000057, E06000018, E10000024, E08000004, E10000025, E06000031, E06000026, E06000029, E06000044, E06000038, E09000026, E06000003, E09000027, E08000005, E08000018, E06000017, E08000006, E08000028, E08000014, E08000019, E06000051, E06000039, E08000029, E10000027, E06000025, E08000023, E06000045, E06000033, E09000028, E08000013, E10000028, E08000007, E06000004, E06000021, E10000029, E08000024, E10000030, E09000029, E06000030, E08000008, E06000020, E06000034, E06000027, E09000030, E08000009, E08000036, E08000030, E09000031, E09000032, E06000007, E10000031, E06000037, E10000032, E09000033, E08000010, E06000054, E06000040, E08000015, E06000041, E08000031, E10000034, E06000014, 301, 302, 370, 800, 822, 303, 330, 889, 890, 350, 837, 867, 380, 304, 846, 801, 305, 825, 351, 381, 873, 202, 823, 895, 896, 201, 908, 331, 306, 909, 841, 831, 830, 878, 371, 835, 332, 840, 307, 811, 845, 308, 881, 390, 390, 916, 203, 204, 876, 205, 850, 309, 310, 805, 311, 884, 919, 312, 313, 921, 420, 206, 207, 886, 810, 314, 382, 340, 208, 888, 383, 856, 855, 209, 925, 341, 821, 352, 887, 315, 806, 826, 391, 316, 926, 812, 813, 802, 392, 815, 928, 929, 929, 892, 891, 353, 931, 874, 879, 836, 851, 870, 317, 807, 318, 354, 372, 857, 355, 333, 343, 373, 893, 871, 334, 933, 803, 393, 852, 882, 210, 342, 860, 356, 808, 861, 935, 394, 936, 319, 866, 357, 894, 883, 880, 211, 358, 384, 335, 320, 212, 877, 937, 869, 938, 213, 359, 865, 868, 344, 872, 336, 885, 816
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              dP0Zw, England, E92000001
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                rg3Nj, 3Dh8u, S4HVx, O7CLF, o2nmX, wCgbx, yM9aB, RZoka, GFDgr, NGqDC, East Midlands, East of England, Inner London, North East, North West, Outer London, South East, South West, West Midlands, Yorkshire and the Humber, E12000004, E12000006, E13000001, E12000001, E12000002, E13000002, E12000008, E12000009, E12000005, E12000003
#> 
#> $timePeriods
#>    code    period   label
#> 1    AY 2006/2007 2006/07
#> 2    AY 2007/2008 2007/08
#> 3    AY 2008/2009 2008/09
#> 4    AY 2009/2010 2009/10
#> 5    AY 2010/2011 2010/11
#> 6    AY 2011/2012 2011/12
#> 7    AY 2012/2013 2012/13
#> 8    AY 2013/2014 2013/14
#> 9    AY 2014/2015 2014/15
#> 10   AY 2015/2016 2015/16
#> 11   AY 2016/2017 2016/17
#>