Get the base API response for a data set's meta data

Get the metadata information for a data set available from the EES API.

Usage

get_meta_response(
  dataset_id,
  dataset_version = NULL,
  api_version = NULL,
  parse = TRUE,
  verbose = FALSE
)

Arguments

dataset_id

ID of data set to be connected to. This is required if the endpoint is one of "get-summary", "get-meta", "get-data" or "post-data"

dataset_version

Version of data set to be connected to

api_version

EES API version

parse

Parse result into structured list

verbose

Run with additional contextual messaging, logical, default = FALSE

Value

Results of query to API meta data endpoint

Examples

get_meta_response(example_id())
#> $filters
#>      id             column              label hint
#> 1 tu1Em establishment_type establishment type     
#> 2 f2GYF               type               type     
#>                                                                                                                                                                                                                                                                                                                                                                                         options
#> 1                                                                                                                                                                                                                            9RpY4, 23odl, Poz7e, NwcLD, wEZcb, o4vFm, Academy Converter, Academy Sponsor Led, Community School, Independent School, Sixth Form College, Voluntary Aided School
#> 2 9Rpn4, dPff0, OBNDC, 7YbEX, p5IXS, IXQEB, iNW86, 6D2xr, 0eYST, LleUW, HJjXz, C7DgI, YFqfH, qGKqj, a2eWL, Bfbf7, zvGnF, ka0Wh, u9vRo, TuzUP, cZ4i3, XHEvf, 1618, 1618_DIS, 1618_NOTDIS, ACAD, ACAD_DIS, ACAD_NOTDIS, AGEN, AGEN_DIS, AGEN_NOTDIS, AGENDENOM, ALEV, ALEV_DIS, ALEV_NOTDIS, ANYL2VOC, SCOPE_L3M, TechCert, TechCert_DIS, TechCert_NOTDIS, TLEV, TLEV_DIS, TLEV_NOTDIS, TLEVDENOM
#> 
#> $indicators
#>      id               column                label
#> 1 E40qF            ppe_value            ppe value
#> 2 QuVwb         ppegrd_value         ppegrd value
#> 3 9Aw4v retained_2ndyr_value retained 2ndyr value
#> 4 2YIlz       retained_value       retained value
#> 5 PbNeb    total_pupil_value    total pupil value
#> 
#> $geographicLevels
#>   code  label
#> 1  SCH School
#> 
#> $locations
#>   level.code     level.label
#> 1         LA Local authority
#> 2        NAT        National
#> 3        SCH          School
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  options
#> 1                                                                                                                                                                                  ea0uS, 4c9UZ, QIqwb, rGcNj, eLGuS, H0RzL, VUG5X, lrocB, jfhAM, Croydon, Enfield, Harrow, Leeds, Lincolnshire, Manchester, Somerset, West Sussex, York, E09000008, E09000010, E09000015, E08000035, E10000019, E08000003, E10000027, E10000032, E06000014, 306, 308, 310, 383, 925, 352, 933, 938, 816
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                              dP0Zw, England, E92000001
#> 3 42nEU, y8Jra, k2qHh, mLJt9, 59BKd, dNu0Z, 0UXNT, BzBp7, hciHy, HS5ez, Bootham School, Brigshaw High School, Felpham Community College, Loreto College, North London Collegiate School, The Latymer School, The Laurels School, The Priory Ruskin Academy, University Academy Holbeach, West Somerset College, 121722, 143238, 126080, 130503, 102257, 102055, 140356, 136194, 137282, 136791, 8166000, 3834113, 9384059, 3528601, 3106075, 3085400, 3066028, 9256910, 9254001, 9334291
#> 
#> $timePeriods
#>   code    period   label
#> 1   AY 2022/2023 2022/23
#>