Get the base API response for a data set's meta data

Get the metadata information for a data set available from the EES API.

Usage

get_meta_response(
  dataset_id,
  dataset_version = NULL,
  api_version = NULL,
  parse = TRUE
)

Arguments

dataset_id

ID of data set to be connected to

dataset_version

Version of data set to be connected to

api_version

EES API version

parse

Parse result into structured list

Value

Results of query to API meta data endpoint

Examples

get_meta_response(example_id())
#> $filters
#>                    id               label hint
#> 1   attendance_status   Attendance status     
#> 2     attendance_type     Attendance type     
#> 3          day_number          Day number     
#> 4 establishment_phase Establishment phase     
#> 5              reason              Reason     
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        options
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     uRBo4, lxdcB, ULFRF, okYmX, rtaNj, Absence, Attendance, Late sessions, Not determined, Possible sessions
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          G7Kgr, 3Km8u, SPIVx, ycTaB, RWdka, Approved educational activity, Authorised, Not determined, Present, Unauthorised
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     bGMtT, DqsQe, eUpuS, 5uFdi, 4aPUZ, mhT9K, 1, 2, 3, 4, 5, Total, NA, NA, NA, NA, NA, TRUE
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             0mST5, WZy2b, EObqF, QjDwb, Primary, Secondary, Special, Total, NA, NA, NA, TRUE
#> 5 9y94v, 2xclz, PUoeb, NqbDC, wGjbx, ohkmX, 9aa4v, dm90Z, OYzCL, 7JpXo, pPmSo, IzpBz, ibG6X, LZ6Wj, 6tCrf, HSmzL, 09WT5, CH0Id, Yp9HK, qISjG, a8bLP, BFm7J, zVyFQ, klbhs, uk6o4, XvhfK, T4iPJ, cK731, 5o1di, B education off site, C authorised other, D dual registration, E authorised excluded, Excluded, F legacy family holiday, G unauthorised holiday, H authorised holiday, I authorised illness, J interview, L present late before registers closed, M authorised medical dental, N no reason yet, Not attending planned closed, O other unauthorised, P approved sporting activity, Pa, Present, Present am, Present pm, R authorised religious observance, S authorised study leave, T authorised grt absence, Total, U unauthorised late after registers closed, V educational visit trip, W work experience, Y not attending enforced closure, Z not attending not on roll, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, TRUE, NA, NA, NA, NA, NA
#> 
#> $indicators
#>                id           label
#> 1   session_count   Session count
#> 2 session_percent Session percent
#> 3  session_scaled  Session scaled
#> 
#> $geographicLevels
#>   code           label
#> 1   LA Local authority
#> 2  NAT        National
#> 3  REG        Regional
#> 
#> $locations
#>   level.code     level.label
#> 1         LA Local authority
#> 2        NAT        National
#> 3        REG        Regional
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               options
#> 1 09WT5, HSmzL, bgKtT, wgAbx, YS4HK, 7zXob, eRtuS, WxL2b, EIEqF, uk7o4, k3Ehs, 1b9i3, 4LeUZ, CH0Id, tCoEm, asELP, Yp9HK, daf0Z, lM0cB, myP9K, oyzmX, ohkmX, qBVjG, qFjG7, NAEDC, wGjbx, wOGbx, uyNo4, T4bPJ, qISjG, lQ4cB, c1X31, z5aFQ, HTzLj, Su7Vx, DjxQe, zdLFQ, hc7yW, a8bLP, LsrWj, 5midi, BFm7J, YPHKM, YVUHK, 9WW4v, 9aa4v, dP0Zw, Gfigr, dm90Z, fpcYF, zVyFQ, pTSoj, OHFCL, klbhs, XpafK, d4U0Z, uk6o4, XvhfK, OzJCL, q75jG, OYzCL, 7JpXo, 4WuUZ, Bv47J, T4iPJ, 09VT5, Vut5X, O7CLF, 045T5, Wsa2b, h3XyW, TUtPJ, pPmSo, 7nVXo, PUgeb, CzTId, UWJRF, mae9K, cK731, 7vSXo, BJ97J, QAQwb, 5o1di, pu4So, 0kT5D, 6EIrf, cif31, 9LL4v, 9LU4v, 1gai3, TRrPJ, 9WR4v, Qgbwb, ohHmX, LxNWj, rqWNj, y8eaB, j7RAM, CpId1, IzBzg, pe9So, tOuEm, rELNj, eRSuS, maO9K, GX1gr, myg9K, 1wmi3, 5oJdi, 2K8lz, lKTcB, gf2O9, ucvo4, d940Z, 2M3lz, ekEuS, bO1tT, IzpBz, LxWjE, gYlO9, 3198u, IFSBz, AX2GK, 6qBrf, PRmeb, 0UlT5, fnlYF, PROeb, SeBVx, Pk9eb, DAMQe, RlTka, ibG6X, y3PaB, ECtqF, it6Xr, jBUAM, LZ6Wj, 3i48u, H54zL, uhXo4, 596di, CFPId, 6tCrf, Ux0RF, R20ka, VoE5X, d9g0Z, dmG0Z, U2vRF, A7WGK, 5mvdi, HV9zL, Barking and Dagenham, Barnet, Barnsley, Bath and North East Somerset, Bedford, Bexley, Birmingham, Blackburn with Darwen, Blackpool, Bolton, Bournemouth, Christchurch and Poole, Bracknell Forest, Bradford, Brent, Brighton and Hove, Bristol, City of, Bromley, Buckinghamshire, Bury, Calderdale, Cambridgeshire, Camden, Central Bedfordshire, Cheshire East, Cheshire West and Chester, City of London, Cornwall, County Durham, Coventry, Croydon, Cumberland, Darlington, Derby, Derbyshire, Devon, Doncaster, Dorset, Dudley, Ealing, East Riding of Yorkshire, East Sussex, Enfield, Essex, Gateshead, Gloucestershire, Greenwich, Hackney, Halton, Hammersmith and Fulham, Hampshire, Haringey, Harrow, Hartlepool, Havering, Herefordshire, County of, Hertfordshire, Hillingdon, Hounslow, Isle of Wight, Isles of Scilly, Islington, Kensington and Chelsea, Kent, Kingston upon Hull, City of, Kingston upon Thames, Kirklees, Knowsley, Lambeth, Lancashire, Leeds, Leicester, Leicestershire, Lewisham, Lincolnshire, Liverpool, Luton, Manchester, Medway, Merton, Middlesbrough, Milton Keynes, Newcastle upon Tyne, Newham, Norfolk, North East Lincolnshire, North Lincolnshire, North Northamptonshire, North Somerset, North Tyneside, North Yorkshire, Northumberland, Nottingham, Nottinghamshire, Oldham, Oxfordshire, Peterborough, Plymouth, Portsmouth, Reading, Redbridge, Redcar and Cleveland, Richmond upon Thames, Rochdale, Rotherham, Rutland, Salford, Sandwell, Sefton, Sheffield, Shropshire, Slough, Solihull, Somerset, South Gloucestershire, South Tyneside, Southampton, Southend-on-Sea, Southwark, St. Helens, Staffordshire, Stockport, Stockton-on-Tees, Stoke-on-Trent, Suffolk, Sunderland, Surrey, Sutton, Swindon, Tameside, Telford and Wrekin, Thurrock, Torbay, Tower Hamlets, Trafford, Wakefield, Walsall, Waltham Forest, Wandsworth, Warrington, Warwickshire, West Berkshire, West Northamptonshire, West Sussex, Westminster, Westmorland and Furness, Wigan, Wiltshire, Windsor and Maidenhead, Wirral, Wokingham, Wolverhampton, Worcestershire, York, E09000002, E09000003, E08000016, E06000022, E06000055, E09000004, E08000025, E06000008, E06000009, E08000001, E06000058, E06000036, E08000032, E09000005, E06000043, E06000023, E09000006, E06000060, E08000002, E08000033, E10000003, E09000007, E06000056, E06000049, E06000050, E09000001, E06000052, E06000047, E08000026, E09000008, E06000063, E06000005, E06000015, E10000007, E10000008, E08000017, E06000059, E08000027, E09000009, E06000011, E10000011, E09000010, E10000012, E08000037, E10000013, E09000011, E09000012, E06000006, E09000013, E10000014, E09000014, E09000015, E06000001, E09000016, E06000019, E10000015, E09000017, E09000018, E06000046, E06000053, E09000019, E09000020, E10000016, E06000010, E09000021, E08000034, E08000011, E09000022, E10000017, E08000035, E06000016, E10000018, E09000023, E10000019, E08000012, E06000032, E08000003, E06000035, E09000024, E06000002, E06000042, E08000021, E09000025, E10000020, E06000012, E06000013, E06000061, E06000024, E08000022, E06000065, E06000057, E06000018, E10000024, E08000004, E10000025, E06000031, E06000026, E06000044, E06000038, E09000026, E06000003, E09000027, E08000005, E08000018, E06000017, E08000006, E08000028, E08000014, E08000019, E06000051, E06000039, E08000029, E06000066, E06000025, E08000023, E06000045, E06000033, E09000028, E08000013, E10000028, E08000007, E06000004, E06000021, E10000029, E08000024, E10000030, E09000029, E06000030, E08000008, E06000020, E06000034, E06000027, E09000030, E08000009, E08000036, E08000030, E09000031, E09000032, E06000007, E10000031, E06000037, E06000062, E10000032, E09000033, E06000064, E08000010, E06000054, E06000040, E08000015, E06000041, E08000031, E10000034, E06000014, 301, 302, 370, 800, 822, 303, 330, 889, 890, 350, 839, 867, 380, 304, 846, 801, 305, 825, 351, 381, 873, 202, 823, 895, 896, 201, 908, 840, 331, 306, 942, 841, 831, 830, 878, 371, 838, 332, 307, 811, 845, 308, 881, 390, 916, 203, 204, 876, 205, 850, 309, 310, 805, 311, 884, 919, 312, 313, 921, 420, 206, 207, 886, 810, 314, 382, 340, 208, 888, 383, 856, 855, 209, 925, 341, 821, 352, 887, 315, 806, 826, 391, 316, 926, 812, 813, 940, 802, 392, 815, 929, 892, 891, 353, 931, 874, 879, 851, 870, 317, 807, 318, 354, 372, 857, 355, 333, 343, 373, 893, 871, 334, 933, 803, 393, 852, 882, 210, 342, 860, 356, 808, 861, 935, 394, 936, 319, 866, 357, 894, 883, 880, 211, 358, 384, 335, 320, 212, 877, 937, 869, 941, 938, 213, 943, 359, 865, 868, 344, 872, 336, 885, 816
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           BT7J3, England, E92000001
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   X9fKb, cg31S, o3VmX, z4FQE, krhsL, 5Zdi9, tdEm5, TsPJP, jXkAM, East Midlands, East of England, London, North East, North West, South East, South West, West Midlands, Yorkshire and The Humber, E12000004, E12000006, E12000007, E12000001, E12000002, E12000008, E12000009, E12000005, E12000003
#> 
#> $timePeriods
#>    code period        label
#> 1    W1   2024  2024 Week 1
#> 2   W10   2024 2024 Week 10
#> 3   W11   2024 2024 Week 11
#> 4   W12   2024 2024 Week 12
#> 5   W13   2024 2024 Week 13
#> 6   W15   2024 2024 Week 15
#> 7   W16   2024 2024 Week 16
#> 8   W17   2024 2024 Week 17
#> 9   W18   2024 2024 Week 18
#> 10  W19   2024 2024 Week 19
#> 11   W2   2024  2024 Week 2
#> 12  W20   2024 2024 Week 20
#> 13  W21   2024 2024 Week 21
#> 14  W23   2024 2024 Week 23
#> 15  W24   2024 2024 Week 24
#> 16  W25   2024 2024 Week 25
#> 17  W26   2024 2024 Week 26
#> 18   W3   2024  2024 Week 3
#> 19  W37   2023 2023 Week 37
#> 20  W38   2023 2023 Week 38
#> 21  W39   2023 2023 Week 39
#> 22   W4   2024  2024 Week 4
#> 23  W40   2023 2023 Week 40
#> 24  W41   2023 2023 Week 41
#> 25  W42   2023 2023 Week 42
#> 26  W43   2023 2023 Week 43
#> 27  W44   2023 2023 Week 44
#> 28  W45   2023 2023 Week 45
#> 29  W46   2023 2023 Week 46
#> 30  W47   2023 2023 Week 47
#> 31  W48   2023 2023 Week 48
#> 32  W49   2023 2023 Week 49
#> 33   W5   2024  2024 Week 5
#> 34  W50   2023 2023 Week 50
#> 35  W51   2023 2023 Week 51
#> 36   W6   2024  2024 Week 6
#> 37   W7   2024  2024 Week 7
#> 38   W8   2024  2024 Week 8
#> 39   W9   2024  2024 Week 9
#>