Get the base API response for a data set's meta data

Get the metadata information for a data set available from the EES API.

Usage

get_meta_response(
  dataset_id,
  dataset_version = NULL,
  preview_token = NULL,
  ees_environment = NULL,
  api_version = NULL,
  parse = TRUE,
  verbose = FALSE
)

Arguments

dataset_id

ID of data set to be connected to. This is required if the endpoint is one of "get-dataset-versions", "get-summary", "get-meta", "get-csv", "get-data" or "post-data"

dataset_version

Version of data set to be connected to

preview_token

Preview token required for access to private data sets

ees_environment

EES ees_environment to connect to: "dev", "test", "preprod" or "prod"

api_version

EES API version

parse

Parse result into structured list

verbose

Run with additional contextual messaging. Logical, default = FALSE

Value

Results of query to API meta data endpoint

Examples

eesyapi:::get_meta_response(example_id())
#> $filters
#>      id          column           label hint
#> 1 BT7J3 education_phase Education phase     
#>                                                                options
#> 1 UyHRF, oUXmX, rwhNj, GIxgr, All schools, Primary, Secondary, Special
#> 
#> $indicators
#>      id                     column                      label unit
#> 1 uxo41 persistent_absence_percent Persistent absence percent    %
#>   decimalPlaces
#> 1             1
#> 
#> $geographicLevels
#>   code           label
#> 1   LA Local authority
#> 2  NAT        National
#> 3  REG        Regional
#> 
#> $locations
#>   level.code     level.label
#> 1         LA Local authority
#> 2        NAT        National
#> 3        REG        Regional
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               options
#> 1 rQkNj, GCsgr, u9Mo4, tdEm5, WU2bo, 3se8u, emDuS, 0kT5D, HTzLj, 7YFXo, 9e4vZ, dt0Zw, Poqeb, SdgVx, 3P8uH, fyYFZ, yW5aB, P0ebW, p5PSo, 1rPi3, HvezL, RySka, ENqFR, DcQeg, eFuSW, o47mX, 45UZU, b8tTc, TuNPJ, bqZtT, o49mX, it6Xr, z4FQE, CXVId, YJuHK, XH4fK, 2alzG, hLOyW, Db3Qe, YPHKM, qgvjG, emJuS, alDLP, UhYRF, BYe7J, 5TYdi, 4kdUZ, LxWjE, mU59K, znRFQ, 0evT5, WcR2b, O7CLF, EADqF, TsPJP, kLDhs, Qt6wb, 9Ru4v, Rpka2, mV9KC, 23Llz, PoTeb, u4Ko4, CpId1, NwNDC, dPE0Z, qGJjG, wqtbx, XzUfK, u9Oo4, krhsL, ToMPJ, oU4mX, csC31, a2NLP, AkGK0, IXFBz, 1Ti3Z, 9kk4v, 7zXob, GZgru, 5Tsdi, deu0Z, 5Pxdi, qFjG7, arLPb, NrDCj, juAMt, tj0Em, dPe0Z, Qiwb4, X9fKb, tATEm, iNE6X, fFLYF, ghO9G, TrPJP, SGVxt, lpcBn, O7fCL, pTSoj, 7HJXo, LlMWj, TuxPJ, uxo41, 6DArf, mUq9K, zvUFQ, cZO31, 5fdi9, UfRFO, gIyO9, OBXCL, eNuSW, fzaYF, yVaBm, VN5XE, pdeSo, BfP7J, jOmAM, 0ehT5, IzBzg, 5Zdi9, e9QuS, jgoAM, ToDPJ, IfJBz, mE9KC, HJPzL, cg31S, PEebW, hgyW6, iDq6X, C7SId, l8ScB, AOhGK, LcKWj, 6QGrf, 6jrfe, hWCyW, ugo41, wwbxT, mRj9K, 0uyT5, 9RR4v, YFeHK, 0JT5D, oBmXk, kaNhs, rRNj1, V0D5X, gBaO9, BT7J3, Barking and Dagenham, Barnet, Barnsley, Bath and North East Somerset, Bedford, Bexley, Birmingham, Blackburn with Darwen, Blackpool, Bolton, Bournemouth, Christchurch and Poole, Bracknell Forest, Bradford, Brent, Brighton and Hove, Bristol, City of, Bromley, Buckinghamshire, Bury, Calderdale, Cambridgeshire, Camden, Central Bedfordshire, Cheshire East, Cheshire West and Chester, City of London, Cornwall, County Durham, Coventry, Croydon, Cumberland, Darlington, Derby, Derbyshire, Devon, Doncaster, Dorset, Dudley, Ealing, East Riding of Yorkshire, East Sussex, Enfield, Essex, Gateshead, Gloucestershire, Greenwich, Hackney, Halton, Hammersmith and Fulham, Hampshire, Haringey, Harrow, Hartlepool, Havering, Herefordshire, County of, Hertfordshire, Hillingdon, Hounslow, Isle of Wight, Isles of Scilly, Islington, Kensington and Chelsea, Kent, Kingston upon Hull, City of, Kingston upon Thames, Kirklees, Knowsley, Lambeth, Lancashire, Leeds, Leicester, Leicestershire, Lewisham, Lincolnshire, Liverpool, Luton, Manchester, Medway, Merton, Middlesbrough, Milton Keynes, Newcastle upon Tyne, Newham, Norfolk, North East Lincolnshire, North Lincolnshire, North Northamptonshire, North Somerset, North Tyneside, North Yorkshire, Northumberland, Nottingham, Nottinghamshire, Oldham, Oxfordshire, Peterborough, Plymouth, Portsmouth, Reading, Redbridge, Redcar and Cleveland, Richmond upon Thames, Rochdale, Rotherham, Rutland, Salford, Sandwell, Sefton, Sheffield, Shropshire, Slough, Solihull, Somerset, South Gloucestershire, South Tyneside, Southampton, Southend-on-Sea, Southwark, St. Helens, Staffordshire, Stockport, Stockton-on-Tees, Stoke-on-Trent, Suffolk, Sunderland, Surrey, Sutton, Swindon, Tameside, Telford and Wrekin, Thurrock, Torbay, Tower Hamlets, Trafford, Wakefield, Walsall, Waltham Forest, Wandsworth, Warrington, Warwickshire, West Berkshire, West Northamptonshire, West Sussex, Westminster, Westmorland and Furness, Wigan, Wiltshire, Windsor and Maidenhead, Wirral, Wokingham, Wolverhampton, Worcestershire, York, E09000002, E09000003, E08000016, E06000022, E06000055, E09000004, E08000025, E06000008, E06000009, E08000001, E06000058, E06000036, E08000032, E09000005, E06000043, E06000023, E09000006, E06000060, E08000002, E08000033, E10000003, E09000007, E06000056, E06000049, E06000050, E09000001, E06000052, E06000047, E08000026, E09000008, E06000063, E06000005, E06000015, E10000007, E10000008, E08000017, E06000059, E08000027, E09000009, E06000011, E10000011, E09000010, E10000012, E08000037, E10000013, E09000011, E09000012, E06000006, E09000013, E10000014, E09000014, E09000015, E06000001, E09000016, E06000019, E10000015, E09000017, E09000018, E06000046, E06000053, E09000019, E09000020, E10000016, E06000010, E09000021, E08000034, E08000011, E09000022, E10000017, E08000035, E06000016, E10000018, E09000023, E10000019, E08000012, E06000032, E08000003, E06000035, E09000024, E06000002, E06000042, E08000021, E09000025, E10000020, E06000012, E06000013, E06000061, E06000024, E08000022, E06000065, E06000057, E06000018, E10000024, E08000004, E10000025, E06000031, E06000026, E06000044, E06000038, E09000026, E06000003, E09000027, E08000005, E08000018, E06000017, E08000006, E08000028, E08000014, E08000019, E06000051, E06000039, E08000029, E06000066, E06000025, E08000023, E06000045, E06000033, E09000028, E08000013, E10000028, E08000007, E06000004, E06000021, E10000029, E08000024, E10000030, E09000029, E06000030, E08000008, E06000020, E06000034, E06000027, E09000030, E08000009, E08000036, E08000030, E09000031, E09000032, E06000007, E10000031, E06000037, E06000062, E10000032, E09000033, E06000064, E08000010, E06000054, E06000040, E08000015, E06000041, E08000031, E10000034, E06000014, 301, 302, 370, 800, 822, 303, 330, 889, 890, 350, 839, 867, 380, 304, 846, 801, 305, 825, 351, 381, 873, 202, 823, 895, 896, 201, 908, 840, 331, 306, 942, 841, 831, 830, 878, 371, 838, 332, 307, 811, 845, 308, 881, 390, 916, 203, 204, 876, 205, 850, 309, 310, 805, 311, 884, 919, 312, 313, 921, 420, 206, 207, 886, 810, 314, 382, 340, 208, 888, 383, 856, 855, 209, 925, 341, 821, 352, 887, 315, 806, 826, 391, 316, 926, 812, 813, 940, 802, 392, 815, 929, 892, 891, 353, 931, 874, 879, 851, 870, 317, 807, 318, 354, 372, 857, 355, 333, 343, 373, 893, 871, 334, 933, 803, 393, 852, 882, 210, 342, 860, 356, 808, 861, 935, 394, 936, 319, 866, 357, 894, 883, 880, 211, 358, 384, 335, 320, 212, 877, 937, 869, 941, 938, 213, 943, 359, 865, 868, 344, 872, 336, 885, 816
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           dP0Zw, England, E92000001
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   1Nei3, u4Co4, lZVcB, ACyGK, VTQ5X, UyHRF, oUXmX, deq0Z, Pmjeb, East Midlands, East of England, London, North East, North West, South East, South West, West Midlands, Yorkshire and The Humber, E12000004, E12000006, E12000007, E12000001, E12000002, E12000008, E12000009, E12000005, E12000003
#> 
#> $timePeriods
#>   code period        label
#> 1  W10   2025 2025 Week 10
#> 2  W12   2025 2025 Week 12
#> 3  W14   2025 2025 Week 14
#> 4  W15   2025 2025 Week 15
#> 5  W18   2025 2025 Week 18
#> 6   W8   2025  2025 Week 8
#>