Get publications

Get publications

Usage

get_publications(page_size = 40, page = NULL, verbose = FALSE)

Arguments

page_size

Number of results to collect in a single query to the API (max 40)

page

Page number to return (Default is NULL which will loop through until all pages of the query are collated).

verbose

Add extra contextual information whilst running

Value

Data frame listing all available publications

Examples

get_publications()
#>                                      id
#> 1  b6d9ed96-be68-4791-abc3-08dcaba68c04
#> 2  82c3b19e-1060-4f5e-3410-08dc0b96cfea
#> 3  8b657684-3b15-48d1-153b-08d9fa80c621
#> 4  0427bade-6d68-46bf-eadd-08dcd89c0dc5
#> 5  d823e4df-626f-4450-9b21-08dc8b95fc02
#> 6  2a1a743c-9da5-49e2-fb97-08dce75f2dbb
#> 7  15f4cefb-7d43-4bc0-507e-08dceda89d2d
#> 8  d12cb4ed-6d68-4b2d-50aa-08dceda89d2d
#> 9  3de97aa6-8f78-4e5f-50ce-08dceda89d2d
#> 10 fe1f9c47-715f-451c-50e3-08dceda89d2d
#>                                                             title
#> 1                                             Attendance API data
#> 2                            CSV filename validation test-Updated
#> 3                                         Nick's Test Publication
#> 4                                                    Public API 2
#> 5                        Public API dataset testing-Updated title
#> 6                                                 sv api test new
#> 7  UI tests - Public API - preview token 20241016-061014-u0tsh4-0
#> 8  UI tests - Public API - preview token 20241016-061014-u0tsh4-1
#> 9  UI tests - Public API - preview token 20241016-061014-u0tsh4-2
#> 10 UI tests - Public API - preview token 20241016-061014-u0tsh4-3
#>                                                          slug
#> 1                                         attendance-api-data
#> 2                        csv-filename-validation-test-updated
#> 3                                     nick-s-test-publication
#> 4                                                public-api-2
#> 5                    public-api-dataset-testing-updated-title
#> 6                                             sv-api-test-new
#> 7  ui-tests-public-api-preview-token-20241016-061014-u0tsh4-0
#> 8  ui-tests-public-api-preview-token-20241016-061014-u0tsh4-1
#> 9  ui-tests-public-api-preview-token-20241016-061014-u0tsh4-2
#> 10 ui-tests-public-api-preview-token-20241016-061014-u0tsh4-3
#>                                                                   summary
#> 1                         A place to demo tidy attendance data on the API
#> 2                                                                    TEST
#> 3                                                                        
#> 4                                                                    asdf
#> 5                                                                    test
#> 6                                                                    test
#> 7  UI tests - Public API - preview token 20241016-061014-u0tsh4-0 summary
#> 8  UI tests - Public API - preview token 20241016-061014-u0tsh4-1 summary
#> 9  UI tests - Public API - preview token 20241016-061014-u0tsh4-2 summary
#> 10 UI tests - Public API - preview token 20241016-061014-u0tsh4-3 summary
#>                        lastPublished
#> 1  2024-09-17T11:13:41.8711175+00:00
#> 2  2024-10-14T11:57:51.3142157+00:00
#> 3  2022-11-08T22:47:26.2316722+00:00
#> 4  2024-09-19T15:48:16.1376152+00:00
#> 5  2024-10-09T12:00:43.9538526+00:00
#> 6  2024-10-15T10:42:43.7291296+00:00
#> 7  2024-10-16T06:14:15.1475552+00:00
#> 8  2024-10-16T06:23:32.9953579+00:00
#> 9  2024-10-16T06:32:27.4237216+00:00
#> 10 2024-10-16T06:41:00.8197776+00:00