Get publications

Get publications

Usage

get_publications(
  ees_environment = NULL,
  api_version = NULL,
  page_size = 40,
  page = NULL,
  verbose = FALSE
)

Arguments

ees_environment

EES ees_environment to connect to: "dev", "test", "preprod" or "prod"

api_version

EES API version

page_size

Number of results to return in a single query

page

Page number of query results to return

verbose

Run with additional contextual messaging. Logical, default = FALSE

Value

Data frame listing all available publications

Examples

get_publications()
#>                                     id                           title
#> 1 2c75dd3d-03e8-4f23-eeb2-08dce39219d4          API data sets' Testing
#> 2 25d0e40b-643a-4f73-3ae5-08dcf1c4d57f       Data sets for API testing
#> 3 f715cb51-1f0a-4fa0-0df9-08dce77dad59 Task 2 - second time publishing
#>                            slug
#> 1         api-data-sets-testing
#> 2     data-sets-for-api-testing
#> 3 task-2-second-time-publishing
#>                                                                     summary
#> 1                                        Testing for the first time on Test
#> 2 A collection of example data sets for testing the public facing API with.
#> 3                                           Task 2 - second time publishing
#>                       lastPublished
#> 1 2024-10-15T10:30:30.9450847+00:00
#> 2 2024-10-22T22:34:23.2140687+00:00
#> 3 2024-10-08T10:21:06.9135415+00:00