Get publications

Get publications

Usage

get_publications(
  ees_environment = NULL,
  api_version = NULL,
  search = NULL,
  page_size = 40,
  page = NULL,
  verbose = FALSE
)

Arguments

ees_environment

EES ees_environment to connect to: "dev", "test", "preprod" or "prod"

api_version

EES API version

search

String for filtering the publication list for publication titles and summaries containing the search string provided (strings separated by spaces are combined with OR logic).

page_size

Number of results to return in a single query

page

Page number of query results to return

verbose

Run with additional contextual messaging. Logical, default = FALSE

Value

Data frame listing all available publications

Examples

get_publications()
#>                                     id                       title
#> 1 9676af6b-d563-41f4-d071-08da8f468680 Pupil attendance in schools
#>                          slug
#> 1 pupil-attendance-in-schools
#>                                                                                                                                                       summary
#> 1 Pupil attendance and absence data including termly national statistics and fortnightly statistics in development derived from DfE’s regular attendance data
#>                       lastPublished
#> 1 2025-06-12T08:30:10.9368397+00:00
get_publications(search = "attendance")
#>                                     id                       title
#> 1 9676af6b-d563-41f4-d071-08da8f468680 Pupil attendance in schools
#>                          slug
#> 1 pupil-attendance-in-schools
#>                                                                                                                                                       summary
#> 1 Pupil attendance and absence data including termly national statistics and fortnightly statistics in development derived from DfE’s regular attendance data
#>                       lastPublished
#> 1 2025-06-12T08:30:10.9368397+00:00