Parse contents of API data set json output

This takes the json results output from the API and converts it into a usable data frame. It's primary use is as a helper to get_dataset and post_dataset, but can be used in isolation from those by running an API query and passing the following to parse_api_dataset(): response |> httr::content("text") |> jsonlite::fromJSON() |> parse_api_dataset()

Usage

parse_api_dataset(api_data_result, dataset_id = NULL, verbose = FALSE)

Arguments

api_data_result

A json data result list as returned from the API

dataset_id

ID of data set to be connected to.

verbose

Run in verbose mode, logical, default = FALSE

Value

Data frame containing API data results

Examples

api_url(
  "get-data",
  dataset_id = example_id(), indicators = example_id("indicator"), page_size = 10
) |>
  httr::GET() |>
  httr::content("text") |>
  jsonlite::fromJSON() |>
  parse_api_dataset()
#>    code    period geographic_level    LA   NAT   REG deq0Z VTQ5X Pmjeb 1Nei3
#> 1    AY 2019/2020               LA 5USdi dP0Zw S4HVx zo8FQ bfQtT LrEWj qFujG
#> 2    AY 2019/2020               LA 9ll4v dP0Zw GFDgr kKOhs 9ZZ4v LrEWj qFujG
#> 3    AY 2019/2020               LA uWzo4 dP0Zw GFDgr ucEo4 70eXo LrEWj qFujG
#> 4    AY 2019/2020               LA uWzo4 dP0Zw GFDgr ucEo4 OS5CL LrEWj qFujG
#> 5    AY 2019/2020               LA k8qhs dP0Zw o2nmX TWtPJ 9ZW4v LrEWj qFujG
#> 6    AY 2019/2020               LA uWzo4 dP0Zw GFDgr kKOhs QXGwb LrEWj qFujG
#> 7    AY 2019/2020               LA uWzo4 dP0Zw GFDgr kKOhs EzCqF LrEWj qFujG
#> 8    AY 2019/2020               LA uWzo4 dP0Zw GFDgr kKOhs WNZ2b LrEWj qFujG
#> 9    AY 2019/2020               LA uWzo4 dP0Zw GFDgr kKOhs 0LoT5 LrEWj qFujG
#> 10   AY 2019/2020               LA uWzo4 dP0Zw GFDgr kKOhs m8R9K LrEWj qFujG
#>    mRj9K
#> 1    112
#> 2     97
#> 3  29333
#> 4     14
#> 5     86
#> 6    945
#> 7    196
#> 8     84
#> 9    448
#> 10   163