Parse contents of API data set json output

This takes the json results output from the API and converts it into a usable data frame. It's primary use is as a helper to get_dataset and post_dataset, but can be used in isolation from those by running an API query and passing the following to parse_api_dataset(): response |> httr::content("text") |> jsonlite::fromJSON() |> parse_api_dataset()

Usage

parse_api_dataset(
  api_data_result,
  dataset_id,
  dataset_version = NULL,
  api_version = NULL,
  verbose = FALSE
)

Arguments

api_data_result

A json data result list as returned from the API

dataset_id

ID of data set to be connected to. This is required if the endpoint is one of "get-summary", "get-meta", "get-data" or "post-data"

dataset_version

Version of data set to be connected to

api_version

EES API version

verbose

Run with additional contextual messaging. Logical, default = FALSE

Value

Data frame containing API data results

Examples

example_data_raw(group = "attendance") |>
  parse_api_dataset(example_id(group = "attendance"))
#>    code period geographic_level  la_name   la_code la_oldCode nat_name
#> 1   W26   2024               LA Havering E09000016        311  England
#> 2   W26   2024               LA Havering E09000016        311  England
#> 3   W26   2024               LA Havering E09000016        311  England
#> 4   W26   2024               LA Havering E09000016        311  England
#> 5   W26   2024               LA Havering E09000016        311  England
#> 6   W26   2024               LA Havering E09000016        311  England
#> 7   W26   2024               LA Havering E09000016        311  England
#> 8   W26   2024               LA Havering E09000016        311  England
#> 9   W26   2024               LA Havering E09000016        311  England
#> 10  W26   2024               LA Havering E09000016        311  England
#> 11  W26   2024               LA Havering E09000016        311  England
#> 12  W26   2024               LA Havering E09000016        311  England
#> 13  W26   2024               LA Havering E09000016        311  England
#> 14  W26   2024               LA Havering E09000016        311  England
#> 15  W26   2024               LA Havering E09000016        311  England
#> 16  W26   2024               LA Havering E09000016        311  England
#> 17  W26   2024               LA Havering E09000016        311  England
#> 18  W26   2024               LA Havering E09000016        311  England
#> 19  W26   2024               LA Havering E09000016        311  England
#> 20  W26   2024               LA Havering E09000016        311  England
#> 21  W26   2024               LA Havering E09000016        311  England
#> 22  W26   2024               LA Havering E09000016        311  England
#> 23  W26   2024               LA Havering E09000016        311  England
#> 24  W26   2024               LA Havering E09000016        311  England
#> 25  W26   2024               LA Havering E09000016        311  England
#> 26  W26   2024               LA Havering E09000016        311  England
#> 27  W26   2024               LA Havering E09000016        311  England
#> 28  W26   2024               LA Havering E09000016        311  England
#> 29  W26   2024               LA Havering E09000016        311  England
#> 30  W26   2024               LA Havering E09000016        311  England
#> 31  W26   2024               LA Havering E09000016        311  England
#> 32  W26   2024               LA Havering E09000016        311  England
#>     nat_code reg_name  reg_code attendance_status               attendance_type
#> 1  E92000001   London E12000007        Attendance                       Present
#> 2  E92000001   London E12000007        Attendance Approved educational activity
#> 3  E92000001   London E12000007        Attendance                Not determined
#> 4  E92000001   London E12000007           Absence                Not determined
#> 5  E92000001   London E12000007           Absence                    Authorised
#> 6  E92000001   London E12000007           Absence                  Unauthorised
#> 7  E92000001   London E12000007     Late sessions                Not determined
#> 8  E92000001   London E12000007 Possible sessions                Not determined
#> 9  E92000001   London E12000007        Attendance                       Present
#> 10 E92000001   London E12000007        Attendance                       Present
#> 11 E92000001   London E12000007        Attendance                       Present
#> 12 E92000001   London E12000007        Attendance                       Present
#> 13 E92000001   London E12000007           Absence                    Authorised
#> 14 E92000001   London E12000007           Absence                    Authorised
#> 15 E92000001   London E12000007           Absence                    Authorised
#> 16 E92000001   London E12000007           Absence                    Authorised
#> 17 E92000001   London E12000007           Absence                    Authorised
#> 18 E92000001   London E12000007           Absence                    Authorised
#> 19 E92000001   London E12000007           Absence                    Authorised
#> 20 E92000001   London E12000007           Absence                    Authorised
#> 21 E92000001   London E12000007        Attendance Approved educational activity
#> 22 E92000001   London E12000007        Attendance Approved educational activity
#> 23 E92000001   London E12000007        Attendance Approved educational activity
#> 24 E92000001   London E12000007        Attendance Approved educational activity
#> 25 E92000001   London E12000007        Attendance Approved educational activity
#> 26 E92000001   London E12000007           Absence                  Unauthorised
#> 27 E92000001   London E12000007           Absence                  Unauthorised
#> 28 E92000001   London E12000007           Absence                  Unauthorised
#> 29 E92000001   London E12000007    Not determined                Not determined
#> 30 E92000001   London E12000007    Not determined                Not determined
#> 31 E92000001   London E12000007    Not determined                Not determined
#> 32 E92000001   London E12000007    Not determined                Not determined
#>    day_number establishment_phase                                     reason
#> 1           1             Primary                                      Total
#> 2           1             Primary                                      Total
#> 3           1             Primary                                      Total
#> 4           1             Primary                                      Total
#> 5           1             Primary                                      Total
#> 6           1             Primary                                      Total
#> 7           1             Primary                                      Total
#> 8           1             Primary                                      Total
#> 9           1             Primary                                 Present am
#> 10          1             Primary                                 Present pm
#> 11          1             Primary                                    Present
#> 12          1             Primary     L present late before registers closed
#> 13          1             Primary                       I authorised illness
#> 14          1             Primary                M authorised medical dental
#> 15          1             Primary          R authorised religious observance
#> 16          1             Primary                   S authorised study leave
#> 17          1             Primary                   T authorised grt absence
#> 18          1             Primary                       H authorised holiday
#> 19          1             Primary                      E authorised excluded
#> 20          1             Primary                         C authorised other
#> 21          1             Primary                       B education off site
#> 22          1             Primary                        D dual registration
#> 23          1             Primary               P approved sporting activity
#> 24          1             Primary                   V educational visit trip
#> 25          1             Primary                          W work experience
#> 26          1             Primary                     G unauthorised holiday
#> 27          1             Primary U unauthorised late after registers closed
#> 28          1             Primary                       O other unauthorised
#> 29          1             Primary                            N no reason yet
#> 30          1             Primary               Not attending planned closed
#> 31          1             Primary           Y not attending enforced closure
#> 32          1             Primary                Z not attending not on roll
#>    session_count
#> 1          32342
#> 2            605
#> 3          32947
#> 4           1955
#> 5           1455
#> 6            500
#> 7            404
#> 8          34902
#> 9          15783
#> 10         16180
#> 11         31963
#> 12           379
#> 13          1153
#> 14            84
#> 15             4
#> 16             0
#> 17            14
#> 18            30
#> 19             3
#> 20           167
#> 21           120
#> 22             0
#> 23            16
#> 24           383
#> 25             0
#> 26           241
#> 27            25
#> 28           192
#> 29            42
#> 30             0
#> 31             2
#> 32           124