Parse contents of API data set json output

This takes the json results output from the API and converts it into a usable data frame. It's primary use is as a helper to get_dataset and post_dataset, but can be used in isolation from those by running an API query and passing the following to parse_api_dataset(): response |> httr::content("text") |> jsonlite::fromJSON() |> parse_api_dataset()

Usage

parse_api_dataset(
  api_data_result,
  dataset_id,
  dataset_version = NULL,
  preview_token = NULL,
  ees_environment = NULL,
  api_version = NULL,
  verbose = FALSE
)

Arguments

api_data_result

A json data result list as returned from the API

dataset_id

ID of data set to be connected to. This is required if the endpoint is one of "get-dataset-versions", "get-summary", "get-meta", "get-csv", "get-data" or "post-data"

dataset_version

Version of data set to be connected to

preview_token

Preview token required for access to private data sets

ees_environment

EES ees_environment to connect to: "dev", "test", "preprod" or "prod"

api_version

EES API version

verbose

Run with additional contextual messaging. Logical, default = FALSE

Value

Data frame containing API data results

Examples

example_data_raw(group = "attendance") |>
  eesyapi:::parse_api_dataset(example_id(group = "attendance"))
#>    time_period time_identifier geographic_level nat_name  nat_code
#> 1         2025         Week 14         National  England E92000001
#> 2         2025         Week 14         National  England E92000001
#> 3         2025         Week 14         National  England E92000001
#> 4         2025         Week 14         National  England E92000001
#> 5         2025         Week 14         National  England E92000001
#> 6         2025         Week 14         National  England E92000001
#> 7         2025         Week 14         National  England E92000001
#> 8         2025         Week 14         National  England E92000001
#> 9         2025         Week 14         National  England E92000001
#> 10        2025         Week 14         National  England E92000001
#> 11        2025         Week 14         National  England E92000001
#> 12        2025         Week 14         National  England E92000001
#> 13        2025         Week 14         National  England E92000001
#> 14        2025         Week 14         National  England E92000001
#> 15        2025         Week 14         National  England E92000001
#> 16        2025         Week 14         National  England E92000001
#> 17        2025         Week 14         National  England E92000001
#> 18        2025         Week 14         National  England E92000001
#> 19        2025         Week 14         National  England E92000001
#> 20        2025         Week 14         National  England E92000001
#> 21        2025         Week 14         National  England E92000001
#> 22        2025         Week 14         National  England E92000001
#> 23        2025         Week 14         National  England E92000001
#> 24        2025         Week 14         National  England E92000001
#> 25        2025         Week 14         National  England E92000001
#> 26        2025         Week 14         National  England E92000001
#> 27        2025         Week 14         National  England E92000001
#> 28        2025         Week 14         National  England E92000001
#> 29        2025         Week 14         National  England E92000001
#> 30        2025         Week 14         National  England E92000001
#> 31        2025         Week 14         National  England E92000001
#> 32        2025         Week 14         National  England E92000001
#>                           attendance_reason attendance_status
#> 1                               All present        Attendance
#> 2         All approved educational activity        Attendance
#> 3                        Overall attendance        Attendance
#> 4                           Overall absence           Absence
#> 5                            All authorised           Absence
#> 6                          All unauthorised           Absence
#> 7                         All late sessions     Late sessions
#> 8                     All possible sessions Possible sessions
#> 9                                Present am        Attendance
#> 10                               Present pm        Attendance
#> 11                                  Present        Attendance
#> 12 Present late before registers closed (l)        Attendance
#> 13                              Illness (i)           Absence
#> 14                       Medical dental (m)           Absence
#> 15                 Religious observance (r)           Absence
#> 16                          Study leave (s)           Absence
#> 17                   Authorised holiday (h)           Absence
#> 18                             Excluded (e)           Absence
#> 19                     Other authorised (c)           Absence
#> 20               Regulated performance (c1)           Absence
#> 21                           Interview (j1)           Absence
#> 22         Temporary reduced timetable (c2)           Absence
#> 23                   Education off site (b)        Attendance
#> 24                            Interview (j)        Attendance
#> 25           Approved sporting activity (p)        Attendance
#> 26               Educational visit trip (v)        Attendance
#> 27                      Work experience (w)        Attendance
#> 28             Education arranged by LA (k)        Attendance
#> 29                 Unauthorised holiday (g)           Absence
#> 30          Late after registers closed (u)           Absence
#> 31                   Other unauthorised (o)           Absence
#> 32                        No reason yet (n)           Absence
#>                  attendance_type education_phase   time_frame session_count
#> 1                        Present     All schools Year to date    1619003783
#> 2  Approved educational activity     All schools Year to date      14815860
#> 3             Overall attendance     All schools Year to date    1633819643
#> 4                Overall absence     All schools Year to date     118548447
#> 5                     Authorised     All schools Year to date      80814242
#> 6                   Unauthorised     All schools Year to date      37734205
#> 7              All late sessions     All schools Year to date      37028063
#> 8          All possible sessions     All schools Year to date    1752368090
#> 9                        Present     All schools Year to date     782878151
#> 10                       Present     All schools Year to date     802942904
#> 11                       Present     All schools Year to date    1585821055
#> 12                       Present     All schools Year to date      33182728
#> 13                    Authorised     All schools Year to date      60111966
#> 14                    Authorised     All schools Year to date       5569344
#> 15                    Authorised     All schools Year to date       1211933
#> 16                    Authorised     All schools Year to date         36993
#> 17                    Authorised     All schools Year to date          1715
#> 18                    Authorised     All schools Year to date       2731162
#> 19                    Authorised     All schools Year to date       6872530
#> 20                    Authorised     All schools Year to date         67637
#> 21                    Authorised     All schools Year to date        211774
#> 22                    Authorised     All schools Year to date       3871881
#> 23 Approved educational activity     All schools Year to date       3417207
#> 24 Approved educational activity     All schools Year to date           553
#> 25 Approved educational activity     All schools Year to date       1479929
#> 26 Approved educational activity     All schools Year to date       8624874
#> 27 Approved educational activity     All schools Year to date        668876
#> 28 Approved educational activity     All schools Year to date        624421
#> 29                  Unauthorised     All schools Year to date       6361189
#> 30                  Unauthorised     All schools Year to date       3845335
#> 31                  Unauthorised     All schools Year to date      26318365
#> 32                  Unauthorised     All schools Year to date       1209316