Parse contents of API data set json output

This takes the json results output from the API and converts it into a usable data frame. It's primary use is as a helper to get_dataset and post_dataset, but can be used in isolation from those by running an API query and passing the following to parse_api_dataset(): response |> httr::content("text") |> jsonlite::fromJSON() |> parse_api_dataset()

Usage

parse_api_dataset(
  api_data_result,
  dataset_id,
  dataset_version = NULL,
  preview_token = NULL,
  ees_environment = NULL,
  api_version = NULL,
  verbose = FALSE
)

Arguments

api_data_result

A json data result list as returned from the API

dataset_id

ID of data set to be connected to. This is required if the endpoint is one of "get-dataset-versions", "get-summary", "get-meta", "get-csv", "get-data" or "post-data"

dataset_version

Version of data set to be connected to

preview_token

Preview token required for access to private data sets

ees_environment

EES ees_environment to connect to: "dev", "test", "preprod" or "prod"

api_version

EES API version

verbose

Run with additional contextual messaging. Logical, default = FALSE

Value

Data frame containing API data results

Examples

example_data_raw(group = "attendance") |>
  eesyapi:::parse_api_dataset(example_id(group = "attendance"))
#>    time_period time_identifier geographic_level la_name   la_code la_oldCode
#> 1         2025         Week 18  Local authority  Wirral E08000015        344
#> 2         2025         Week 18  Local authority  Wirral E08000015        344
#> 3         2025         Week 18  Local authority  Wirral E08000015        344
#> 4         2025         Week 18  Local authority  Wirral E08000015        344
#> 5         2025         Week 18  Local authority  Wirral E08000015        344
#> 6         2025         Week 18  Local authority  Wirral E08000015        344
#> 7         2025         Week 18  Local authority  Wirral E08000015        344
#> 8         2025         Week 18  Local authority  Wirral E08000015        344
#> 9         2025         Week 18  Local authority  Wirral E08000015        344
#> 10        2025         Week 18  Local authority  Wirral E08000015        344
#> 11        2025         Week 18  Local authority  Wirral E08000015        344
#> 12        2025         Week 18  Local authority  Wirral E08000015        344
#> 13        2025         Week 18  Local authority  Wirral E08000015        344
#> 14        2025         Week 18  Local authority  Wirral E08000015        344
#> 15        2025         Week 18  Local authority  Wirral E08000015        344
#> 16        2025         Week 18  Local authority  Wirral E08000015        344
#> 17        2025         Week 18  Local authority  Wirral E08000015        344
#> 18        2025         Week 18  Local authority  Wirral E08000015        344
#> 19        2025         Week 18  Local authority  Wirral E08000015        344
#> 20        2025         Week 18  Local authority  Wirral E08000015        344
#> 21        2025         Week 18  Local authority  Wirral E08000015        344
#> 22        2025         Week 18  Local authority  Wirral E08000015        344
#> 23        2025         Week 18  Local authority  Wirral E08000015        344
#> 24        2025         Week 18  Local authority  Wirral E08000015        344
#> 25        2025         Week 18  Local authority  Wirral E08000015        344
#> 26        2025         Week 18  Local authority  Wirral E08000015        344
#> 27        2025         Week 18  Local authority  Wirral E08000015        344
#> 28        2025         Week 18  Local authority  Wirral E08000015        344
#> 29        2025         Week 18  Local authority  Wirral E08000015        344
#> 30        2025         Week 18  Local authority  Wirral E08000015        344
#> 31        2025         Week 18  Local authority  Wirral E08000015        344
#> 32        2025         Week 18  Local authority  Wirral E08000015        344
#>    nat_name  nat_code   reg_name  reg_code
#> 1   England E92000001 North West E12000002
#> 2   England E92000001 North West E12000002
#> 3   England E92000001 North West E12000002
#> 4   England E92000001 North West E12000002
#> 5   England E92000001 North West E12000002
#> 6   England E92000001 North West E12000002
#> 7   England E92000001 North West E12000002
#> 8   England E92000001 North West E12000002
#> 9   England E92000001 North West E12000002
#> 10  England E92000001 North West E12000002
#> 11  England E92000001 North West E12000002
#> 12  England E92000001 North West E12000002
#> 13  England E92000001 North West E12000002
#> 14  England E92000001 North West E12000002
#> 15  England E92000001 North West E12000002
#> 16  England E92000001 North West E12000002
#> 17  England E92000001 North West E12000002
#> 18  England E92000001 North West E12000002
#> 19  England E92000001 North West E12000002
#> 20  England E92000001 North West E12000002
#> 21  England E92000001 North West E12000002
#> 22  England E92000001 North West E12000002
#> 23  England E92000001 North West E12000002
#> 24  England E92000001 North West E12000002
#> 25  England E92000001 North West E12000002
#> 26  England E92000001 North West E12000002
#> 27  England E92000001 North West E12000002
#> 28  England E92000001 North West E12000002
#> 29  England E92000001 North West E12000002
#> 30  England E92000001 North West E12000002
#> 31  England E92000001 North West E12000002
#> 32  England E92000001 North West E12000002
#>                           attendance_reason           attendance_status
#> 1         All approved educational activity                  Attendance
#> 2                        Overall attendance                  Attendance
#> 3                           Overall absence                     Absence
#> 4                            All authorised                     Absence
#> 5                          All unauthorised                     Absence
#> 6                         All late sessions               Late sessions
#> 7                     All possible sessions           Possible sessions
#> 8                                Present am                  Attendance
#> 9                                Present pm                  Attendance
#> 10                                  Present                  Attendance
#> 11 Present late before registers closed (l)                  Attendance
#> 12                              Illness (i)                     Absence
#> 13                       Medical dental (m)                     Absence
#> 14                 Religious observance (r)                     Absence
#> 15                          Study leave (s)                     Absence
#> 16                   Authorised holiday (h)                     Absence
#> 17                             Excluded (e)                     Absence
#> 18                     Other authorised (c)                     Absence
#> 19               Regulated performance (c1)                     Absence
#> 20                           Interview (j1)                     Absence
#> 21         Temporary reduced timetable (c2)                     Absence
#> 22                   Education off site (b)                  Attendance
#> 23                            Interview (j)                  Attendance
#> 24           Approved sporting activity (p)                  Attendance
#> 25               Educational visit trip (v)                  Attendance
#> 26                      Work experience (w)                  Attendance
#> 27             Education arranged by LA (k)                  Attendance
#> 28                 Unauthorised holiday (g)                     Absence
#> 29          Late after registers closed (u)                     Absence
#> 30                   Other unauthorised (o)                     Absence
#> 31                        No reason yet (n)                     Absence
#> 32                    Dual registration (d) Management and legacy codes
#>                  attendance_type education_phase time_frame session_count
#> 1  Approved educational activity       Secondary   Thursday           529
#> 2             Overall attendance       Secondary   Thursday         34756
#> 3                Overall absence       Secondary   Thursday          3419
#> 4                     Authorised       Secondary   Thursday          1836
#> 5                   Unauthorised       Secondary   Thursday          1583
#> 6              All late sessions       Secondary   Thursday          1505
#> 7          All possible sessions       Secondary   Thursday         38175
#> 8                        Present       Secondary   Thursday         16057
#> 9                        Present       Secondary   Thursday         16814
#> 10                       Present       Secondary   Thursday         32871
#> 11                       Present       Secondary   Thursday          1356
#> 12                    Authorised       Secondary   Thursday          1032
#> 13                    Authorised       Secondary   Thursday           184
#> 14                    Authorised       Secondary   Thursday             0
#> 15                    Authorised       Secondary   Thursday            14
#> 16                    Authorised       Secondary   Thursday             0
#> 17                    Authorised       Secondary   Thursday            79
#> 18                    Authorised       Secondary   Thursday           313
#> 19                    Authorised       Secondary   Thursday             0
#> 20                    Authorised       Secondary   Thursday             2
#> 21                    Authorised       Secondary   Thursday           212
#> 22 Approved educational activity       Secondary   Thursday           324
#> 23 Approved educational activity       Secondary   Thursday             0
#> 24 Approved educational activity       Secondary   Thursday           108
#> 25 Approved educational activity       Secondary   Thursday            75
#> 26 Approved educational activity       Secondary   Thursday             8
#> 27 Approved educational activity       Secondary   Thursday            14
#> 28                  Unauthorised       Secondary   Thursday           120
#> 29                  Unauthorised       Secondary   Thursday           149
#> 30                  Unauthorised       Secondary   Thursday          1244
#> 31                  Unauthorised       Secondary   Thursday            70
#> 32   Management and legacy codes       Secondary   Thursday           253