Parse contents of API data set json output

This takes the json results output from the API and converts it into a usable data frame. It's primary use is as a helper to get_dataset and post_dataset, but can be used in isolation from those by running an API query and passing the following to parse_api_dataset(): response |> httr::content("text") |> jsonlite::fromJSON() |> parse_api_dataset()

Usage

parse_api_dataset(
  api_data_result,
  dataset_id,
  dataset_version = NULL,
  ees_environment = NULL,
  api_version = NULL,
  verbose = FALSE
)

Arguments

api_data_result

A json data result list as returned from the API

dataset_id

ID of data set to be connected to. This is required if the endpoint is one of "get-dataset-versions", "get-summary", "get-meta", "get-csv", "get-data" or "post-data"

dataset_version

Version of data set to be connected to

ees_environment

EES ees_environment to connect to: "dev", "test", "preprod" or "prod"

api_version

EES API version

verbose

Run with additional contextual messaging. Logical, default = FALSE

Value

Data frame containing API data results

Examples

example_data_raw(group = "attendance") |>
  eesyapi:::parse_api_dataset(example_id(group = "attendance"))
#>    time_period time_identifier geographic_level    la_name   la_code la_oldCode
#> 1         2025          Week 8  Local authority Hartlepool E06000001        805
#> 2         2025          Week 8  Local authority Hartlepool E06000001        805
#> 3         2025          Week 8  Local authority Hartlepool E06000001        805
#> 4         2025          Week 8  Local authority Hartlepool E06000001        805
#> 5         2025          Week 8  Local authority Hartlepool E06000001        805
#> 6         2025          Week 8  Local authority Hartlepool E06000001        805
#> 7         2025          Week 8  Local authority Hartlepool E06000001        805
#> 8         2025          Week 8  Local authority Hartlepool E06000001        805
#> 9         2025          Week 8  Local authority Hartlepool E06000001        805
#> 10        2025          Week 8  Local authority Hartlepool E06000001        805
#> 11        2025          Week 8  Local authority Hartlepool E06000001        805
#> 12        2025          Week 8  Local authority Hartlepool E06000001        805
#> 13        2025          Week 8  Local authority Hartlepool E06000001        805
#> 14        2025          Week 8  Local authority Hartlepool E06000001        805
#> 15        2025          Week 8  Local authority Hartlepool E06000001        805
#> 16        2025          Week 8  Local authority Hartlepool E06000001        805
#> 17        2025          Week 8  Local authority Hartlepool E06000001        805
#> 18        2025          Week 8  Local authority Hartlepool E06000001        805
#> 19        2025          Week 8  Local authority Hartlepool E06000001        805
#> 20        2025          Week 8  Local authority Hartlepool E06000001        805
#> 21        2025          Week 8  Local authority Hartlepool E06000001        805
#> 22        2025          Week 8  Local authority Hartlepool E06000001        805
#> 23        2025          Week 8  Local authority Hartlepool E06000001        805
#> 24        2025          Week 8  Local authority Hartlepool E06000001        805
#> 25        2025          Week 8  Local authority Hartlepool E06000001        805
#> 26        2025          Week 8  Local authority Hartlepool E06000001        805
#> 27        2025          Week 8  Local authority Hartlepool E06000001        805
#> 28        2025          Week 8  Local authority Hartlepool E06000001        805
#> 29        2025          Week 8  Local authority Hartlepool E06000001        805
#> 30        2025          Week 8  Local authority Hartlepool E06000001        805
#> 31        2025          Week 8  Local authority Hartlepool E06000001        805
#> 32        2025          Week 8  Local authority Hartlepool E06000001        805
#>    nat_name  nat_code   reg_name  reg_code
#> 1   England E92000001 North East E12000001
#> 2   England E92000001 North East E12000001
#> 3   England E92000001 North East E12000001
#> 4   England E92000001 North East E12000001
#> 5   England E92000001 North East E12000001
#> 6   England E92000001 North East E12000001
#> 7   England E92000001 North East E12000001
#> 8   England E92000001 North East E12000001
#> 9   England E92000001 North East E12000001
#> 10  England E92000001 North East E12000001
#> 11  England E92000001 North East E12000001
#> 12  England E92000001 North East E12000001
#> 13  England E92000001 North East E12000001
#> 14  England E92000001 North East E12000001
#> 15  England E92000001 North East E12000001
#> 16  England E92000001 North East E12000001
#> 17  England E92000001 North East E12000001
#> 18  England E92000001 North East E12000001
#> 19  England E92000001 North East E12000001
#> 20  England E92000001 North East E12000001
#> 21  England E92000001 North East E12000001
#> 22  England E92000001 North East E12000001
#> 23  England E92000001 North East E12000001
#> 24  England E92000001 North East E12000001
#> 25  England E92000001 North East E12000001
#> 26  England E92000001 North East E12000001
#> 27  England E92000001 North East E12000001
#> 28  England E92000001 North East E12000001
#> 29  England E92000001 North East E12000001
#> 30  England E92000001 North East E12000001
#> 31  England E92000001 North East E12000001
#> 32  England E92000001 North East E12000001
#>                           attendance_reason attendance_status
#> 1                               All present        Attendance
#> 2         All approved educational activity        Attendance
#> 3                        Overall attendance        Attendance
#> 4                           Overall absence           Absence
#> 5                            All authorised           Absence
#> 6                          All unauthorised           Absence
#> 7                         All late sessions     Late sessions
#> 8                     All possible sessions Possible sessions
#> 9                                Present am        Attendance
#> 10                               Present pm        Attendance
#> 11                                  Present        Attendance
#> 12 Present late before registers closed (l)        Attendance
#> 13                              Illness (i)           Absence
#> 14                       Medical dental (m)           Absence
#> 15                 Religious observance (r)           Absence
#> 16                          Study leave (s)           Absence
#> 17                   Authorised holiday (h)           Absence
#> 18                             Excluded (e)           Absence
#> 19                     Other authorised (c)           Absence
#> 20               Regulated performance (c1)           Absence
#> 21                           Interview (j1)           Absence
#> 22         Temporary reduced timetable (c2)           Absence
#> 23                   Education off site (b)        Attendance
#> 24                            Interview (j)        Attendance
#> 25           Approved sporting activity (p)        Attendance
#> 26               Educational visit trip (v)        Attendance
#> 27                      Work experience (w)        Attendance
#> 28             Education arranged by LA (k)        Attendance
#> 29                 Unauthorised holiday (g)           Absence
#> 30          Late after registers closed (u)           Absence
#> 31                   Other unauthorised (o)           Absence
#> 32                        No reason yet (n)           Absence
#>                  attendance_type education_phase   time_frame session_count
#> 1                        Present         Primary Year to date       1273300
#> 2  Approved educational activity         Primary Year to date          6642
#> 3             Overall attendance         Primary Year to date       1279942
#> 4                Overall absence         Primary Year to date         70857
#> 5                     Authorised         Primary Year to date         49494
#> 6                   Unauthorised         Primary Year to date         21363
#> 7              All late sessions         Primary Year to date         17533
#> 8          All possible sessions         Primary Year to date       1350799
#> 9                        Present         Primary Year to date        621998
#> 10                       Present         Primary Year to date        635842
#> 11                       Present         Primary Year to date       1257840
#> 12                       Present         Primary Year to date         15460
#> 13                    Authorised         Primary Year to date         42823
#> 14                    Authorised         Primary Year to date          2884
#> 15                    Authorised         Primary Year to date             7
#> 16                    Authorised         Primary Year to date             0
#> 17                    Authorised         Primary Year to date             0
#> 18                    Authorised         Primary Year to date            84
#> 19                    Authorised         Primary Year to date          3302
#> 20                    Authorised         Primary Year to date           115
#> 21                    Authorised         Primary Year to date            33
#> 22                    Authorised         Primary Year to date           245
#> 23 Approved educational activity         Primary Year to date           843
#> 24 Approved educational activity         Primary Year to date             0
#> 25 Approved educational activity         Primary Year to date           734
#> 26 Approved educational activity         Primary Year to date          5006
#> 27 Approved educational activity         Primary Year to date             0
#> 28 Approved educational activity         Primary Year to date            59
#> 29                  Unauthorised         Primary Year to date          8110
#> 30                  Unauthorised         Primary Year to date          2073
#> 31                  Unauthorised         Primary Year to date         10283
#> 32                  Unauthorised         Primary Year to date           897