Parse IDs in a set of filters

The API uses unique IDs (sqids) to identify each filter column and its contents (filter items). This function parses those into the data creators' id and item labels based on the meta data stored on the API for the data set.

Usage

parse_sqids_filters(filters, meta, verbose = FALSE)

Arguments

filters

A set of filter item columns as taken from a data set downloaded from the API

meta

Meta data for the data set as provided by get_meta()

verbose

Run in verbose mode with debugging messages

Value

Data frame

Examples

example_data_raw() |>
  magrittr::use_series("filters") |>
  eesyapi:::parse_sqids_filters(get_meta(example_id(group = "attendance")))
#>                           attendance_reason attendance_status
#> 1                               All present        Attendance
#> 2         All approved educational activity        Attendance
#> 3                        Overall attendance        Attendance
#> 4                           Overall absence           Absence
#> 5                            All authorised           Absence
#> 6                          All unauthorised           Absence
#> 7                         All late sessions     Late sessions
#> 8                     All possible sessions Possible sessions
#> 9                                Present am        Attendance
#> 10                               Present pm        Attendance
#> 11                                  Present        Attendance
#> 12 Present late before registers closed (l)        Attendance
#> 13                              Illness (i)           Absence
#> 14                       Medical dental (m)           Absence
#> 15                 Religious observance (r)           Absence
#> 16                          Study leave (s)           Absence
#> 17                   Authorised holiday (h)           Absence
#> 18                             Excluded (e)           Absence
#> 19                     Other authorised (c)           Absence
#> 20               Regulated performance (c1)           Absence
#> 21                           Interview (j1)           Absence
#> 22         Temporary reduced timetable (c2)           Absence
#> 23                   Education off site (b)        Attendance
#> 24                            Interview (j)        Attendance
#> 25           Approved sporting activity (p)        Attendance
#> 26               Educational visit trip (v)        Attendance
#> 27                      Work experience (w)        Attendance
#> 28             Education arranged by LA (k)        Attendance
#> 29                 Unauthorised holiday (g)           Absence
#> 30          Late after registers closed (u)           Absence
#> 31                   Other unauthorised (o)           Absence
#> 32                        No reason yet (n)           Absence
#>                  attendance_type education_phase   time_frame
#> 1                        Present     All schools Year to date
#> 2  Approved educational activity     All schools Year to date
#> 3             Overall attendance     All schools Year to date
#> 4                Overall absence     All schools Year to date
#> 5                     Authorised     All schools Year to date
#> 6                   Unauthorised     All schools Year to date
#> 7              All late sessions     All schools Year to date
#> 8          All possible sessions     All schools Year to date
#> 9                        Present     All schools Year to date
#> 10                       Present     All schools Year to date
#> 11                       Present     All schools Year to date
#> 12                       Present     All schools Year to date
#> 13                    Authorised     All schools Year to date
#> 14                    Authorised     All schools Year to date
#> 15                    Authorised     All schools Year to date
#> 16                    Authorised     All schools Year to date
#> 17                    Authorised     All schools Year to date
#> 18                    Authorised     All schools Year to date
#> 19                    Authorised     All schools Year to date
#> 20                    Authorised     All schools Year to date
#> 21                    Authorised     All schools Year to date
#> 22                    Authorised     All schools Year to date
#> 23 Approved educational activity     All schools Year to date
#> 24 Approved educational activity     All schools Year to date
#> 25 Approved educational activity     All schools Year to date
#> 26 Approved educational activity     All schools Year to date
#> 27 Approved educational activity     All schools Year to date
#> 28 Approved educational activity     All schools Year to date
#> 29                  Unauthorised     All schools Year to date
#> 30                  Unauthorised     All schools Year to date
#> 31                  Unauthorised     All schools Year to date
#> 32                  Unauthorised     All schools Year to date