Query a data set using POST and a query json

This function provides a method for generating and sending a json based data query to the EES API. As a minimum, it requires the dataset_id flag and either the indicators flag or a json file containing a query to be provided.

Usage

post_dataset(
  dataset_id,
  indicators = NULL,
  time_periods = NULL,
  geographies = NULL,
  filter_items = NULL,
  json_query = NULL,
  dataset_version = NULL,
  api_version = NULL,
  page = NULL,
  page_size = 1000,
  parse = TRUE,
  debug = FALSE,
  verbose = FALSE
)

Arguments

dataset_id

ID of data set to be connected to. This is required if the endpoint is one of "get-summary", "get-meta", "get-data" or "post-data"

indicators

Indicators required as a string or vector of strings (required)

time_periods

Time periods required as a string ("period|code") or vector of strings

geographies

String, vector or data frame containing the geographic levels and locations to be queried.

filter_items

Filter items required as a string or vector of strings

json_query

Optional path to a json file containing the query parameters

dataset_version

Version of data set to be connected to

api_version

EES API version

page

Page number of query results to return

page_size

Number of results to return in a single query

parse

Logical flag to activate parsing of the results. Default: TRUE

debug

Run POST query in debug mode: logic, default: FALSE

verbose

Run with additional contextual messaging, logical, default = FALSE

Value

Data frame containing query results of an API data set

Examples

post_dataset(
  example_id(group = "attendance"),
  json_query = example_json_query()
)
#> Parsing query options
#>   code period geographic_level   NAT   REG Db3Qe emJuS 4kdUZ 5TYdi mU59K
#> 1  W23   2024              REG dP0Zw rg3Nj pmRSo CvuId ThDPJ uLQo4 bBrtT
#> 2  W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf ThDPJ uLQo4 bBrtT
#> 3  W23   2024              REG dP0Zw rg3Nj pmRSo CvuId crH31 uLQo4 bBrtT
#> 4  W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf crH31 uLQo4 bBrtT
#> 5  W23   2024              NAT dP0Zw  <NA> pmRSo CvuId crH31 uLQo4 bBrtT
#> 6  W23   2024              NAT dP0Zw  <NA> pmRSo 6AXrf crH31 uLQo4 bBrtT
#> 7  W23   2024              NAT dP0Zw  <NA> pmRSo CvuId ThDPJ uLQo4 bBrtT
#> 8  W23   2024              NAT dP0Zw  <NA> pmRSo 6AXrf ThDPJ uLQo4 bBrtT
#>        bqZtT
#> 1    1862256
#> 2      35007
#> 3      74898
#> 4       1020
#> 5     870024
#> 6      17506
#> 7   20145467
#> 8     401209

# Run post_dataset() to select rows containing either of two geographic locations and either of
# two filter items.
post_dataset(
  example_id(group = "attendance"),
  indicators = example_id("indicator", group = "attendance"),
  time_periods = "2024|W23",
  geographies = c("NAT|id|dP0Zw", "REG|id|rg3Nj"),
  filter_items = c("CvuId", "6AXrf"),
  page = 1,
  page_size = 32
)
#>    code period geographic_level   NAT   REG Db3Qe emJuS 4kdUZ 5TYdi mU59K
#> 1   W23   2024              REG dP0Zw rg3Nj pmRSo CvuId XVWfK qf0jG bBrtT
#> 2   W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf XVWfK qf0jG bBrtT
#> 3   W23   2024              REG dP0Zw rg3Nj pmRSo CvuId XVWfK qf0jG GzNgr
#> 4   W23   2024              REG dP0Zw rg3Nj pmRSo CvuId XVWfK qf0jG 3Nk8u
#> 5   W23   2024              REG dP0Zw rg3Nj pmRSo CvuId XVWfK qf0jG rGcNj
#> 6   W23   2024              REG dP0Zw rg3Nj pmRSo CvuId XVWfK qf0jG PyBeb
#> 7   W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf XVWfK qf0jG t74Em
#> 8   W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf XVWfK qf0jG jYyAM
#> 9   W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf XVWfK qf0jG oWvmX
#> 10  W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf XVWfK qf0jG ea0uS
#> 11  W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf XVWfK qf0jG 5USdi
#> 12  W23   2024              REG dP0Zw rg3Nj pmRSo CvuId crH31 qf0jG bBrtT
#> 13  W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf crH31 qf0jG bBrtT
#> 14  W23   2024              REG dP0Zw rg3Nj pmRSo CvuId crH31 qf0jG GzNgr
#> 15  W23   2024              REG dP0Zw rg3Nj pmRSo CvuId crH31 qf0jG 3Nk8u
#> 16  W23   2024              REG dP0Zw rg3Nj pmRSo CvuId crH31 qf0jG rGcNj
#> 17  W23   2024              REG dP0Zw rg3Nj pmRSo CvuId crH31 qf0jG PyBeb
#> 18  W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf crH31 qf0jG t74Em
#> 19  W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf crH31 qf0jG jYyAM
#> 20  W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf crH31 qf0jG oWvmX
#> 21  W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf crH31 qf0jG ea0uS
#> 22  W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf crH31 qf0jG 5USdi
#> 23  W23   2024              NAT dP0Zw  <NA> pmRSo CvuId crH31 qf0jG bBrtT
#> 24  W23   2024              NAT dP0Zw  <NA> pmRSo 6AXrf crH31 qf0jG bBrtT
#> 25  W23   2024              NAT dP0Zw  <NA> pmRSo CvuId crH31 qf0jG GzNgr
#> 26  W23   2024              NAT dP0Zw  <NA> pmRSo CvuId crH31 qf0jG 3Nk8u
#> 27  W23   2024              NAT dP0Zw  <NA> pmRSo CvuId crH31 qf0jG rGcNj
#> 28  W23   2024              NAT dP0Zw  <NA> pmRSo CvuId crH31 qf0jG PyBeb
#> 29  W23   2024              NAT dP0Zw  <NA> pmRSo 6AXrf crH31 qf0jG t74Em
#> 30  W23   2024              NAT dP0Zw  <NA> pmRSo 6AXrf crH31 qf0jG jYyAM
#> 31  W23   2024              NAT dP0Zw  <NA> pmRSo 6AXrf crH31 qf0jG oWvmX
#> 32  W23   2024              NAT dP0Zw  <NA> pmRSo 6AXrf crH31 qf0jG ea0uS
#>         bqZtT
#> 1      475991
#> 2        5235
#> 3      234417
#> 4      236991
#> 5      471408
#> 6        4583
#> 7         452
#> 8         226
#> 9         552
#> 10       4225
#> 11          0
#> 12      14497
#> 13        130
#> 14       7206
#> 15       7203
#> 16      14409
#> 17         88
#> 18        103
#> 19         26
#> 20          1
#> 21         24
#> 22          0
#> 23     157670
#> 24       2611
#> 25      77609
#> 26      78597
#> 27     156206
#> 28       1464
#> 29       2132
#> 30        287
#> 31         46
#> 32        343

# Run post_dataset() using set parameters giving a combination of filter options
post_dataset(
  example_id(group = "attendance"),
  indicators = example_id("indicator", group = "attendance"),
  time_periods = "2024|W23",
  geographies = c("NAT|id|dP0Zw", "REG|id|rg3Nj"),
  filter_items = list(
    attendance_status = c("pmRSo", "7SdXo"),
    attendance_type = c("CvuId", "6AXrf", "0k3T5", "YdkHK"),
    education_phase = c("ThDPJ", "crH31"),
    day_number = c("uLQo4"),
    reason = c("bBrtT")
  )
)
#>    code period geographic_level   NAT   REG 5TYdi mU59K Db3Qe emJuS 4kdUZ
#> 1   W23   2024              REG dP0Zw rg3Nj uLQo4 bBrtT pmRSo CvuId ThDPJ
#> 2   W23   2024              REG dP0Zw rg3Nj uLQo4 bBrtT pmRSo 6AXrf ThDPJ
#> 3   W23   2024              REG dP0Zw rg3Nj uLQo4 bBrtT 7SdXo 0k3T5 ThDPJ
#> 4   W23   2024              REG dP0Zw rg3Nj uLQo4 bBrtT 7SdXo YdkHK ThDPJ
#> 5   W23   2024              REG dP0Zw rg3Nj uLQo4 bBrtT pmRSo CvuId crH31
#> 6   W23   2024              REG dP0Zw rg3Nj uLQo4 bBrtT pmRSo 6AXrf crH31
#> 7   W23   2024              REG dP0Zw rg3Nj uLQo4 bBrtT 7SdXo 0k3T5 crH31
#> 8   W23   2024              REG dP0Zw rg3Nj uLQo4 bBrtT 7SdXo YdkHK crH31
#> 9   W23   2024              NAT dP0Zw  <NA> uLQo4 bBrtT pmRSo CvuId crH31
#> 10  W23   2024              NAT dP0Zw  <NA> uLQo4 bBrtT pmRSo 6AXrf crH31
#> 11  W23   2024              NAT dP0Zw  <NA> uLQo4 bBrtT 7SdXo 0k3T5 crH31
#> 12  W23   2024              NAT dP0Zw  <NA> uLQo4 bBrtT 7SdXo YdkHK crH31
#> 13  W23   2024              NAT dP0Zw  <NA> uLQo4 bBrtT pmRSo CvuId ThDPJ
#> 14  W23   2024              NAT dP0Zw  <NA> uLQo4 bBrtT pmRSo 6AXrf ThDPJ
#> 15  W23   2024              NAT dP0Zw  <NA> uLQo4 bBrtT 7SdXo 0k3T5 ThDPJ
#> 16  W23   2024              NAT dP0Zw  <NA> uLQo4 bBrtT 7SdXo YdkHK ThDPJ
#>         bqZtT
#> 1     1862256
#> 2       35007
#> 3      105543
#> 4       83756
#> 5       74898
#> 6        1020
#> 7        7905
#> 8        2871
#> 9      870024
#> 10      17506
#> 11      88094
#> 12      37748
#> 13   20145467
#> 14     401209
#> 15    1093633
#> 16     956076