Query a data set using POST and a query json

This function provides a method for generating and sending a json based data query to the EES API. As a minimum, it requires the dataset_id flag and either the indicators flag or a json file containing a query to be provided.

Usage

post_dataset(
  dataset_id,
  indicators = NULL,
  time_periods = NULL,
  geographies = NULL,
  filter_items = NULL,
  json_query = NULL,
  dataset_version = NULL,
  api_version = NULL,
  page = NULL,
  page_size = 1000,
  parse = TRUE,
  debug = FALSE,
  verbose = FALSE
)

Arguments

dataset_id

ID of data set to be connected to. This is required if the endpoint is one of "get-summary", "get-meta", "get-data" or "post-data"

indicators

Indicators required as a string or vector of strings (required)

time_periods

Time periods required as a string ("period|code") or vector of strings

geographies

String, vector or data frame containing the geographic levels and locations to be queried.

filter_items

Filter items required as a string or vector of strings

json_query

Optional path to a json file containing the query parameters

dataset_version

Version of data set to be connected to

api_version

EES API version

page

Page number of query results to return

page_size

Number of results to return in a single query

parse

Logical flag to activate parsing of the results. Default: TRUE

debug

Run POST query in debug mode. Logical, default = FALSE

verbose

Run with additional contextual messaging. Logical, default = FALSE

Value

Data frame containing query results of an API data set

Examples

post_dataset(
  example_id(group = "attendance"),
  json_query = example_json_query()
)
#>   code period geographic_level nat_name  nat_code      reg_name  reg_code
#> 1  W23   2024              REG  England E92000001 East Midlands E12000004
#> 2  W23   2024              REG  England E92000001 East Midlands E12000004
#> 3  W23   2024              REG  England E92000001 East Midlands E12000004
#> 4  W23   2024              REG  England E92000001 East Midlands E12000004
#> 5  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 6  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 7  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 8  W23   2024              NAT  England E92000001          <NA>      <NA>
#>   attendance_status               attendance_type day_number
#> 1        Attendance                       Present      Total
#> 2        Attendance Approved educational activity      Total
#> 3        Attendance                       Present      Total
#> 4        Attendance Approved educational activity      Total
#> 5        Attendance                       Present      Total
#> 6        Attendance Approved educational activity      Total
#> 7        Attendance                       Present      Total
#> 8        Attendance Approved educational activity      Total
#>   establishment_phase reason session_count
#> 1           Secondary  Total       1862256
#> 2           Secondary  Total         35007
#> 3             Special  Total         74898
#> 4             Special  Total          1020
#> 5             Special  Total        870024
#> 6             Special  Total         17506
#> 7           Secondary  Total      20145467
#> 8           Secondary  Total        401209

# Run post_dataset() to select rows containing either of two geographic locations and either of
# two filter items.
post_dataset(
  example_id(group = "attendance"),
  indicators = example_id("indicator", group = "attendance"),
  time_periods = "2024|W23",
  geographies = c("NAT|id|dP0Zw", "REG|id|rg3Nj"),
  filter_items = c("CvuId", "6AXrf"),
  page = 1,
  page_size = 32
)
#>    code period geographic_level nat_name  nat_code      reg_name  reg_code
#> 1   W23   2024              REG  England E92000001 East Midlands E12000004
#> 2   W23   2024              REG  England E92000001 East Midlands E12000004
#> 3   W23   2024              REG  England E92000001 East Midlands E12000004
#> 4   W23   2024              REG  England E92000001 East Midlands E12000004
#> 5   W23   2024              REG  England E92000001 East Midlands E12000004
#> 6   W23   2024              REG  England E92000001 East Midlands E12000004
#> 7   W23   2024              REG  England E92000001 East Midlands E12000004
#> 8   W23   2024              REG  England E92000001 East Midlands E12000004
#> 9   W23   2024              REG  England E92000001 East Midlands E12000004
#> 10  W23   2024              REG  England E92000001 East Midlands E12000004
#> 11  W23   2024              REG  England E92000001 East Midlands E12000004
#> 12  W23   2024              REG  England E92000001 East Midlands E12000004
#> 13  W23   2024              REG  England E92000001 East Midlands E12000004
#> 14  W23   2024              REG  England E92000001 East Midlands E12000004
#> 15  W23   2024              REG  England E92000001 East Midlands E12000004
#> 16  W23   2024              REG  England E92000001 East Midlands E12000004
#> 17  W23   2024              REG  England E92000001 East Midlands E12000004
#> 18  W23   2024              REG  England E92000001 East Midlands E12000004
#> 19  W23   2024              REG  England E92000001 East Midlands E12000004
#> 20  W23   2024              REG  England E92000001 East Midlands E12000004
#> 21  W23   2024              REG  England E92000001 East Midlands E12000004
#> 22  W23   2024              REG  England E92000001 East Midlands E12000004
#> 23  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 24  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 25  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 26  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 27  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 28  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 29  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 30  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 31  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 32  W23   2024              NAT  England E92000001          <NA>      <NA>
#>    attendance_status               attendance_type day_number
#> 1         Attendance                       Present          1
#> 2         Attendance Approved educational activity          1
#> 3         Attendance                       Present          1
#> 4         Attendance                       Present          1
#> 5         Attendance                       Present          1
#> 6         Attendance                       Present          1
#> 7         Attendance Approved educational activity          1
#> 8         Attendance Approved educational activity          1
#> 9         Attendance Approved educational activity          1
#> 10        Attendance Approved educational activity          1
#> 11        Attendance Approved educational activity          1
#> 12        Attendance                       Present          1
#> 13        Attendance Approved educational activity          1
#> 14        Attendance                       Present          1
#> 15        Attendance                       Present          1
#> 16        Attendance                       Present          1
#> 17        Attendance                       Present          1
#> 18        Attendance Approved educational activity          1
#> 19        Attendance Approved educational activity          1
#> 20        Attendance Approved educational activity          1
#> 21        Attendance Approved educational activity          1
#> 22        Attendance Approved educational activity          1
#> 23        Attendance                       Present          1
#> 24        Attendance Approved educational activity          1
#> 25        Attendance                       Present          1
#> 26        Attendance                       Present          1
#> 27        Attendance                       Present          1
#> 28        Attendance                       Present          1
#> 29        Attendance Approved educational activity          1
#> 30        Attendance Approved educational activity          1
#> 31        Attendance Approved educational activity          1
#> 32        Attendance Approved educational activity          1
#>    establishment_phase                                 reason session_count
#> 1              Primary                                  Total        475991
#> 2              Primary                                  Total          5235
#> 3              Primary                             Present am        234417
#> 4              Primary                             Present pm        236991
#> 5              Primary                                Present        471408
#> 6              Primary L present late before registers closed          4583
#> 7              Primary                   B education off site           452
#> 8              Primary                    D dual registration           226
#> 9              Primary           P approved sporting activity           552
#> 10             Primary               V educational visit trip          4225
#> 11             Primary                      W work experience             0
#> 12             Special                                  Total         14497
#> 13             Special                                  Total           130
#> 14             Special                             Present am          7206
#> 15             Special                             Present pm          7203
#> 16             Special                                Present         14409
#> 17             Special L present late before registers closed            88
#> 18             Special                   B education off site           103
#> 19             Special                    D dual registration            26
#> 20             Special           P approved sporting activity             1
#> 21             Special               V educational visit trip            24
#> 22             Special                      W work experience             0
#> 23             Special                                  Total        157670
#> 24             Special                                  Total          2611
#> 25             Special                             Present am         77609
#> 26             Special                             Present pm         78597
#> 27             Special                                Present        156206
#> 28             Special L present late before registers closed          1464
#> 29             Special                   B education off site          2132
#> 30             Special                    D dual registration           287
#> 31             Special           P approved sporting activity            46
#> 32             Special               V educational visit trip           343

# Run post_dataset() using set parameters giving a combination of filter options
post_dataset(
  example_id(group = "attendance"),
  indicators = example_id("indicator", group = "attendance"),
  time_periods = "2024|W23",
  geographies = c("NAT|id|dP0Zw", "REG|id|rg3Nj"),
  filter_items = list(
    attendance_status = c("pmRSo", "7SdXo"),
    attendance_type = c("CvuId", "6AXrf", "0k3T5", "YdkHK"),
    education_phase = c("ThDPJ", "crH31"),
    day_number = c("uLQo4"),
    reason = c("bBrtT")
  )
)
#>    code period geographic_level nat_name  nat_code      reg_name  reg_code
#> 1   W23   2024              REG  England E92000001 East Midlands E12000004
#> 2   W23   2024              REG  England E92000001 East Midlands E12000004
#> 3   W23   2024              REG  England E92000001 East Midlands E12000004
#> 4   W23   2024              REG  England E92000001 East Midlands E12000004
#> 5   W23   2024              REG  England E92000001 East Midlands E12000004
#> 6   W23   2024              REG  England E92000001 East Midlands E12000004
#> 7   W23   2024              REG  England E92000001 East Midlands E12000004
#> 8   W23   2024              REG  England E92000001 East Midlands E12000004
#> 9   W23   2024              NAT  England E92000001          <NA>      <NA>
#> 10  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 11  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 12  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 13  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 14  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 15  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 16  W23   2024              NAT  England E92000001          <NA>      <NA>
#>    attendance_status               attendance_type day_number
#> 1         Attendance                       Present      Total
#> 2         Attendance Approved educational activity      Total
#> 3            Absence                    Authorised      Total
#> 4            Absence                  Unauthorised      Total
#> 5         Attendance                       Present      Total
#> 6         Attendance Approved educational activity      Total
#> 7            Absence                    Authorised      Total
#> 8            Absence                  Unauthorised      Total
#> 9         Attendance                       Present      Total
#> 10        Attendance Approved educational activity      Total
#> 11           Absence                    Authorised      Total
#> 12           Absence                  Unauthorised      Total
#> 13        Attendance                       Present      Total
#> 14        Attendance Approved educational activity      Total
#> 15           Absence                    Authorised      Total
#> 16           Absence                  Unauthorised      Total
#>    establishment_phase reason session_count
#> 1            Secondary  Total       1862256
#> 2            Secondary  Total         35007
#> 3            Secondary  Total        105543
#> 4            Secondary  Total         83756
#> 5              Special  Total         74898
#> 6              Special  Total          1020
#> 7              Special  Total          7905
#> 8              Special  Total          2871
#> 9              Special  Total        870024
#> 10             Special  Total         17506
#> 11             Special  Total         88094
#> 12             Special  Total         37748
#> 13           Secondary  Total      20145467
#> 14           Secondary  Total        401209
#> 15           Secondary  Total       1093633
#> 16           Secondary  Total        956076