Query a data set using POST and a query json

This function provides a method for generating and sending a json based data query to the EES API. As a minimum, it requires the dataset_id flag and either the indicators flag or a json file containing a query to be provided.

Usage

post_dataset(
  dataset_id,
  indicators = NULL,
  time_periods = NULL,
  geographies = NULL,
  filter_items = NULL,
  json_query = NULL,
  dataset_version = NULL,
  ees_environment = NULL,
  api_version = NULL,
  page = NULL,
  page_size = 10000,
  parse = TRUE,
  debug = FALSE,
  verbose = FALSE
)

Arguments

dataset_id

ID of data set to be connected to. This is required if the endpoint is one of "get-dataset-versions", "get-summary", "get-meta", "get-csv", "get-data" or "post-data"

indicators

Indicators required as a string or vector of strings (required)

time_periods

Time periods required as a string ("period|code") or vector of strings

geographies

String, vector or data frame containing the geographic levels and locations to be queried.

filter_items

Filter items required as a string or vector of strings

json_query

Optional path to a json file containing the query parameters

dataset_version

Version of data set to be connected to

ees_environment

EES ees_environment to connect to: "dev", "test", "preprod" or "prod"

api_version

EES API version

page

Page number of query results to return

page_size

Number of results to return in a single query

parse

Logical flag to activate parsing of the results. Default: TRUE

debug

Run POST query in debug mode. Logical, default = FALSE

verbose

Run with additional contextual messaging. Logical, default = FALSE

Value

Data frame containing query results of an API data set

Examples

eesyapi:::post_dataset(
  example_id(group = "attendance"),
  json_query = example_json_query()
)
#>   time_period time_identifier geographic_level nat_name  nat_code   reg_name
#> 1        2024         Week 23         Regional  England E92000001 North East
#> 2        2024         Week 23         Regional  England E92000001 North East
#> 3        2024         Week 23         Regional  England E92000001 North West
#> 4        2024         Week 23         Regional  England E92000001 North West
#> 5        2024         Week 23         Regional  England E92000001 North East
#> 6        2024         Week 23         Regional  England E92000001 North East
#> 7        2024         Week 23         Regional  England E92000001 North West
#> 8        2024         Week 23         Regional  England E92000001 North West
#>    reg_code attendance_status attendance_type day_number education_phase
#> 1 E12000001           Absence      Authorised      Total       Secondary
#> 2 E12000001           Absence    Unauthorised      Total       Secondary
#> 3 E12000002           Absence      Authorised      Total       Secondary
#> 4 E12000002           Absence    Unauthorised      Total       Secondary
#> 5 E12000001           Absence      Authorised      Total           Total
#> 6 E12000001           Absence    Unauthorised      Total           Total
#> 7 E12000002           Absence      Authorised      Total           Total
#> 8 E12000002           Absence    Unauthorised      Total           Total
#>   attendance_reason session_count
#> 1             Total         53326
#> 2             Total         64441
#> 3             Total        152262
#> 4             Total        167394
#> 5             Total        107165
#> 6             Total        117871
#> 7             Total        273327
#> 8             Total        273633

# Run post_dataset() to select rows containing either of two geographic locations and a single
# filter item.
eesyapi:::post_dataset(
  example_id(group = "attendance"),
  indicators = example_id("indicator", group = "attendance"),
  time_periods = example_id("time_period", group = "attendance"),
  geographies = example_id("location_code", group = "attendance"),
  filter_items = example_id("filter_item", group = "attendance"),
  page = 1,
  page_size = 32
)
#>    time_period time_identifier geographic_level nat_name  nat_code
#> 1         2024         Week 23         National  England E92000001
#> 2         2024         Week 23         National  England E92000001
#> 3         2024         Week 23         National  England E92000001
#> 4         2024         Week 23         National  England E92000001
#> 5         2024         Week 23         National  England E92000001
#> 6         2024         Week 23         National  England E92000001
#> 7         2024         Week 23         National  England E92000001
#> 8         2024         Week 23         National  England E92000001
#> 9         2024         Week 23         National  England E92000001
#> 10        2024         Week 23         National  England E92000001
#> 11        2024         Week 23         National  England E92000001
#> 12        2024         Week 23         National  England E92000001
#> 13        2024         Week 23         National  England E92000001
#> 14        2024         Week 23         National  England E92000001
#> 15        2024         Week 23         National  England E92000001
#> 16        2024         Week 23         National  England E92000001
#> 17        2024         Week 23         National  England E92000001
#> 18        2024         Week 23         National  England E92000001
#> 19        2024         Week 23         National  England E92000001
#> 20        2024         Week 23         National  England E92000001
#> 21        2024         Week 23         National  England E92000001
#> 22        2024         Week 23         National  England E92000001
#> 23        2024         Week 23         National  England E92000001
#> 24        2024         Week 23         National  England E92000001
#>    attendance_status attendance_type day_number education_phase
#> 1            Absence    Unauthorised          1         Special
#> 2            Absence    Unauthorised          1         Primary
#> 3            Absence    Unauthorised          1       Secondary
#> 4            Absence    Unauthorised          2         Special
#> 5            Absence    Unauthorised          2         Primary
#> 6            Absence    Unauthorised          2       Secondary
#> 7            Absence    Unauthorised          3       Secondary
#> 8            Absence    Unauthorised          3         Special
#> 9            Absence    Unauthorised          3         Primary
#> 10           Absence    Unauthorised          4       Secondary
#> 11           Absence    Unauthorised          4         Primary
#> 12           Absence    Unauthorised          4         Special
#> 13           Absence    Unauthorised          5       Secondary
#> 14           Absence    Unauthorised          5         Special
#> 15           Absence    Unauthorised          5         Primary
#> 16           Absence    Unauthorised      Total         Special
#> 17           Absence    Unauthorised      Total         Primary
#> 18           Absence    Unauthorised      Total       Secondary
#> 19           Absence    Unauthorised          2           Total
#> 20           Absence    Unauthorised          3           Total
#> 21           Absence    Unauthorised          4           Total
#> 22           Absence    Unauthorised          5           Total
#> 23           Absence    Unauthorised      Total           Total
#> 24           Absence    Unauthorised          1           Total
#>         attendance_reason session_count
#> 1  G unauthorised holiday          1574
#> 2  G unauthorised holiday         84811
#> 3  G unauthorised holiday         44458
#> 4  G unauthorised holiday          1443
#> 5  G unauthorised holiday         98308
#> 6  G unauthorised holiday         37788
#> 7  G unauthorised holiday         33274
#> 8  G unauthorised holiday          1372
#> 9  G unauthorised holiday         86219
#> 10 G unauthorised holiday         29610
#> 11 G unauthorised holiday         75852
#> 12 G unauthorised holiday          1345
#> 13 G unauthorised holiday         29339
#> 14 G unauthorised holiday          1410
#> 15 G unauthorised holiday         77983
#> 16 G unauthorised holiday          7144
#> 17 G unauthorised holiday        423173
#> 18 G unauthorised holiday        174469
#> 19 G unauthorised holiday        137539
#> 20 G unauthorised holiday        120865
#> 21 G unauthorised holiday        106807
#> 22 G unauthorised holiday        108732
#> 23 G unauthorised holiday        604786
#> 24 G unauthorised holiday        130843