Query a data set

Create and send a query to the EES API. Queries can be supplied and run in one of 4 ways:

Supplying a json query in a file to be sent with the POST method.

Supplying a json query in a string variable to be sent with the POST method.

Supplying parameters (time_periods, geographies, filter_items, indicators) to build a json query is then sent with the POST method.

Supplying parameters (time_periods, geographies, filter_items, indicators) to build a json query is then sent with the GET method.

In all cases, the data set id must be supplied explicitly using the dataset_id.

Details on the format of each parameter for the latter two methods are as follows.

indicators

This must be supplied as a vector of sqids, which can be identified from the meta data using get_meta().

time_periods

Time periods should be supplied as a vector of periods and codes in the form:

"period|code"

For example, selecting the 2023 and 2024 academic years would require: time_period = c("2023|AY", "2024|AY")

geographies

Geographies can be supplied as a vector or a data frame depending on the complexity of the desired query.

A vector will run a query returning any rows meeting any of the given geographies, i.e. geographies = c("NAT", "REG") will return all national and regional level rows, whilst c("NAT", "REG|code|E120000001") will return all national level rows and all North East rows. Specific locations are required to be supplied in the format "LEVEL|identifier_type|identifier", where identifier type can be either code or id and the corresponding identifier is then the standard ONS code or the sqid given in the meta data respectively. Using England as an example, these would be:

"NAT|code|E92000001"

"NAT|id|dP0Zw"

If you require a more complex selection, for example all LAs in a given region, then a data frame should be supplied, with a row for each selection. Note however, that this will only work when using the default POST method. The GET method is much more limited and can not process more complex queries.

The geography query data frame should contain the following columns:

return_level: the geographic level to return (e.g. LA in the example above).

search_level: the geographic level of the search location (e.g. REG in the example above).

identifier_type: "code" or "id".

identifier: the code or id (sqid) for the search location (e.g. the code or sqid of the region in the above example).

Further rows can be added to add other geography searches to include in results.

An example of a working geographies data frame can be obtained using example_geography_query().

filter_items

Similarly to geographies, criteria for querying on filter items can be provided either via a vector for simple queries or, for more complex queries, as a list. In both cases, vector or list, filter items can only be supplied as sqids, i.e. the ids found in the meta data using get_meta(dataset_id).

Providing a vector of sqids will effectively run a query returning any rows containing any of the listed sqids. This therefore does not allow narrow searches based on a row or set of rows matching multiple criteria.

Providing a list structure can provide a more narrow query selecting individual rows based on a combination of criteria. For example if we want rows that contain sqid1 in one column and sqid2 in another, then we would pass a list of: filter_query <- list(column1 = c("sqid1"), column2 = c("sqid2")) Note that the naming of the entries in the list is not necessary, but may help in creating more readable code.

If we wish to create a query whereby we receive rows containing sqid1 in column 1 and sqid2 or sqid3 in column 2, then the required list would be: filter_query <- list(column1 = c("sqid1"), column2 = c("sqid2", "sqid3"))

In this way, we can build up combinations of OR and AND criteria for a query across multiple filter columns.

Note again that the more complex querying using a list variable will only function when using the POST method.

Controlling paging

You can request a specific set of rows using the page and page_size parameters. Keeping the default of page = NULL will return all rows matching the query. Setting page and page_size to numerical values will attempt to return a subset of rows, with page_size defining the number of rows and page defining which subset of rows to return from the query (i.e. page = 1, page_size = 20 will return the first 20 rows, page = 2 and page_size = 20 will return the second 20 rows and so on).

Usage

query_dataset(
  dataset_id,
  indicators = NULL,
  time_periods = NULL,
  geographies = NULL,
  geographic_levels = NULL,
  locations = NULL,
  filter_items = NULL,
  json_query = NULL,
  method = "POST",
  dataset_version = NULL,
  api_version = NULL,
  page_size = 1000,
  page = NULL,
  debug = FALSE,
  verbose = FALSE
)

Arguments

dataset_id

ID of data set to be connected to. This is required if the endpoint is one of "get-summary", "get-meta", "get-csv", "get-data" or "post-data"

indicators

Indicators required as a string or vector of strings (required)

time_periods

Time periods required as a string ("period|code") or vector of strings

geographies

String, vector or data frame containing the geographic levels and locations to be queried.

geographic_levels

Geographic levels required as a string or vector of strings

locations

Location code required as a string or vector of strings

filter_items

Filter items required as a string or vector of strings

json_query

Optional path to a json file containing the query parameters

method

The API query method to be used. Can be "POST" or "GET". Default: "POST".

dataset_version

Version of data set to be connected to

api_version

EES API version

page_size

Number of results to return in a single query

page

Page number of query results to return

debug

Run POST query in debug mode. Logical, default = FALSE

verbose

Run with additional contextual messaging. Logical, default = FALSE

Value

Data frame containing query results

Examples

# Run query_dataset() using a json query string input to json_query (this can also be done by
# passing a filename of a file containing your json query string).
query_dataset(
  example_id(group = "attendance"),
  json_query = example_json_query()
)
#>   code period geographic_level nat_name  nat_code      reg_name  reg_code
#> 1  W23   2024              REG  England E92000001 East Midlands E12000004
#> 2  W23   2024              REG  England E92000001 East Midlands E12000004
#> 3  W23   2024              REG  England E92000001 East Midlands E12000004
#> 4  W23   2024              REG  England E92000001 East Midlands E12000004
#> 5  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 6  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 7  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 8  W23   2024              NAT  England E92000001          <NA>      <NA>
#>   attendance_status               attendance_type day_number
#> 1        Attendance                       Present      Total
#> 2        Attendance Approved educational activity      Total
#> 3        Attendance                       Present      Total
#> 4        Attendance Approved educational activity      Total
#> 5        Attendance                       Present      Total
#> 6        Attendance Approved educational activity      Total
#> 7        Attendance                       Present      Total
#> 8        Attendance Approved educational activity      Total
#>   establishment_phase reason session_count
#> 1           Secondary  Total       1862256
#> 2           Secondary  Total         35007
#> 3             Special  Total         74898
#> 4             Special  Total          1020
#> 5             Special  Total        870024
#> 6             Special  Total         17506
#> 7           Secondary  Total      20145467
#> 8           Secondary  Total        401209

# If you don't want to have tio write your own json query, the rest of the examples illustrate
# how to use query_dataset() with parameters to construct queries in R.

# Run query_dataset() to select rows containing either of two geographic locations and either of
# two filter items.
query_dataset(
  example_id(group = "attendance"),
  indicators = example_id("indicator", group = "attendance"),
  time_periods = "2024|W23",
  geographies = c("NAT|id|dP0Zw", "REG|id|rg3Nj"),
  filter_items = c("CvuId", "6AXrf"),
  page = 1,
  page_size = 32
)
#>    code period geographic_level nat_name  nat_code      reg_name  reg_code
#> 1   W23   2024              REG  England E92000001 East Midlands E12000004
#> 2   W23   2024              REG  England E92000001 East Midlands E12000004
#> 3   W23   2024              REG  England E92000001 East Midlands E12000004
#> 4   W23   2024              REG  England E92000001 East Midlands E12000004
#> 5   W23   2024              REG  England E92000001 East Midlands E12000004
#> 6   W23   2024              REG  England E92000001 East Midlands E12000004
#> 7   W23   2024              REG  England E92000001 East Midlands E12000004
#> 8   W23   2024              REG  England E92000001 East Midlands E12000004
#> 9   W23   2024              REG  England E92000001 East Midlands E12000004
#> 10  W23   2024              REG  England E92000001 East Midlands E12000004
#> 11  W23   2024              REG  England E92000001 East Midlands E12000004
#> 12  W23   2024              REG  England E92000001 East Midlands E12000004
#> 13  W23   2024              REG  England E92000001 East Midlands E12000004
#> 14  W23   2024              REG  England E92000001 East Midlands E12000004
#> 15  W23   2024              REG  England E92000001 East Midlands E12000004
#> 16  W23   2024              REG  England E92000001 East Midlands E12000004
#> 17  W23   2024              REG  England E92000001 East Midlands E12000004
#> 18  W23   2024              REG  England E92000001 East Midlands E12000004
#> 19  W23   2024              REG  England E92000001 East Midlands E12000004
#> 20  W23   2024              REG  England E92000001 East Midlands E12000004
#> 21  W23   2024              REG  England E92000001 East Midlands E12000004
#> 22  W23   2024              REG  England E92000001 East Midlands E12000004
#> 23  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 24  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 25  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 26  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 27  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 28  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 29  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 30  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 31  W23   2024              NAT  England E92000001          <NA>      <NA>
#> 32  W23   2024              NAT  England E92000001          <NA>      <NA>
#>    attendance_status               attendance_type day_number
#> 1         Attendance                       Present          1
#> 2         Attendance Approved educational activity          1
#> 3         Attendance                       Present          1
#> 4         Attendance                       Present          1
#> 5         Attendance                       Present          1
#> 6         Attendance                       Present          1
#> 7         Attendance Approved educational activity          1
#> 8         Attendance Approved educational activity          1
#> 9         Attendance Approved educational activity          1
#> 10        Attendance Approved educational activity          1
#> 11        Attendance Approved educational activity          1
#> 12        Attendance                       Present          1
#> 13        Attendance Approved educational activity          1
#> 14        Attendance                       Present          1
#> 15        Attendance                       Present          1
#> 16        Attendance                       Present          1
#> 17        Attendance                       Present          1
#> 18        Attendance Approved educational activity          1
#> 19        Attendance Approved educational activity          1
#> 20        Attendance Approved educational activity          1
#> 21        Attendance Approved educational activity          1
#> 22        Attendance Approved educational activity          1
#> 23        Attendance                       Present          1
#> 24        Attendance Approved educational activity          1
#> 25        Attendance                       Present          1
#> 26        Attendance                       Present          1
#> 27        Attendance                       Present          1
#> 28        Attendance                       Present          1
#> 29        Attendance Approved educational activity          1
#> 30        Attendance Approved educational activity          1
#> 31        Attendance Approved educational activity          1
#> 32        Attendance Approved educational activity          1
#>    establishment_phase                                 reason session_count
#> 1              Primary                                  Total        475991
#> 2              Primary                                  Total          5235
#> 3              Primary                             Present am        234417
#> 4              Primary                             Present pm        236991
#> 5              Primary                                Present        471408
#> 6              Primary L present late before registers closed          4583
#> 7              Primary                   B education off site           452
#> 8              Primary                    D dual registration           226
#> 9              Primary           P approved sporting activity           552
#> 10             Primary               V educational visit trip          4225
#> 11             Primary                      W work experience             0
#> 12             Special                                  Total         14497
#> 13             Special                                  Total           130
#> 14             Special                             Present am          7206
#> 15             Special                             Present pm          7203
#> 16             Special                                Present         14409
#> 17             Special L present late before registers closed            88
#> 18             Special                   B education off site           103
#> 19             Special                    D dual registration            26
#> 20             Special           P approved sporting activity             1
#> 21             Special               V educational visit trip            24
#> 22             Special                      W work experience             0
#> 23             Special                                  Total        157670
#> 24             Special                                  Total          2611
#> 25             Special                             Present am         77609
#> 26             Special                             Present pm         78597
#> 27             Special                                Present        156206
#> 28             Special L present late before registers closed          1464
#> 29             Special                   B education off site          2132
#> 30             Special                    D dual registration           287
#> 31             Special           P approved sporting activity            46
#> 32             Special               V educational visit trip           343

# Run query_dataset() using set parameters giving a combination of filter options
query_dataset(
  example_id(group = "attendance"),
  indicators = example_id("indicator", group = "attendance"),
  time_periods = "2024|W23",
  geographies = c("NAT"),
  filter_items = list(
    attendance_status = c("pmRSo", "7SdXo"),
    attendance_type = c("CvuId", "6AXrf", "0k3T5", "YdkHK"),
    education_phase = c("ThDPJ", "crH31"),
    day_number = c("uLQo4"),
    reason = c("bBrtT")
  )
)
#>   code period geographic_level nat_name  nat_code attendance_status
#> 1  W23   2024              NAT  England E92000001        Attendance
#> 2  W23   2024              NAT  England E92000001        Attendance
#> 3  W23   2024              NAT  England E92000001           Absence
#> 4  W23   2024              NAT  England E92000001           Absence
#> 5  W23   2024              NAT  England E92000001        Attendance
#> 6  W23   2024              NAT  England E92000001        Attendance
#> 7  W23   2024              NAT  England E92000001           Absence
#> 8  W23   2024              NAT  England E92000001           Absence
#>                 attendance_type day_number establishment_phase reason
#> 1                       Present      Total             Special  Total
#> 2 Approved educational activity      Total             Special  Total
#> 3                    Authorised      Total             Special  Total
#> 4                  Unauthorised      Total             Special  Total
#> 5                       Present      Total           Secondary  Total
#> 6 Approved educational activity      Total           Secondary  Total
#> 7                    Authorised      Total           Secondary  Total
#> 8                  Unauthorised      Total           Secondary  Total
#>   session_count
#> 1        870024
#> 2         17506
#> 3         88094
#> 4         37748
#> 5      20145467
#> 6        401209
#> 7       1093633
#> 8        956076

# Run a query with a more complex geography selection. Return data for all of:
#   - England
#   - Yorkshire and the Humber
#   - All LAs in Yorkshire and the Humber
example_geography_query("nat_yorks_yorkslas")
#>   return_level search_level identifier_type identifier
#> 1          NAT          NAT            code  E92000001
#> 2          REG          REG            code  E12000003
#> 3           LA          REG            code  E12000003
query_dataset(
  example_id(group = "attendance"),
  indicators = example_id("indicator", group = "attendance"),
  time_periods = "2024|W23",
  geographies = example_geography_query("nat_yorks_yorkslas"),
  filter_items = list(
    attendance_status = c("pmRSo"),
    attendance_type = c("CvuId"),
    education_phase = c("ThDPJ"),
    day_number = c("uLQo4"),
    reason = c("bBrtT")
  )
)
#>    code period geographic_level                     la_name   la_code
#> 1   W23   2024               LA                   Doncaster E08000017
#> 2   W23   2024               LA    East Riding of Yorkshire E06000011
#> 3   W23   2024               LA     North East Lincolnshire E06000012
#> 4   W23   2024               LA                    Kirklees E08000034
#> 5   W23   2024               LA                   Rotherham E08000018
#> 6   W23   2024               LA                  Calderdale E08000033
#> 7   W23   2024               LA                    Bradford E08000032
#> 8   W23   2024               LA                       Leeds E08000035
#> 9   W23   2024               LA                   Wakefield E08000036
#> 10  W23   2024               LA          North Lincolnshire E06000013
#> 11  W23   2024               LA Kingston upon Hull, City of E06000010
#> 12  W23   2024               LA                        York E06000014
#> 13  W23   2024               LA                    Barnsley E08000016
#> 14  W23   2024               LA             North Yorkshire E06000065
#> 15  W23   2024               LA                   Sheffield E08000019
#> 16  W23   2024              REG                        <NA>      <NA>
#> 17  W23   2024              NAT                        <NA>      <NA>
#>    la_oldCode nat_name  nat_code                 reg_name  reg_code
#> 1         371  England E92000001 Yorkshire and The Humber E12000003
#> 2         811  England E92000001 Yorkshire and The Humber E12000003
#> 3         812  England E92000001 Yorkshire and The Humber E12000003
#> 4         382  England E92000001 Yorkshire and The Humber E12000003
#> 5         372  England E92000001 Yorkshire and The Humber E12000003
#> 6         381  England E92000001 Yorkshire and The Humber E12000003
#> 7         380  England E92000001 Yorkshire and The Humber E12000003
#> 8         383  England E92000001 Yorkshire and The Humber E12000003
#> 9         384  England E92000001 Yorkshire and The Humber E12000003
#> 10        813  England E92000001 Yorkshire and The Humber E12000003
#> 11        810  England E92000001 Yorkshire and The Humber E12000003
#> 12        816  England E92000001 Yorkshire and The Humber E12000003
#> 13        370  England E92000001 Yorkshire and The Humber E12000003
#> 14        815  England E92000001 Yorkshire and The Humber E12000003
#> 15        373  England E92000001 Yorkshire and The Humber E12000003
#> 16       <NA>  England E92000001 Yorkshire and The Humber E12000003
#> 17       <NA>  England E92000001                     <NA>      <NA>
#>    attendance_status attendance_type day_number establishment_phase reason
#> 1         Attendance         Present      Total           Secondary  Total
#> 2         Attendance         Present      Total           Secondary  Total
#> 3         Attendance         Present      Total           Secondary  Total
#> 4         Attendance         Present      Total           Secondary  Total
#> 5         Attendance         Present      Total           Secondary  Total
#> 6         Attendance         Present      Total           Secondary  Total
#> 7         Attendance         Present      Total           Secondary  Total
#> 8         Attendance         Present      Total           Secondary  Total
#> 9         Attendance         Present      Total           Secondary  Total
#> 10        Attendance         Present      Total           Secondary  Total
#> 11        Attendance         Present      Total           Secondary  Total
#> 12        Attendance         Present      Total           Secondary  Total
#> 13        Attendance         Present      Total           Secondary  Total
#> 14        Attendance         Present      Total           Secondary  Total
#> 15        Attendance         Present      Total           Secondary  Total
#> 16        Attendance         Present      Total           Secondary  Total
#> 17        Attendance         Present      Total           Secondary  Total
#>    session_count
#> 1         100362
#> 2         122064
#> 3          64103
#> 4         193988
#> 5         114645
#> 6          91976
#> 7         259606
#> 8         337686
#> 9         144054
#> 10         73286
#> 11        100724
#> 12         68239
#> 13         95796
#> 14        222434
#> 15        194561
#> 16       2183524
#> 17      20145467

# Run a basic query using GET instead of POST
query_dataset(
  example_id(),
  method = "GET",
  geographic_levels = c("SCH"),
  filter_items = example_id("filter_item"),
  indicators = example_id("indicator"),
  page = 1,
  page_size = 10
)
#> Warning: Using GET to query a data set offers limited functionality, we recommend using POST alongside a JSON structured query instead:
#>    - query_dataset(..., method = 'POST')
#>    code    period geographic_level    la_name   la_code la_oldCode nat_name
#> 1    AY 2022/2023              SCH Manchester E08000003        352  England
#> 2    AY 2022/2023              SCH Manchester E08000003        352  England
#> 3    AY 2022/2023              SCH Manchester E08000003        352  England
#> 4    AY 2022/2023              SCH Manchester E08000003        352  England
#> 5    AY 2022/2023              SCH Manchester E08000003        352  England
#> 6    AY 2022/2023              SCH Manchester E08000003        352  England
#> 7    AY 2022/2023              SCH Manchester E08000003        352  England
#> 8    AY 2022/2023              SCH Manchester E08000003        352  England
#> 9    AY 2022/2023              SCH Manchester E08000003        352  England
#> 10   AY 2022/2023              SCH Manchester E08000003        352  England
#>     nat_code       sch_name sch_urn sch_laEstab establishment_type      type
#> 1  E92000001 Loreto College  130503     3528601 Sixth Form College      1618
#> 2  E92000001 Loreto College  130503     3528601 Sixth Form College      ALEV
#> 3  E92000001 Loreto College  130503     3528601 Sixth Form College      ACAD
#> 4  E92000001 Loreto College  130503     3528601 Sixth Form College      AGEN
#> 5  E92000001 Loreto College  130503     3528601 Sixth Form College      TLEV
#> 6  E92000001 Loreto College  130503     3528601 Sixth Form College TLEVDENOM
#> 7  E92000001 Loreto College  130503     3528601 Sixth Form College AGENDENOM
#> 8  E92000001 Loreto College  130503     3528601 Sixth Form College SCOPE_L3M
#> 9  E92000001 Loreto College  130503     3528601 Sixth Form College  TechCert
#> 10 E92000001 Loreto College  130503     3528601 Sixth Form College  ANYL2VOC
#>    total_pupil_value
#> 1               1777
#> 2               1505
#> 3               1505
#> 4                510
#> 5                  z
#> 6                 18
#> 7                524
#> 8                  z
#> 9                  z
#> 10                 z