Query a data set

Create and send a query to the EES API. Queries can be supplied and run in one of 4 ways:

Supplying a json query in a file to be sent with the POST method.

Supplying a json query in a string variable to be sent with the POST method.

Supplying parameters (time_periods, geographies, filter_items, indicators) to build a json query is then sent with the POST method.

Supplying parameters (time_periods, geographies, filter_items, indicators) to build a json query is then sent with the GET method.

In all cases, the data set id must be supplied explicitly using the dataset_id.

Details on the format of each parameter for the latter two methods are as follows.

indicators

This must be supplied as a vector of sqids, which can be identified from the meta data using get_meta().

time_periods

Time periods should be supplied as a vector of periods and codes in the form:

"period|code"

For example, selecting the 2023 and 2024 academic years would require: time_period = c("2023|AY", "2024|AY")

geographies

Geographies can be supplied as a vector or a data frame depending on the complexity of the desired query.

A vector will run a query returning any rows meeting any of the given geographies, i.e. geographies = c("NAT", "REG") will return all national and regional level rows, whilst c("NAT", "REG|code|E120000001") will return all national level rows and all North East rows. Specific locations are required to be supplied in the format "LEVEL|identifier_type|identifier", where identifier type can be either code or id and the corresponding identifier is then the standard ONS code or the sqid given in the meta data respectively. Using England as an example, these would be:

"NAT|code|E92000001"

"NAT|id|dP0Zw"

If you require a more complex selection, for example all LAs in a given region, then a data frame should be supplied, with a row for each selection. Note however, that this will only work when using the default POST method. The GET method is much more limited and can not process more complex queries.

The geography query data frame should contain the following columns:

return_level: the geographic level to return (e.g. LA in the example above).

search_level: the geographic level of the search location (e.g. REG in the example above).

identifier_type: "code" or "id".

identifier: the code or id (sqid) for the search location (e.g. the code or sqid of the region in the above example).

Further rows can be added to add other geography searches to include in results.

An example of a working geographies data frame can be obtained using example_geography_query().

filter_items

Similarly to geographies, criteria for querying on filter items can be provided either via a vector for simple queries or, for more complex queries, as a list. In both cases, vector or list, filter items can only be supplied as sqids, i.e. the ids found in the meta data using get_meta(dataset_id).

Providing a vector of sqids will effectively run a query returning any rows containing any of the listed sqids. This therefore does not allow narrow searches based on a row or set of rows matching multiple criteria.

Providing a list structure can provide a more narrow query selecting individual rows based on a combination of criteria. For example if we want rows that contain sqid1 in one column and sqid2 in another, then we would pass a list of: filter_query <- list(column1 = c("sqid1"), column2 = c("sqid2")) Note that the naming of the entries in the list is not necessary, but may help in creating more readable code.

If we wish to create a query whereby we receive rows containing sqid1 in column 1 and sqid2 or sqid3 in column 2, then the required list would be: filter_query <- list(column1 = c("sqid1"), column2 = c("sqid2", "sqid3"))

In this way, we can build up combinations of OR and AND criteria for a query across multiple filter columns.

Note again that the more complex querying using a list variable will only function when using the POST method.

Controlling paging

You can request a specific set of rows using the page and page_size parameters. Keeping the default of page = NULL will return all rows matching the query. Setting page and page_size to numerical values will attempt to return a subset of rows, with page_size defining the number of rows and page defining which subset of rows to return from the query (i.e. page = 1, page_size = 20 will return the first 20 rows, page = 2 and page_size = 20 will return the second 20 rows and so on).

Usage

query_dataset(
  dataset_id,
  indicators = NULL,
  time_periods = NULL,
  geographies = NULL,
  geographic_levels = NULL,
  locations = NULL,
  filter_items = NULL,
  json_query = NULL,
  method = "POST",
  dataset_version = NULL,
  api_version = NULL,
  page_size = 1000,
  page = NULL,
  debug = FALSE,
  verbose = FALSE
)

Arguments

dataset_id

ID of data set to be connected to. This is required if the endpoint is one of "get-summary", "get-meta", "get-data" or "post-data"

indicators

Indicators required as a string or vector of strings (required)

time_periods

Time periods required as a string ("period|code") or vector of strings

geographies

String, vector or data frame containing the geographic levels and locations to be queried.

geographic_levels

Geographic levels required as a string or vector of strings

locations

Location code required as a string or vector of strings

filter_items

Filter items required as a string or vector of strings

json_query

Optional path to a json file containing the query parameters

method

The API query method to be used. Can be "POST" or "GET". Default: "POST".

dataset_version

Version of data set to be connected to

api_version

EES API version

page_size

Number of results to return in a single query

page

Page number of query results to return

debug

Run POST query in debug mode: logic, default: FALSE

verbose

Run with additional contextual messaging, logical, default = FALSE

Value

Data frame containing query results

Examples

# Run query_dataset() using a json query string input to json_query (this can also be done by
# passing a filename of a file containing your json query string).
query_dataset(
  example_id(group = "attendance"),
  json_query = example_json_query()
)
#> Parsing query options
#>   code period geographic_level   NAT   REG Db3Qe emJuS 4kdUZ 5TYdi mU59K
#> 1  W23   2024              REG dP0Zw rg3Nj pmRSo CvuId ThDPJ uLQo4 bBrtT
#> 2  W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf ThDPJ uLQo4 bBrtT
#> 3  W23   2024              REG dP0Zw rg3Nj pmRSo CvuId crH31 uLQo4 bBrtT
#> 4  W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf crH31 uLQo4 bBrtT
#> 5  W23   2024              NAT dP0Zw  <NA> pmRSo CvuId crH31 uLQo4 bBrtT
#> 6  W23   2024              NAT dP0Zw  <NA> pmRSo 6AXrf crH31 uLQo4 bBrtT
#> 7  W23   2024              NAT dP0Zw  <NA> pmRSo CvuId ThDPJ uLQo4 bBrtT
#> 8  W23   2024              NAT dP0Zw  <NA> pmRSo 6AXrf ThDPJ uLQo4 bBrtT
#>        bqZtT
#> 1    1862256
#> 2      35007
#> 3      74898
#> 4       1020
#> 5     870024
#> 6      17506
#> 7   20145467
#> 8     401209

# If you don't want to have tio write your own json query, the rest of the examples illustrate
# how to use query_dataset() with parameters to construct queries in R.

# Run query_dataset() to select rows containing either of two geographic locations and either of
# two filter items.
query_dataset(
  example_id(group = "attendance"),
  indicators = example_id("indicator", group = "attendance"),
  time_periods = "2024|W23",
  geographies = c("NAT|id|dP0Zw", "REG|id|rg3Nj"),
  filter_items = c("CvuId", "6AXrf"),
  page = 1,
  page_size = 32
)
#>    code period geographic_level   NAT   REG Db3Qe emJuS 4kdUZ 5TYdi mU59K
#> 1   W23   2024              REG dP0Zw rg3Nj pmRSo CvuId XVWfK qf0jG bBrtT
#> 2   W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf XVWfK qf0jG bBrtT
#> 3   W23   2024              REG dP0Zw rg3Nj pmRSo CvuId XVWfK qf0jG GzNgr
#> 4   W23   2024              REG dP0Zw rg3Nj pmRSo CvuId XVWfK qf0jG 3Nk8u
#> 5   W23   2024              REG dP0Zw rg3Nj pmRSo CvuId XVWfK qf0jG rGcNj
#> 6   W23   2024              REG dP0Zw rg3Nj pmRSo CvuId XVWfK qf0jG PyBeb
#> 7   W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf XVWfK qf0jG t74Em
#> 8   W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf XVWfK qf0jG jYyAM
#> 9   W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf XVWfK qf0jG oWvmX
#> 10  W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf XVWfK qf0jG ea0uS
#> 11  W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf XVWfK qf0jG 5USdi
#> 12  W23   2024              REG dP0Zw rg3Nj pmRSo CvuId crH31 qf0jG bBrtT
#> 13  W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf crH31 qf0jG bBrtT
#> 14  W23   2024              REG dP0Zw rg3Nj pmRSo CvuId crH31 qf0jG GzNgr
#> 15  W23   2024              REG dP0Zw rg3Nj pmRSo CvuId crH31 qf0jG 3Nk8u
#> 16  W23   2024              REG dP0Zw rg3Nj pmRSo CvuId crH31 qf0jG rGcNj
#> 17  W23   2024              REG dP0Zw rg3Nj pmRSo CvuId crH31 qf0jG PyBeb
#> 18  W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf crH31 qf0jG t74Em
#> 19  W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf crH31 qf0jG jYyAM
#> 20  W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf crH31 qf0jG oWvmX
#> 21  W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf crH31 qf0jG ea0uS
#> 22  W23   2024              REG dP0Zw rg3Nj pmRSo 6AXrf crH31 qf0jG 5USdi
#> 23  W23   2024              NAT dP0Zw  <NA> pmRSo CvuId crH31 qf0jG bBrtT
#> 24  W23   2024              NAT dP0Zw  <NA> pmRSo 6AXrf crH31 qf0jG bBrtT
#> 25  W23   2024              NAT dP0Zw  <NA> pmRSo CvuId crH31 qf0jG GzNgr
#> 26  W23   2024              NAT dP0Zw  <NA> pmRSo CvuId crH31 qf0jG 3Nk8u
#> 27  W23   2024              NAT dP0Zw  <NA> pmRSo CvuId crH31 qf0jG rGcNj
#> 28  W23   2024              NAT dP0Zw  <NA> pmRSo CvuId crH31 qf0jG PyBeb
#> 29  W23   2024              NAT dP0Zw  <NA> pmRSo 6AXrf crH31 qf0jG t74Em
#> 30  W23   2024              NAT dP0Zw  <NA> pmRSo 6AXrf crH31 qf0jG jYyAM
#> 31  W23   2024              NAT dP0Zw  <NA> pmRSo 6AXrf crH31 qf0jG oWvmX
#> 32  W23   2024              NAT dP0Zw  <NA> pmRSo 6AXrf crH31 qf0jG ea0uS
#>         bqZtT
#> 1      475991
#> 2        5235
#> 3      234417
#> 4      236991
#> 5      471408
#> 6        4583
#> 7         452
#> 8         226
#> 9         552
#> 10       4225
#> 11          0
#> 12      14497
#> 13        130
#> 14       7206
#> 15       7203
#> 16      14409
#> 17         88
#> 18        103
#> 19         26
#> 20          1
#> 21         24
#> 22          0
#> 23     157670
#> 24       2611
#> 25      77609
#> 26      78597
#> 27     156206
#> 28       1464
#> 29       2132
#> 30        287
#> 31         46
#> 32        343

# Run query_dataset() using set parameters giving a combination of filter options
query_dataset(
  example_id(group = "attendance"),
  indicators = example_id("indicator", group = "attendance"),
  time_periods = "2024|W23",
  geographies = c("NAT"),
  filter_items = list(
    attendance_status = c("pmRSo", "7SdXo"),
    attendance_type = c("CvuId", "6AXrf", "0k3T5", "YdkHK"),
    education_phase = c("ThDPJ", "crH31"),
    day_number = c("uLQo4"),
    reason = c("bBrtT")
  )
)
#>   code period geographic_level   NAT Db3Qe emJuS 4kdUZ 5TYdi mU59K      bqZtT
#> 1  W23   2024              NAT dP0Zw pmRSo CvuId crH31 uLQo4 bBrtT     870024
#> 2  W23   2024              NAT dP0Zw pmRSo 6AXrf crH31 uLQo4 bBrtT      17506
#> 3  W23   2024              NAT dP0Zw 7SdXo 0k3T5 crH31 uLQo4 bBrtT      88094
#> 4  W23   2024              NAT dP0Zw 7SdXo YdkHK crH31 uLQo4 bBrtT      37748
#> 5  W23   2024              NAT dP0Zw pmRSo CvuId ThDPJ uLQo4 bBrtT   20145467
#> 6  W23   2024              NAT dP0Zw pmRSo 6AXrf ThDPJ uLQo4 bBrtT     401209
#> 7  W23   2024              NAT dP0Zw 7SdXo 0k3T5 ThDPJ uLQo4 bBrtT    1093633
#> 8  W23   2024              NAT dP0Zw 7SdXo YdkHK ThDPJ uLQo4 bBrtT     956076

# Run a query with a more complex geography selection. Return data for all of:
#   - England
#   - Yorkshire and the Humber
#   - All LAs in Yorkshire and the Humber
example_geography_query("nat_yorks_yorkslas")
#>   return_level search_level identifier_type identifier
#> 1          NAT          NAT            code  E92000001
#> 2          REG          REG            code  E12000004
#> 3           LA          REG            code  E12000004
query_dataset(
  example_id(group = "attendance"),
  indicators = example_id("indicator", group = "attendance"),
  time_periods = "2024|W23",
  geographies = example_geography_query("nat_yorks_yorkslas"),
  filter_items = list(
    attendance_status = c("pmRSo"),
    attendance_type = c("CvuId"),
    education_phase = c("ThDPJ"),
    day_number = c("uLQo4"),
    reason = c("bBrtT")
  )
)
#>    code period geographic_level    LA   NAT   REG Db3Qe emJuS 4kdUZ 5TYdi mU59K
#> 1   W23   2024               LA mWI9K dP0Zw rg3Nj pmRSo CvuId ThDPJ uLQo4 bBrtT
#> 2   W23   2024               LA hZYyW dP0Zw rg3Nj pmRSo CvuId ThDPJ uLQo4 bBrtT
#> 3   W23   2024               LA 5RVdi dP0Zw rg3Nj pmRSo CvuId ThDPJ uLQo4 bBrtT
#> 4   W23   2024               LA TkqPJ dP0Zw rg3Nj pmRSo CvuId ThDPJ uLQo4 bBrtT
#> 5   W23   2024               LA eyjuS dP0Zw rg3Nj pmRSo CvuId ThDPJ uLQo4 bBrtT
#> 6   W23   2024               LA eLGuS dP0Zw rg3Nj pmRSo CvuId ThDPJ uLQo4 bBrtT
#> 7   W23   2024               LA BSk7J dP0Zw rg3Nj pmRSo CvuId ThDPJ uLQo4 bBrtT
#> 8   W23   2024               LA gJZO9 dP0Zw rg3Nj pmRSo CvuId ThDPJ uLQo4 bBrtT
#> 9   W23   2024              REG  <NA> dP0Zw rg3Nj pmRSo CvuId ThDPJ uLQo4 bBrtT
#> 10  W23   2024              NAT  <NA> dP0Zw  <NA> pmRSo CvuId ThDPJ uLQo4 bBrtT
#> 11  W23   2024               LA eLTuS dP0Zw rg3Nj pmRSo CvuId ThDPJ uLQo4 bBrtT
#> 12  W23   2024               LA jHaAM dP0Zw rg3Nj pmRSo CvuId ThDPJ uLQo4 bBrtT
#>         bqZtT
#> 1      131228
#> 2       21456
#> 3      194011
#> 4      147391
#> 5      121030
#> 6      294233
#> 7      244947
#> 8      296357
#> 9     1862256
#> 10   20145467
#> 11     134245
#> 12     277358

# Run a basic query using GET instead of POST
query_dataset(
  example_id(),
  method = "GET",
  geographic_levels = c("SCH"),
  filter_items = example_id("filter_item"),
  indicators = example_id("indicator"),
  page = 1,
  page_size = 10
)
#> Warning: Using GET to query a data set offers limited functionality, we recommend using POST alongside a JSON structured query instead:
#>    - query_dataset(..., method = 'POST')
#>    code    period geographic_level    LA   NAT   SCH f2GYF tu1Em PbNeb
#> 1    AY 2022/2023              SCH H0RzL dP0Zw mLJt9 9Rpn4 wEZcb  1777
#> 2    AY 2022/2023              SCH H0RzL dP0Zw mLJt9 HJjXz wEZcb  1505
#> 3    AY 2022/2023              SCH H0RzL dP0Zw mLJt9 7YbEX wEZcb  1505
#> 4    AY 2022/2023              SCH H0RzL dP0Zw mLJt9 iNW86 wEZcb   510
#> 5    AY 2022/2023              SCH H0RzL dP0Zw mLJt9 u9vRo wEZcb     z
#> 6    AY 2022/2023              SCH H0RzL dP0Zw mLJt9 XHEvf wEZcb    18
#> 7    AY 2022/2023              SCH H0RzL dP0Zw mLJt9 LleUW wEZcb   524
#> 8    AY 2022/2023              SCH H0RzL dP0Zw mLJt9 a2eWL wEZcb     z
#> 9    AY 2022/2023              SCH H0RzL dP0Zw mLJt9 Bfbf7 wEZcb     z
#> 10   AY 2022/2023              SCH H0RzL dP0Zw mLJt9 qGKqj wEZcb     z